Exodus.jl

Documentation for Exodus.jl

Documentation

Exodus.Block — Type

Block{I <: ExoInt, B <: ExoInt}

Container for reading in blocks

Exodus.Block — Method

Block(exo::ExodusDatabase{M, I, B, F}, block_id::I) where {M <: ExoInt, I <: ExoInt,
                                                           B <: ExoInt, F <: ExoFloat}

Init method for block container.

Exodus.put_coordinate_names — Method

put_coordinate_names(exo::ExodusDatabase{M, I, B, F}, 
                     coord_names::Vector{String}) where {M <: ExoInt, I <: ExoInt,
                                                         B <: ExoInt, F <: ExoFloat}

Work in progress...

Exodus.put_coordinates — Method

put_coordinates(exo::ExodusDatabase{M, I, B, F}, 
                     coords::Matrix{F}) where {M <: ExoInt, I <: ExoInt,
                                               B <: ExoInt, F <: ExoFloat}

Work in progress... not that well tested

Exodus.read_block_ids — Method

read_block_ids(exo::ExodusDatabase{M, I, B, F}, 
               init::Initialization) where {M <: ExoInt, I <: ExoInt, 
                                            B <: ExoInt, F <: ExoFloat}

Retrieves numerical block ids.

Exodus.read_blocks — Method

read_blocks(exo::ExodusDatabase{M, I, B, F}, 
            block_ids::Vector{I}) where {M <: ExoInt, I <: ExoInt,
                                         B <: ExoInt, F <: ExoFloat}

Helper method for initializing blocks.

Exodus.read_coordinate_names — Method

read_coordinate_names(exo::ExodusDatabase{M, I, B, F}, 
                      init::Initialization) where {M <: ExoInt, I <: ExoInt,
                                                   B <: ExoInt, F <: ExoFloat}

Exodus.read_coordinates — Method

read_coordinates(exo::ExodusDatabase{M, I, B, F}, 
                 init::Initialization) where {M <: ExoInt, I <: ExoInt,
                                              B <: ExoInt, F <: ExoFloat}

Method to read coordinates. Returns a matrix that is nnodes x ndim.

TODO... This method should really return ndim x nnodes but there's TODO... issues encountered downstream with some views found in Tardigrade