Exodus.jl Methods

Public methods

Base.close — Method

close(exo::ExodusDatabase)

Used to close and ExodusDatabase.

Base.copy — Method

copy(
    exo::ExodusDatabase,
    new_file_name::String;
    mesh_only_flag
)

Used to copy an ExodusDatabase. As of right now this is the best way to create a new ExodusDatabase for output. Not all of the put methods have been wrapped and properly tested. This one has though.

Base.length — Method

Base.length — Method

Exodus.collect_element_connectivities — Method

collect_element_connectivities(
    exo::ExodusDatabase{M, I, B, F}
) -> Any

collects all blocks by default

Exodus.collect_element_to_element_connectivities — Method

collect_element_to_element_connectivities(
    exo::ExodusDatabase{M, I, B, F}
) -> Any

collect all blocks by default

Exodus.collect_node_to_element_connectivities — Method

collect_node_to_element_connectivities(
    exo::ExodusDatabase{M, I, B, F}
) -> Any

collect all blocks by default

Exodus.copy_mesh — Method

copy_mesh(file_name::String, new_file_name::String)

Simpler copy method to only copy a mesh for output later on

Exodus.decomp — Method

decomp(file_name::String, n_procs::Integer)

Exodus.epu — Method

epu(file_name::String)

Exodus.epu — Method

epu() -> Any

Prints epu help message

Exodus.exodiff — Method

exodiff(
    ex_1::String,
    ex_2::String,
    cli_args::Vector{String}
) -> Bool

Exodus.exodiff — Method

exodiff(ex_1::String, ex_2::String; command_file) -> Bool

Return true if the two files pass the exodiff test. Otherwise it returns false

Exodus.exodiff — Method

exodiff() -> Any

Prints exodiff help message

Exodus.read_block — Method

read_block(exo::ExodusDatabase, block_id::Integer) -> Block

Exodus.read_block — Method

read_block(exo::ExodusDatabase, block_name::String) -> Block

Exodus.read_block_id_map — Method

read_block_id_map(
    exo::ExodusDatabase,
    block_id::Integer
) -> Any

TODO do this one later... it depends on a few things

TODO fix this to not use void_int... use a proper type

Exodus.read_blocks — Method

read_blocks(
    exo::ExodusDatabase,
    block_ids::Union{Vector{<:Integer}, var"#s130"} where var"#s130"<:Integer
) -> Any

Helper method for initializing blocks.

TODO: change name to readelementblocks

Exodus.read_coordinates — Method

read_coordinates(exo::ExodusDatabase) -> Matrix

Method to read coordinates. Returns a matrix that is ndim x nnodes.

Exodus.read_element_type — Method

read_element_type(
    exo::ExodusDatabase,
    block_id::Integer
) -> String

Exodus.read_id_map — Method

Exodus.read_ids — Method

Exodus.read_info — Method

Exodus.read_map — Method

TODO change to not use void_int

Exodus.read_name — Method

Exodus.read_name — Method

General method to read the name of a variable in index var_index for a given variable type V.

Examples: julia> readname(exo, ElementVariable, 1) "stressxx"

julia> readname(exo, GlobalVariable, 2) "reactionforce"

julia> readname(exo, NodalVariable, 1) "displx"

julia> readname(exo, NodeSetVariable, 1) "nsetdispl_x"

julia> read_name(exo, SideSetVariable, 1) "pressure"

Exodus.read_names — Method

Exodus.read_names — Method

General method to read the names of variables for a given variable type V.

Examples: julia> readnames(exo, ElementVariable) "stressxx" "stressyy" "stresszz" "stressxy" "stressyz" "stress_zx"

julia> readnames(exo, GlobalVariable) "globaldispl" "reaction_force"

julia> readnames(exo, NodalVariable) "displx" "disply" "displz"

julia> readname(exo, NodeSetVariable) "nsetdisplx" "nsetdisply" "nsetdispl_z"

julia> read_name(exo, SideSetVariable) "pressure"

Exodus.read_number_of_time_steps — Method

Exodus.read_number_of_variables — Method

General method to read the number of variables for a given variable type V.

Examples: julia> readnumberof_variables(exo, ElementVariable) 6

julia> readnumberof_variables(exo, GlobalVariable) 5

julia> readnumberof_variables(exo, NodalVariable) 3

julia> readnumberof_variables(exo, NodeSetVariable) 3

julia> readnumberof_variables(exo, SideSetVariable) 6

Exodus.read_qa — Method

Exodus.read_set — Method

Exodus.read_sets — Method

Exodus.read_side_set_node_list — Method

UNTESTED

Exodus.read_time — Method

TODO figure out how to make this not use a vector of length 1 - either a ref or a ptr

Exodus.read_times — Method

Exodus.read_values — Method

Wrapper method for global variables around the main readvalues method readvalues(exo::ExodusDatabase, t::Type{GlobalVariable}, timestep::Integer) = read_values(exo, t, timestep, 1, 1)

Example: read_values(exo, GlobalVariable, 1)

Exodus.read_values — Method

Wrapper method for nodal variables

Exodus.read_values — Method

Wrapper method for nodal variables

Exodus.read_values — Method

Exodus.read_values — Method

Exodus.read_values — Method

Method to read element variables

Exodus.read_values — Method

Method to read global variables

Exodus.read_values — Method

Method to read nodal variables

Exodus.read_values — Method

Method to read nodeset/sideset variables

Exodus.write_block — Method

write_block(exo::ExodusDatabase, block::Block)

WARNING: currently does not support edges, faces and attributes

Exodus.write_block — Method

write_block(
    exo::ExodusDatabase,
    block_id::Integer,
    elem_type::String,
    conn::Array{I<:Integer, 2}
)

Exodus.write_blocks — Method

write_blocks(exo::ExodusDatabase, blocks::Vector{<:Block})

Exodus.write_coordinates — Method

write_coordinates(
    exo::ExodusDatabase,
    coords::Union{Array{F<:AbstractFloat, 1}, Array{F<:AbstractFloat, 2}}
)

Method to write coordinates

Exodus.write_id_map — Method

Exodus.write_info — Method

Exodus.write_name — Method

Exodus.write_name — Method

Exodus.write_name — Method

Exodus.write_names — Method

WARNING: this methods likely does not have good safe guards

Exodus.write_names — Method

Exodus.write_number_of_variables — Method

General method to write the number of variables for a given variable type V.

Examples: julia> writenumberof_variables(exo, ElementVariable, 6)

julia> writenumberof_variables(exo, GlobalVariable, 5)

julia> writenumberof_variables(exo, NodalVariable, 3)

julia> writenumberof_variables(exo, NodeSetVariable, 3)

julia> writenumberof_variables(exo, SideSetVariable, 6)

Exodus.write_qa — Method

Exodus.write_set — Method

Typing ensures we don't write a set with non-matching types to the exodus file.

Exodus.write_sets — Method

Exodus.write_time — Method

Exodus.write_values — Method

Wrapper method for global variables around the main writevalues method writevalues( exo::ExodusDatabase, t::Type{GlobalVariable}, timestep::Integer, varvalues::Vector{<:AbstractFloat} ) = writevalues(exo, t, timestep, 1, 1, var_values)

Note: you need to first run writenumberof_variables(exo, GlobalVariable, n) where n is the number of variables.

Example: writenumberofvariables(exo, GlobalVariable, 5) writevalues(exo, GlobalVariable, 1, [10.0, 20.0, 30.0, 40.0, 50.0])

Exodus.write_values — Method

Wrapper for writing nodal variables by index number

Exodus.write_values — Method

Wrapper method for nodal variables

Exodus.write_values — Method

Exodus.write_values — Method

Exodus.write_values — Method

Private methods

Base.show — Method

Base.show — Method

Base.show — Method

Base.show — Method

Base.show — Method

Base.show — Method

Base.show — Method

Base.show — Method

Base.show — Method

Base.show — Method

Base.show — Method

Base.show — Method

Base.show — Method

Base.show — Method

Base.show — Method

Exodus.InitializationGlobal — Method

Exodus.collect_element_connectivities! — Method

collect_element_connectivities!(
    conns::Array{Array{B, 1}, 1},
    blocks::Array{Block{I, Array{B, 2}}, 1}
)

Exodus.collect_element_to_element_connectivities! — Method

collect_element_to_element_connectivities!(
    elem_to_elem::Array{Array{B, 1}, 1},
    node_to_elem::Array{Array{B, 1}, 1},
    conns::Array{Array{B, 1}, 1}
)

Exodus.collect_global_to_color — Function

For collecting globaltocolor

collect_global_to_color(
    file_name::String,
    n_procs::Int64
) -> Vector{Int64}
collect_global_to_color(
    file_name::String,
    n_procs::Int64,
    n_dofs::Int64
) -> Vector{Int64}

Exodus.collect_node_to_element_connectivities! — Method

collect_node_to_element_connectivities!(
    node_to_elem::Array{Array{B, 1}, 1},
    conns::Array{Array{B, 1}, 1}
)

Exodus.copy_transient — Method

copy_transient(file_name::String, new_file_name::String)

Exodus.epu_error — Method

epu_error(cmd::Cmd)

Exodus.exodiff_error — Method

exodiff_error(cmd::Cmd)

Exodus.exodus_error_check — Method

exodus_error_check(
    exo::ExodusDatabase,
    error_code::Integer,
    method_name::String
)

Exodus.exodus_error_check — Method

exodus_error_check(
    exo::Int32,
    error_code::Integer,
    method_name::String
)

Generic error handling method.

Arguments

error_code::T: error code, usually negative means something went bad

method_name::String: method name that called this

Exodus.exodus_type_check — Method

exodus_type_check(sym, context, type1, type2)

Exodus.nem_slice — Method

nem_slice() -> Any

Exodus.nem_slice — Method

nem_slice(file_name::String, n_procs::Integer) -> Any

Exodus.nem_slice_error — Method

nem_slice_error(cmd::Cmd)

Exodus.nem_spread — Method

nem_spread() -> Any

Exodus.nem_spread — Method

nem_spread(file_name::String, n_procs::Integer) -> Any

Exodus.nem_spread_error — Method

nem_spread_error(cmd::Cmd)

Exodus.open_exodus_file — Method

open_exodus_file(file_name::String, mode) -> Int32

Helper method for opening exodus database

Exodus.read_block_connectivity — Method

read_block_connectivity(
    exo::ExodusDatabase{M, I, B, F},
    block_id::Integer,
    conn_length::Integer
) -> Vector

Exodus.read_block_parameters — Method

read_block_parameters(
    exo::ExodusDatabase{M, I, B, F},
    block_id::Integer
) -> Tuple{String, Vararg{Any, 5}}

Exodus.read_blocks! — Method

read_blocks!(
    blocks::Array{B<:Block, 1},
    exo::ExodusDatabase,
    block_ids::Array{I<:Integer, 1}
)

TODO: change name to readelementblocks!

Exodus.read_coordinate_names — Method

read_coordinate_names(exo::ExodusDatabase) -> Any

Method to read coordinates names Returns a vector of strings

Exodus.read_ghost_nodes_and_procs — Method

read_ghost_nodes_and_procs(rank, exo) -> Tuple{Any, Any}

Exodus.read_init_info — Method

Exodus.read_internal_nodes_and_procs — Method

read_internal_nodes_and_procs(rank, exo) -> Tuple{Any, Any}

Exodus.read_node_cmaps — Method

read_node_cmaps(rank, exo) -> Any

Exodus.read_node_set_nodes — Method

Exodus.read_partial_block_connectivity — Method

read_partial_block_connectivity(
    exo::ExodusDatabase,
    block_id::Integer,
    start_num::Integer,
    num_ent::Integer
) -> Any

Exodus.read_partial_coordinates — Method

read_partial_coordinates(
    exo::ExodusDatabase,
    start_node_num::Integer,
    n_nodes::Integer
) -> Matrix

Method to read a partial set of coordinates that are contiguous. Returns a matrix that is ndim x nnodes

Exodus.read_partial_coordinates_component — Method

read_partial_coordinates_component(
    exo::ExodusDatabase,
    start_node_num::Integer,
    n_nodes::Integer,
    component::Integer
) -> Vector

Method to read a specific component of a partial set of coordinates that are contiguous. Returns a vector of length n_nodes TODO change to not use Cvoid

Exodus.read_partial_coordinates_component — Method

read_partial_coordinates_component(
    exo::ExodusDatabase,
    start_node_num::Integer,
    n_nodes::Integer,
    component::String
) -> Vector

Exodus.read_set_parameters — Method

Exodus.read_sets! — Method

Exodus.read_side_set_elements_and_sides — Method

Exodus.set_exodus_max_name_length — Method

Exodus.set_exodus_options — Method

Exodus.write_block_connectivity — Method

write_block_connectivity(
    exo::ExodusDatabase,
    block_id::Integer,
    conn::Array{I<:Integer, 2}
)

Exodus.write_coordinate_names — Method

write_coordinate_names(
    exo::ExodusDatabase,
    coord_names::Vector{String}
)

Method to write coordinate names, e.g. x, y, z

Exodus.write_initialization! — Method

write_initialization!(exoid::Int32, init::Initialization)

Used to set up a exodus database in write mode

The ccall signatures should reall be B (bulk int type of exo) instead of Clonglong

Exodus.write_partial_coordinates — Method

write_partial_coordinates(
    exo::ExodusDatabase,
    start_node_num::Integer,
    coords::Union{Array{F<:AbstractFloat, 1}, Array{F<:AbstractFloat, 2}}
)

Exodus.write_partial_coordinates_component — Method

write_partial_coordinates_component(
    exo::ExodusDatabase,
    start_node_num::Integer,
    component::Integer,
    coords::Array{F<:AbstractFloat, 1}
)

Exodus.write_partial_coordinates_component — Method

write_partial_coordinates_component(
    exo::ExodusDatabase,
    start_node_num::Integer,
    component::String,
    coords::Array{F<:AbstractFloat, 1}
)

Exodus.write_set_parameters — Method

WARNING: currently doesn't support distance factors