Exodus.jl Types

Public methods

Exodus.Block — Type

struct Block{I, A<:(AbstractMatrix)} <: Exodus.AbstractExodusSet{I, A<:(AbstractMatrix)}

id::Any

num_elem::Int64

num_nodes_per_elem::Int64

elem_type::String

conn::AbstractMatrix

Exodus.Block — Method

Block(exo::ExodusDatabase, block_id::Integer) -> Block

Init method for block container.

Exodus.Block — Method

Block(exo::ExodusDatabase, name::String) -> Block

Exodus.EdgeMap — Type

struct EdgeMap <: Exodus.AbstractExodusMap

Exodus.ElementMap — Type

struct ElementMap <: Exodus.AbstractExodusMap

Exodus.ElementVariable — Type

struct ElementVariable <: Exodus.AbstractExodusVariable

Exodus.ExodusDatabase — Type

struct ExodusDatabase{M, I, B, F}

exo::Int32

mode::String

file_name::String

init::Initialization

block_name_dict::Dict{String}

nset_name_dict::Dict{String}

sset_name_dict::Dict{String}

element_var_name_dict::Dict{String}

global_var_name_dict::Dict{String}

nodal_var_name_dict::Dict{String}

nset_var_name_dict::Dict{String}

sset_var_name_dict::Dict{String}

Exodus.ExodusDatabase — Method

ExodusDatabase(
    f::Function,
    file_name::String,
    mode::String
) -> Any

Method that safely does stuff by ensuring the file closes no matter what.

Exodus.ExodusDatabase — Method

ExodusDatabase(
    file_name::String,
    mode::String
) -> ExodusDatabase

Type unstable helper to eliminate annoying lines of code to get type stability.

If you're looking for a type stable way to to open an exodus file, Simple copy past some of this into a barrier function

Exodus.ExodusDatabase — Method

Exodus.ExodusDatabase — Method

ExodusDatabase(
    file_name::String,
    mode::String
) -> ExodusDatabase

Exodus.ExodusDatabase — Method

ExodusDatabase(
    file_name::String,
    mode::String
) -> ExodusDatabase

Exodus.ExodusDatabase — Method

Exodus.FaceMap — Type

struct FaceMap <: Exodus.AbstractExodusMap

Exodus.GlobalVariable — Type

struct GlobalVariable <: Exodus.AbstractExodusVariable

Exodus.Initialization — Type

struct Initialization{B}

Type that holds initialization information.

Exodus.Initialization — Method

Initialization(exo::Int32, _::Type{B}) -> Initialization

Exodus.Initialization — Method

Initialization(
    exo::ExodusDatabase{M, I, B, F}
) -> Initialization

Exodus.Initialization — Method

Initialization(_::Type{B<:Integer}) -> Initialization

Initialization(_)

Exodus.NodalVariable — Type

struct NodalVariable <: Exodus.AbstractExodusVariable

Exodus.NodalVectorVariable — Type

struct NodalVectorVariable <: Exodus.AbstractExodusVariable

Exodus.NodeMap — Type

struct NodeMap <: Exodus.AbstractExodusMap

Exodus.NodeSet — Type

struct NodeSet{I, A<:(AbstractVector)} <: Exodus.AbstractExodusSet{I, A<:(AbstractVector)}

id::Any

nodes::AbstractVector

Exodus.NodeSet — Method

NodeSet(exo::ExodusDatabase, name::String) -> NodeSet

Exodus.NodeSet — Method

NodeSet(
    exo::ExodusDatabase{M, I, B, F},
    id::Integer
) -> NodeSet

Exodus.NodeSetVariable — Type

struct NodeSetVariable <: Exodus.AbstractExodusVariable

Exodus.SideSet — Type

struct SideSet{I, A<:(AbstractVector)} <: Exodus.AbstractExodusSet{I, A<:(AbstractVector)}

id::Any

elements::AbstractVector

sides::AbstractVector

num_nodes_per_side::AbstractVector

side_nodes::AbstractVector

Exodus.SideSet — Method

SideSet(exo::ExodusDatabase, name::String) -> SideSet

Exodus.SideSet — Method

SideSet(
    exo::ExodusDatabase{M, I, B, F},
    id::Integer
) -> SideSet

Exodus.SideSetVariable — Type

struct SideSetVariable <: Exodus.AbstractExodusVariable

Private methods

Exodus.AbstractExodusMap — Type

abstract type AbstractExodusMap <: Exodus.AbstractExodusType

Exodus.AbstractExodusSet — Type

abstract type AbstractExodusSet{I, A} <: Exodus.AbstractExodusType

Exodus.AbstractExodusType — Type

abstract type AbstractExodusType

Exodus.AbstractExodusVariable — Type

abstract type AbstractExodusVariable <: Exodus.AbstractExodusType

Exodus.CommunicationMapParameters — Type

struct CommunicationMapParameters{B}

node_cmap_ids::Vector

node_cmap_node_cnts::Vector

elem_cmap_ids::Vector

elem_cmap_elem_cnts::Vector

Exodus.CommunicationMapParameters — Method

CommunicationMapParameters(
    exo::ExodusDatabase{M, I, B, F},
    lb_params::Exodus.LoadBalanceParameters{B},
    processor::Integer
) -> Exodus.CommunicationMapParameters

Exodus.EPUException — Type

struct EPUException <: Exception

cmd::Cmd

Exodus.ElementCommunicationMap — Type

struct ElementCommunicationMap{B}

elem_ids::Vector

side_ids::Vector

proc_ids::Vector

Exodus.ElementCommunicationMap — Method

ElementCommunicationMap(
    exo::ExodusDatabase{M, I, B, F},
    elem_map_id::Integer,
    elem_cnt,
    processor::Integer
) -> Exodus.ElementCommunicationMap

Exodus.ExodiffException — Type

struct ExodiffException <: Exception

cmd::Cmd

Exodus.ExodusError — Type

struct ExodusError <: Exception

error_code::Int32

error_msg::String

Exodus.ExodusWindowsError — Type

struct ExodusWindowsError <: Exception

Exodus.LoadBalanceParameters — Type

struct LoadBalanceParameters{B}

num_int_nodes::Any

num_bor_nodes::Any

num_ext_nodes::Any

num_int_elems::Any

num_bor_elems::Any

num_node_cmaps::Any

num_elem_cmaps::Any

processor::Int32

Exodus.LoadBalanceParameters — Method

LoadBalanceParameters(
    exo::ExodusDatabase{M, I, B, F},
    processor::Integer
) -> Exodus.LoadBalanceParameters

Exodus.ModeException — Type

struct ModeException <: Exception

mode::String

Exodus.NemSliceException — Type

struct NemSliceException <: Exception

cmd::Cmd

Exodus.NemSpreadException — Type

struct NemSpreadException <: Exception

cmd::Cmd

Exodus.NodalScalarVariable — Type

struct NodalVariable <: Exodus.AbstractExodusVariable

Exodus.NodeCommunicationMap — Type

struct NodeCommunicationMap{B}

node_ids::Vector

proc_ids::Vector

Exodus.NodeCommunicationMap — Method

NodeCommunicationMap(
    exo::ExodusDatabase{M, I, B, F},
    node_map_id,
    node_cnt,
    processor
) -> Exodus.NodeCommunicationMap

Exodus.ProcessorElementMaps — Type

struct ProcessorElementMaps{B}

elem_map_internal::Vector

elem_map_border::Vector

Exodus.ProcessorElementMaps — Method

ProcessorElementMaps(
    exo::ExodusDatabase{M, I, B, F},
    processor
) -> Exodus.ProcessorElementMaps

Exodus.ProcessorNodeMaps — Type

struct ProcessorNodeMaps{B}

node_map_internal::Vector

node_map_border::Vector

node_map_external::Vector

Exodus.ProcessorNodeMaps — Method

ProcessorNodeMaps(
    exo::ExodusDatabase{M, I, B, F},
    processor
) -> Exodus.ProcessorNodeMaps

Exodus.SetIDException — Type

struct SetIDException{M, I, B, F, V, I1<:Integer} <: Exception

exo::ExodusDatabase

type::Type

id::Integer

Exodus.SetNameException — Type

struct SetNameException{M, I, B, F, V} <: Exception

exo::ExodusDatabase

type::Type

name::String

Exodus.VariableIDException — Type

struct VariableIDException{M, I, B, F, V} <: Exception

exo::ExodusDatabase

type::Type

id::Int64

Exodus.VariableNameException — Type

struct VariableNameException{M, I, B, F, V} <: Exception

exo::ExodusDatabase

type::Type

name::String

Exodus.ex_entity_id — Type

Exodus.ex_entity_type — Type

ex_entity_type

Entity type enums (exentitytype in exodusII.h)

Exodus.ex_inquiry — Type

ex_inquiry

Inquiry enums (ex_inquiry in exodusII.h).

Exodus.void_int — Type

void_int = Cvoid