ExodusUnitfulExt

This extension can be utilized to add units to values read in from exodusII files.

Exodus.read_coordinates — Method

read_coordinates(exo, unit)

Exodus.read_times — Method

read_times(
    exo::ExodusDatabase,
    unit::Unitful.FreeUnits
) -> Any

Exodus.read_values — Method

read_values(
    exo::ExodusDatabase,
    t::Type{GlobalVariable},
    timestep::Integer,
    unit::Unitful.FreeUnits
) -> Any

Wrapper method for global variables around the main readvalues method readvalues(exo::ExodusDatabase, t::Type{GlobalVariable}, timestep::Integer) = read_values(exo, t, timestep, 1, 1)

Example: read_values(exo, GlobalVariable, 1)

Exodus.read_values — Method

Wrapper method for nodal variables

Exodus.read_values — Method

read_values(
    exo::ExodusDatabase,
    t::Type{NodalVariable},
    timestep::Integer,
    name::String,
    unit::Unitful.FreeUnits
) -> Any

Wrapper method for nodal variables

Exodus.read_values — Method

read_values(
    exo::ExodusDatabase,
    type::Type{V<:Exodus.AbstractExodusVariable},
    timestep::Integer,
    id::Integer,
    var_index::Integer,
    unit::Unitful.FreeUnits
) -> Any

Exodus.read_values — Method

read_values(
    exo::ExodusDatabase,
    _::Type{V<:Union{ElementVariable, NodalVariable, NodeSetVariable, SideSetVariable}},
    time_step::Integer,
    id::Integer,
    var_name::String,
    unit::Unitful.FreeUnits
) -> Any

Exodus.read_values — Method

read_values(
    exo::ExodusDatabase,
    _::Type{V<:Union{ElementVariable, NodeSetVariable, SideSetVariable}},
    time_step::Integer,
    set_name::String,
    var_name::String,
    unit::Unitful.FreeUnits
) -> Any

Exodus.write_values — Method

write_values(
    exo::ExodusDatabase,
    t::Type{GlobalVariable},
    timestep::Integer,
    var_values::Vector{<:Quantity}
)

Wrapper method for global variables around the main writevalues method writevalues( exo::ExodusDatabase, t::Type{GlobalVariable}, timestep::Integer, varvalues::Vector{<:AbstractFloat} ) = writevalues(exo, t, timestep, 1, 1, var_values)

Note: you need to first run writenumberof_variables(exo, GlobalVariable, n) where n is the number of variables.

Example: writenumberofvariables(exo, GlobalVariable, 5) writevalues(exo, GlobalVariable, 1, [10.0, 20.0, 30.0, 40.0, 50.0])

Exodus.write_values — Method

Wrapper for writing nodal variables by index number

Exodus.write_values — Method

write_values(
    exo::ExodusDatabase,
    t::Type{NodalVariable},
    timestep::Integer,
    var_name::String,
    var_values::Vector{<:Quantity}
) -> Any

Wrapper method for nodal variables

Exodus.write_values — Method

write_values(
    exo::ExodusDatabase,
    type::Type{V<:Exodus.AbstractExodusVariable},
    timestep::Integer,
    id::Integer,
    var_index::Integer,
    var_values::Vector{<:Quantity}
)

Exodus.write_values — Method

write_values(
    exo::ExodusDatabase,
    type::Type{V<:Exodus.AbstractExodusVariable},
    timestep::Integer,
    id::Integer,
    var_name::String,
    var_value::Vector{<:Quantity}
) -> Any

Exodus.write_values — Method

write_values(
    exo::ExodusDatabase,
    _::Type{V<:Exodus.AbstractExodusVariable},
    time_step::Integer,
    set_name::String,
    var_name::String,
    var_value::Vector{<:Quantity}
) -> Any