Exodus

Documentation for Exodus.

Exodus.Exodus

Exodus.Block

Exodus.Block

Exodus.Block

Exodus.Element

Exodus.ExodusDatabase

Exodus.ExodusDatabase

Exodus.ExodusDatabase

Exodus.Global

Exodus.Initialization

Exodus.Initialization

Exodus.Initialization

Exodus.Nodal

Exodus.NodeSet

Exodus.NodeSet

Exodus.NodeSet

Exodus.NodeSetVariable

Exodus.SideSet

Exodus.SideSet

Exodus.SideSet

Exodus.SideSetVariable

Exodus.ex_entity_id

Exodus.ex_entity_type

Exodus.ex_inquiry

Exodus.void_int

Base.close

Base.copy

Base.length

Base.length

Base.show

Base.show

Base.show

Base.show

Exodus.collect_block_connectivities

Exodus.decomp

Exodus.exo_float_type

Exodus.exo_int_types

Exodus.exodus_error_check

Exodus.read_block

Exodus.read_block

Exodus.read_block_connectivity

Exodus.read_block_id_map

Exodus.read_block_parameters

Exodus.read_blocks

Exodus.read_blocks!

Exodus.read_coordinate_names

Exodus.read_coordinates

Exodus.read_element_type

Exodus.read_ids

Exodus.read_info

Exodus.read_map

Exodus.read_name

Exodus.read_name

Exodus.read_names

Exodus.read_names

Exodus.read_node_set_nodes

Exodus.read_number_of_time_steps

Exodus.read_number_of_variables

Exodus.read_partial_block_connectivity

Exodus.read_partial_coordinates

Exodus.read_partial_coordinates_component

Exodus.read_partial_values

Exodus.read_partial_values

Exodus.read_qa

Exodus.read_set

Exodus.read_sets

Exodus.read_sets!

Exodus.read_side_set_elements_and_sides

Exodus.read_side_set_node_list

Exodus.read_time

Exodus.read_times

Exodus.read_values

Exodus.read_values

Exodus.read_values

Exodus.set_exodus_max_name_length

Exodus.set_exodus_options

Exodus.write_block

Exodus.write_block_connectivity

Exodus.write_blocks

Exodus.write_coordinate_names

Exodus.write_coordinates

Exodus.write_info

Exodus.write_initialization!

Exodus.write_name

Exodus.write_name

Exodus.write_names

Exodus.write_names

Exodus.write_number_of_variables

Exodus.write_partial_coordinates

Exodus.write_partial_coordinates_component

Exodus.write_partial_coordinates_component

Exodus.write_qa

Exodus.write_set

Exodus.write_set_parameters

Exodus.write_sets

Exodus.write_time

Exodus.write_values

Exodus.write_values

Exodus.write_values

Exodus.@decomp

Exodus.@epu

Exodus.@exodiff

Exodus.Exodus — Module

Block

Element

ExodusDatabase

Global

Initialization

Nodal

NodeSet

NodeSetVariable

SideSet

SideSetVariable

close

collect_block_connectivities

copy

@decomp

decomp

@epu

@exodiff

length

read_block

read_block_id_map

read_blocks

read_coordinates

read_element_type

read_ids

read_info

read_map

read_name

read_names

read_number_of_time_steps

read_number_of_variables

read_qa

read_set

read_sets

read_time

read_times

read_values

write_block

write_blocks

write_coordinates

write_info

write_name

write_names

write_number_of_variables

write_qa

write_set

write_sets

write_time

write_values

Exodus.Block — Type

id::Any

num_elem::Int64

num_nodes_per_elem::Int64

elem_type::String

conn::Matrix

Exodus.Block — Method

Init method for block container.

Exodus.Block — Method

Exodus.Element — Type

Exodus.ExodusDatabase — Method

Exodus.ExodusDatabase — Method

Init method for read/read-write

Exodus.ExodusDatabase — Method

Exodus.Global — Type

Exodus.Initialization — Method

Workaround method

Exodus.Initialization — Method

Workaround method

Exodus.Initialization — Method

Exodus.Nodal — Type

Exodus.NodeSet — Type

id::Any

nodes::Vector

Exodus.NodeSet — Method

Exodus.NodeSet — Method

Exodus.NodeSetVariable — Type

Exodus.SideSet — Type

id::Any

elements::Vector

sides::Vector

Exodus.SideSet — Method

Exodus.SideSet — Method

Exodus.SideSetVariable — Type

Exodus.ex_entity_id — Type

Exodus.ex_entity_type — Type

ex_entity_type

Entity type enums (exentitytype in exodusII.h)

Exodus.ex_inquiry — Type

ex_inquiry

Inquiry enums (ex_inquiry in exodusII.h).

Exodus.void_int — Type

void_int = Cvoid

Base.close — Method

close(exo::ExodusDatabase)

Used to close and ExodusDatabase.

close(exo)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/IO.jl:151.

Base.copy — Method

copy(exo::ExodusDatabase, new_file_name::String)

Used to copy an ExodusDatabase. As of right now this is the best way to create a new ExodusDatabase for output. Not all of the put methods have been wrapped and properly tested. This one has though.

copy(exo, new_file_name)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/IO.jl:160.

Base.length — Method

length(n::NodeSet) -> Int64

length(n)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExodusTypes.jl:118.

Base.length — Method

length(s::SideSet) -> Int64

length(s)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExodusTypes.jl:164.

Base.show — Method

show(io::IO, init::Initialization)

show(io, init)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Initializations.jl:41.

Base.show — Method

show(io::IO, node_set::NodeSet)

show(io, node_set)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExodusTypes.jl:122.

Base.show — Method

show(io::IO, sset::SideSet)

show(io, sset)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExodusTypes.jl:168.

Base.show — Method

show(io::IO, block::Block)

show(io, block)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExodusTypes.jl:79.

Exodus.collect_block_connectivities — Method

collect_block_connectivities(exo::ExodusDatabase) -> Any

TODO reduce allocations and check for type instabilities

collect_block_connectivities(exo)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Blocks.jl:169.

Exodus.decomp — Method

decomp(file_name::String, n_procs::Integer) -> Any

decomp(file_name, n_procs)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Decomp.jl:85.

Exodus.exo_float_type — Method

exo_float_type(
    exoid::Int32
) -> Union{Type{Float32}, Type{Float64}}

exo_float_type(exoid)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/IO.jl:34.

Exodus.exo_int_types — Method

exo_int_types(
    exoid::Int32
) -> Tuple{DataType, DataType, DataType}

exo_int_types(exoid)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/IO.jl:19.

Exodus.exodus_error_check — Method

exodus_error_check(error_code::Integer, method_name::String)

Generic error handling method.

Arguments

error_code::T: error code, usually negative means something went bad

method_name::String: method name that called this

exodus_error_check(error_code, method_name)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Errors.jl:7.

Exodus.read_block — Method

read_block(exo::ExodusDatabase, block_id::Integer) -> Block

read_block(exo, block_id)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Blocks.jl:94.

Exodus.read_block — Method

read_block(exo::ExodusDatabase, block_name::String) -> Block

read_block(exo, block_name)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Blocks.jl:98.

Exodus.read_block_connectivity — Method

read_block_connectivity(
    exo::ExodusDatabase{M, I, B, F},
    block_id::Integer
) -> Any

read_block_connectivity(exo, block_id)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Blocks.jl:35.

Exodus.read_block_id_map — Method

read_block_id_map(
    exo::ExodusDatabase,
    block_id::Integer
) -> Vector

read_block_id_map(exo, block_id)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Blocks.jl:3.

Exodus.read_block_parameters — Method

read_block_parameters(
    exo::ExodusDatabase{M, I, B, F},
    block_id::Integer
) -> Tuple{String, Vararg{Any, 5}}

read_block_parameters(exo, block_id)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Blocks.jl:15.

Exodus.read_blocks! — Method

read_blocks!(
    blocks::Array{B<:Block, 1},
    exo::ExodusDatabase,
    block_ids::Array{I<:Integer, 1}
)

TODO: change name to readelementblocks!

read_blocks!(blocks, exo, block_ids)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Blocks.jl:103.

Exodus.read_blocks — Method

read_blocks(
    exo::ExodusDatabase,
    block_ids::Union{Vector{<:Integer}, var"#s16"} where var"#s16"<:Integer
) -> Any

Helper method for initializing blocks.

TODO: change name to readelementblocks

read_blocks(exo, block_ids)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Blocks.jl:116.

Exodus.read_coordinate_names — Method

read_coordinate_names(exo::ExodusDatabase) -> Any

Method to read coordinates names Returns a vector of strings

read_coordinate_names(exo)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Coordinates.jl:117.

Exodus.read_coordinates — Method

read_coordinates(exo::ExodusDatabase) -> Matrix

Method to read coordinates. Returns a matrix that is ndim x nnodes.

read_coordinates(exo)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Coordinates.jl:5.

Exodus.read_element_type — Method

read_element_type(
    exo::ExodusDatabase,
    block_id::Integer
) -> String

read_element_type(exo, block_id)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Blocks.jl:83.

Exodus.read_ids — Method

read_ids(
    exo::ExodusDatabase{M, I, B, F},
    _::Type{S<:Exodus.AbstractSet}
) -> Any

read_ids(exo, _)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:3.

Exodus.read_info — Method

read_info(exo::ExodusDatabase) -> Vector{String}

read_info(exo)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Info.jl:3.

Exodus.read_map — Method

read_map(exo::ExodusDatabase) -> Any

read_map(exo)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Maps.jl:3.

Exodus.read_name — Method

read_name(
    exo::ExodusDatabase,
    _::Type{S<:Exodus.AbstractSet},
    id::Integer
) -> String

read_name(exo, _, id)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:21.

Exodus.read_name — Method

read_name(
    exo::ExodusDatabase,
    _::Type{V<:Exodus.AbstractVariable},
    var_index::Integer
) -> String

read_name(exo, _, var_index)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:14.

Exodus.read_names — Method

read_names(
    exo::ExodusDatabase,
    _::Type{S<:Exodus.AbstractSet}
) -> Any

read_names(exo, _)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:33.

Exodus.read_names — Method

read_names(
    exo::ExodusDatabase,
    _::Type{V<:Exodus.AbstractVariable}
) -> Vector{String}

read_names(exo, _)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:25.

Exodus.read_node_set_nodes — Method

read_node_set_nodes(
    exo::ExodusDatabase{M, I, B, F},
    set_id::Integer
) -> Any

read_node_set_nodes(exo, set_id)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:59.

Exodus.read_number_of_time_steps — Method

read_number_of_time_steps(exo::ExodusDatabase) -> UInt32

read_number_of_time_steps(exo)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Times.jl:3.

Exodus.read_number_of_variables — Method

read_number_of_variables(
    exo::ExodusDatabase,
    _::Type{V<:Exodus.AbstractVariable}
) -> Int32

read_number_of_variables(exo, _)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:3.

Exodus.read_partial_block_connectivity — Method

read_partial_block_connectivity(
    exo::ExodusDatabase,
    block_id::Integer,
    start_num::Integer,
    num_ent::Integer
) -> Any

read_partial_block_connectivity(
    exo,
    block_id,
    start_num,
    num_ent
)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Blocks.jl:66.

Exodus.read_partial_coordinates — Method

read_partial_coordinates(
    exo::ExodusDatabase,
    start_node_num::Integer,
    num_nodes::Integer
) -> Matrix

Method to read a partial set of coordinates that are contiguous. Returns a matrix that is ndim x nnodes

read_partial_coordinates(exo, start_node_num, num_nodes)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Coordinates.jl:43.

Exodus.read_partial_coordinates_component — Method

read_partial_coordinates_component(
    exo::ExodusDatabase,
    start_node_num::Integer,
    num_nodes::Integer,
    component::Integer
) -> Vector

Method to read a specific component of a partial set of coordinates that are contiguous. Returns a vector of length n_nodes

read_partial_coordinates_component(
    exo,
    start_node_num,
    num_nodes,
    component
)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Coordinates.jl:90.

Exodus.read_partial_values — Method

read_partial_values(
    exo::ExodusDatabase,
    _::Type{V<:Exodus.AbstractVariable},
    time_step::Integer,
    id::Integer,
    var_name::String,
    start_node::Integer,
    num_nodes::Integer
) -> Vector

read_partial_values(
    exo,
    _,
    time_step,
    id,
    var_name,
    start_node,
    num_nodes
)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:127.

Exodus.read_partial_values — Method

read_partial_values(
    exo::ExodusDatabase{M, I, B, F},
    _::Type{V<:Exodus.AbstractVariable},
    time_step::Integer,
    id::Integer,
    var_index::Integer,
    start_node::Integer,
    num_nodes::Integer
) -> Vector

read_partial_values(
    exo,
    _,
    time_step,
    id,
    var_index,
    start_node,
    num_nodes
)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:107.

Exodus.read_qa — Method

read_qa(exo::ExodusDatabase) -> Matrix{String}

read_qa(exo)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/QA.jl:3.

Exodus.read_set — Method

read_set(
    exo::ExodusDatabase,
    type::Type{S<:Exodus.AbstractSet},
    set_id
) -> Any

read_set(exo, type, set_id)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:107.

Exodus.read_sets! — Method

read_sets!(
    sets::Array{T<:Exodus.AbstractSet, 1},
    exo::ExodusDatabase,
    set_ids::Array{I, 1}
)

read_sets!(sets, exo, set_ids)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:113.

Exodus.read_sets — Method

read_sets(
    exo::ExodusDatabase{M, I, B, F},
    type::Type{S<:Exodus.AbstractSet}
) -> Any

read_sets(exo, type)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:127.

Exodus.read_side_set_elements_and_sides — Method

read_side_set_elements_and_sides(
    exo::ExodusDatabase{M, I, B, F},
    set_id::Integer
) -> Tuple{Any, Any}

read_side_set_elements_and_sides(exo, set_id)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:73.

Exodus.read_side_set_node_list — Method

read_side_set_node_list(
    exo::ExodusDatabase{M, I, B, F},
    side_set_id::Integer
) -> Tuple{Any, Vector}

UNTESTED

read_side_set_node_list(exo, side_set_id)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:88.

Exodus.read_time — Method

read_time(exo::ExodusDatabase, time_step::Integer) -> Any

TODO figure out how to make this not use a vector of length 1 - either a ref or a ptr

read_time(exo, time_step)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Times.jl:12.

Exodus.read_times — Method

read_times(exo::ExodusDatabase) -> Vector

read_times(exo)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Times.jl:23.

Exodus.read_values — Method

read_values(
    exo::ExodusDatabase,
    _::Type{V<:Exodus.AbstractVariable},
    time_step::Integer,
    id::Integer,
    var_name::String
) -> Any

read_values(exo, _, time_step, id, var_name)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:72.

Exodus.read_values — Method

read_values(
    exo::ExodusDatabase,
    _::Type{V<:Exodus.AbstractVariable},
    time_step::Integer,
    set_name::String,
    var_name::String
) -> Any

read_values(exo, _, time_step, set_name, var_name)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:86.

Exodus.read_values — Method

read_values(
    exo::ExodusDatabase{M, I, B, F},
    _::Type{V<:Exodus.AbstractVariable},
    timestep::Integer,
    id::Integer,
    var_index::Integer
) -> Any

read_values(exo, _, timestep, id, var_index)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:45.

Exodus.set_exodus_max_name_length — Method

set_exodus_max_name_length(exoid::Int32, len::Int32)

set_exodus_max_name_length(exoid, len)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/IO.jl:10.

Exodus.set_exodus_options — Method

set_exodus_options(options)

set_exodus_options(options)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/IO.jl:3.

Exodus.write_block — Method

write_block(exo::ExodusDatabase, block::Block)

WARNING: currently does not support edges, faces and attributes

write_block(exo, block)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Blocks.jl:133.

Exodus.write_block_connectivity — Method

write_block_connectivity(
    exo::ExodusDatabase,
    block_id::Integer,
    conn::Array{I<:Integer, 2}
)

write_block_connectivity(exo, block_id, conn)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Blocks.jl:49.

Exodus.write_blocks — Method

write_blocks(exo::ExodusDatabase, blocks::Vector{<:Block})

write_blocks(exo, blocks)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Blocks.jl:159.

Exodus.write_coordinate_names — Method

write_coordinate_names(
    exo::ExodusDatabase,
    coord_names::Vector{String}
)

Method to write coordinate names, e.g. x, y, z

write_coordinate_names(exo, coord_names)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Coordinates.jl:176.

Exodus.write_coordinates — Method

write_coordinates(
    exo::ExodusDatabase,
    coords::Union{Array{F<:AbstractFloat, 1}, Array{F<:AbstractFloat, 2}}
)

Method to write coordinates

write_coordinates(exo, coords)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Coordinates.jl:135.

Exodus.write_info — Method

write_info(exo::ExodusDatabase, info::Vector{String})

write_info(exo, info)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Info.jl:24.

Exodus.write_initialization! — Method

write_initialization!(exoid::Int32, init::Initialization)

Used to set up a exodus database in write mode

The ccall signatures should reall be B (bulk int type of exo) instead of Clonglong

write_initialization!(exoid, init)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Initializations.jl:57.

Exodus.write_name — Method

write_name(
    exo::ExodusDatabase{M, I, B, F},
    set::Exodus.AbstractSet,
    name::String
)

write_name(exo, set, name)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:170.

Exodus.write_name — Method

write_name(
    exo::ExodusDatabase,
    _::Type{V<:Exodus.AbstractVariable},
    var_index::Integer,
    var_name::String
)

write_name(exo, _, var_index, var_name)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:152.

Exodus.write_names — Method

write_names(
    exo::ExodusDatabase,
    _::Type{S<:Exodus.AbstractSet},
    names::Vector{String}
)

write_names(exo, _, names)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:180.

Exodus.write_names — Method

write_names(
    exo::ExodusDatabase,
    _::Type{V<:Exodus.AbstractVariable},
    var_names::Vector{String}
)

write_names(exo, _, var_names)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:162.

Exodus.write_number_of_variables — Method

write_number_of_variables(
    exo::ExodusDatabase,
    _::Type{V<:Exodus.AbstractVariable},
    num_vars::Integer
)

write_number_of_variables(exo, _, num_vars)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:143.

Exodus.write_partial_coordinates — Method

write_partial_coordinates(
    exo::ExodusDatabase,
    start_node_num::Integer,
    coords::Union{Array{F<:AbstractFloat, 1}, Array{F<:AbstractFloat, 2}}
)

write_partial_coordinates(exo, start_node_num, coords)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Coordinates.jl:187.

Exodus.write_partial_coordinates_component — Method

write_partial_coordinates_component(
    exo::ExodusDatabase,
    start_node_num::Integer,
    component::Integer,
    coords::Array{F<:AbstractFloat, 1}
)

write_partial_coordinates_component(
    exo,
    start_node_num,
    component,
    coords
)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Coordinates.jl:214.

Exodus.write_partial_coordinates_component — Method

write_partial_coordinates_component(
    exo::ExodusDatabase,
    start_node_num::Integer,
    component::String,
    coords::Array{F<:AbstractFloat, 1}
)

write_partial_coordinates_component(
    exo,
    start_node_num,
    component,
    coords
)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Coordinates.jl:225.

Exodus.write_qa — Method

write_qa(exo::ExodusDatabase, qa_record::Matrix{String})

write_qa(exo, qa_record)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/QA.jl:29.

Exodus.write_set — Method

write_set(
    exo::ExodusDatabase{M, I, B, F},
    set::Exodus.AbstractSet
)

Typing ensures we don't write a set with non-matching types to the exodus file.

write_set(exo, set)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:151.

Exodus.write_set_parameters — Method

write_set_parameters(
    exo::ExodusDatabase{M, I, B, F},
    set::Exodus.AbstractSet
)

WARNING: currently doesn't support distance factors

write_set_parameters(exo, set)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:138.

Exodus.write_sets — Method

write_sets(
    exo::ExodusDatabase,
    sets::Array{T<:Exodus.AbstractSet, 1}
)

write_sets(exo, sets)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:162.

Exodus.write_time — Method

write_time(
    exo::ExodusDatabase,
    time_step::Integer,
    time_value::AbstractFloat
)

write_time(exo, time_step, time_value)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Times.jl:33.

Exodus.write_values — Method

write_values(
    exo::ExodusDatabase,
    _::Type{V<:Exodus.AbstractVariable},
    timestep::Integer,
    id::Integer,
    var_index::Integer,
    var_values::Vector{<:AbstractFloat}
)

write_values(exo, _, timestep, id, var_index, var_values)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:172.

Exodus.write_values — Method

write_values(
    exo::ExodusDatabase,
    _::Type{V<:Exodus.AbstractVariable},
    time_step::Integer,
    id::Integer,
    var_name::String,
    var_value::Vector{<:AbstractFloat}
)

write_values(exo, _, time_step, id, var_name, var_value)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:190.

Exodus.write_values — Method

write_values(
    exo::ExodusDatabase,
    _::Type{V<:Exodus.AbstractVariable},
    time_step::Integer,
    var_name::String,
    set_name::String,
    var_value::Vector{<:AbstractFloat}
)

write_values(
    exo,
    _,
    time_step,
    var_name,
    set_name,
    var_value
)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:207.

Exodus.@decomp — Macro

Exodus.@epu — Macro

Exodus.@exodiff — Macro