Exodus

Documentation for Exodus. <!–

Exodus.Exodus

Exodus.Block

Exodus.Block

Exodus.Block

Exodus.CommunicationMapParameters

Exodus.CommunicationMapParameters

Exodus.EPUException

Exodus.EdgeMap

Exodus.ElementCommunicationMap

Exodus.ElementCommunicationMap

Exodus.ElementMap

Exodus.ElementVariable

Exodus.ExodiffException

Exodus.ExodusDatabase

Exodus.ExodusDatabase

Exodus.ExodusError

Exodus.ExodusWindowsError

Exodus.FaceMap

Exodus.GlobalVariable

Exodus.Initialization

Exodus.Initialization

Exodus.Initialization

Exodus.LoadBalanceParameters

Exodus.LoadBalanceParameters

Exodus.ModeException

Exodus.NemSliceException

Exodus.NemSpreadException

Exodus.NodalVariable

Exodus.NodeCommunicationMap

Exodus.NodeCommunicationMap

Exodus.NodeMap

Exodus.NodeSet

Exodus.NodeSet

Exodus.NodeSet

Exodus.NodeSetVariable

Exodus.ParallelExodusDatabase

Exodus.ParallelExodusDatabase

Exodus.ProcessorElementMaps

Exodus.ProcessorElementMaps

Exodus.ProcessorNodeMaps

Exodus.ProcessorNodeMaps

Exodus.SetIDException

Exodus.SetNameException

Exodus.SideSet

Exodus.SideSet

Exodus.SideSet

Exodus.SideSetVariable

Exodus.VariableIDException

Exodus.VariableNameException

Exodus.ex_entity_id

Exodus.ex_entity_type

Exodus.ex_inquiry

Exodus.void_int

Base.close

Base.close

Base.copy

Base.length

Base.length

Base.show

Base.show

Base.show

Base.show

Base.show

Base.show

Base.show

Base.show

Base.show

Base.show

Base.show

Base.show

Base.show

Base.show

Base.show

Base.show

Exodus.InitializationGlobal

Exodus.collect_element_connectivities

Exodus.collect_element_to_element_connectivities

Exodus.collect_node_to_element_connectivities

Exodus.copy_mesh

Exodus.decomp

Exodus.epu

Exodus.epu

Exodus.epu_error

Exodus.exodiff

Exodus.exodiff

Exodus.exodiff_error

Exodus.exodus_error_check

Exodus.int_and_float_modes

Exodus.nem_slice

Exodus.nem_slice

Exodus.nem_slice_error

Exodus.nem_spread

Exodus.nem_spread

Exodus.nem_spread_error

Exodus.open_exodus_file

Exodus.read_block

Exodus.read_block

Exodus.read_block_connectivity

Exodus.read_block_id_map

Exodus.read_block_parameters

Exodus.read_blocks

Exodus.read_blocks!

Exodus.read_coordinate_names

Exodus.read_coordinates

Exodus.read_coordinates

Exodus.read_element_type

Exodus.read_id_map

Exodus.read_ids

Exodus.read_ids

Exodus.read_info

Exodus.read_init_info

Exodus.read_map

Exodus.read_name

Exodus.read_name

Exodus.read_names

Exodus.read_names

Exodus.read_names

Exodus.read_node_set_nodes

Exodus.read_number_of_time_steps

Exodus.read_number_of_variables

Exodus.read_number_of_variables

Exodus.read_partial_block_connectivity

Exodus.read_partial_coordinates

Exodus.read_partial_coordinates_component

Exodus.read_qa

Exodus.read_set

Exodus.read_set

Exodus.read_set_parameters

Exodus.read_sets

Exodus.read_sets

Exodus.read_sets!

Exodus.read_side_set_elements_and_sides

Exodus.read_side_set_node_list

Exodus.read_time

Exodus.read_times

Exodus.read_values

Exodus.read_values

Exodus.read_values

Exodus.read_values

Exodus.read_values

Exodus.read_values

Exodus.read_values

Exodus.read_values

Exodus.read_values

Exodus.read_values

Exodus.read_values

Exodus.read_values

Exodus.read_values

Exodus.read_values

Exodus.read_values

Exodus.set_exodus_max_name_length

Exodus.set_exodus_options

Exodus.write_block

Exodus.write_block_connectivity

Exodus.write_blocks

Exodus.write_coordinate_names

Exodus.write_coordinates

Exodus.write_id_map

Exodus.write_info

Exodus.write_initialization!

Exodus.write_name

Exodus.write_name

Exodus.write_name

Exodus.write_names

Exodus.write_names

Exodus.write_number_of_variables

Exodus.write_partial_coordinates

Exodus.write_partial_coordinates_component

Exodus.write_partial_coordinates_component

Exodus.write_qa

Exodus.write_set

Exodus.write_set_parameters

Exodus.write_sets

Exodus.write_time

Exodus.write_values

Exodus.write_values

Exodus.write_values

Exodus.write_values

Exodus.write_values

Exodus.write_values

–>

Exodus.Exodus — Module

Block

EdgeMap

ElementMap

ElementVariable

ExodusDatabase

FaceMap

GlobalVariable

Initialization

NodalVariable

NodeMap

NodeSet

NodeSetVariable

SideSet

SideSetVariable

close

collect_element_connectivities

collect_element_to_element_connectivities

collect_node_to_element_connectivities

copy

copy_mesh

decomp

epu

exodiff

length

read_block

read_block_id_map

read_blocks

read_coordinates

read_element_type

read_id_map

read_ids

read_info

read_map

read_name

read_names

read_number_of_time_steps

read_number_of_variables

read_qa

read_set

read_sets

read_side_set_node_list

read_time

read_times

read_values

write_block

write_blocks

write_coordinates

write_id_map

write_info

write_name

write_names

write_number_of_variables

write_qa

write_set

write_sets

write_time

write_values

Exodus.Block — Type

id::Any

num_elem::Int64

num_nodes_per_elem::Int64

elem_type::String

conn::AbstractMatrix

Exodus.Block — Method

Init method for block container.

Exodus.Block — Method

Exodus.CommunicationMapParameters — Type

node_cmap_ids::Vector

node_cmap_node_cnts::Vector

elem_cmap_ids::Vector

elem_cmap_elem_cnts::Vector

Exodus.CommunicationMapParameters — Method

Exodus.EPUException — Type

cmd::Cmd

Exodus.EdgeMap — Type

Exodus.ElementCommunicationMap — Type

elem_ids::Vector

side_ids::Vector

proc_ids::Vector

Exodus.ElementCommunicationMap — Method

Exodus.ElementMap — Type

Exodus.ElementVariable — Type

Exodus.ExodiffException — Type

cmd::Cmd

Exodus.ExodusDatabase — Method

Type unstable helper to eliminate annoying lines of code to get type stability.

If you're looking for a type stable way to to open an exodus file, Simple copy past some of this into a barrier function

Exodus.ExodusDatabase — Method

Exodus.ExodusError — Type

error_code::Int32

error_msg::String

Exodus.ExodusWindowsError — Type

Exodus.FaceMap — Type

Exodus.GlobalVariable — Type

Exodus.Initialization — Method

Workaround method

Exodus.Initialization — Method

Workaround method

Exodus.Initialization — Method

Exodus.LoadBalanceParameters — Type

num_int_nodes::Any

num_bor_nodes::Any

num_ext_nodes::Any

num_int_elems::Any

num_bor_elems::Any

num_node_cmaps::Any

num_elem_cmaps::Any

processor::Int32

Exodus.LoadBalanceParameters — Method

Exodus.ModeException — Type

mode::String

Exodus.NemSliceException — Type

cmd::Cmd

Exodus.NemSpreadException — Type

cmd::Cmd

Exodus.NodalVariable — Type

Exodus.NodeCommunicationMap — Type

node_ids::Vector

proc_ids::Vector

Exodus.NodeCommunicationMap — Method

Exodus.NodeMap — Type

Exodus.NodeSet — Type

id::Any

nodes::AbstractVector

Exodus.NodeSet — Method

Exodus.NodeSet — Method

Exodus.NodeSetVariable — Type

Exodus.ParallelExodusDatabase — Type

base_file_name::String

exos::Array{ExodusDatabase{M, I, B, F}, 1} where {M, I, B, F}

nem::ExodusDatabase{M, I, B, Float32} where {M, I, B}

mode::String

init_global::Initialization

lb_params::Array{Exodus.LoadBalanceParameters{B}, 1} where B

cmap_params::Array{Exodus.CommunicationMapParameters{B}, 1} where B

Exodus.ParallelExodusDatabase — Method

Exodus.ProcessorElementMaps — Type

elem_map_internal::Vector

elem_map_border::Vector

Exodus.ProcessorElementMaps — Method

Exodus.ProcessorNodeMaps — Type

node_map_internal::Vector

node_map_border::Vector

node_map_external::Vector

Exodus.ProcessorNodeMaps — Method

Exodus.SetIDException — Type

exo::ExodusDatabase

type::Type

id::Integer

Exodus.SetNameException — Type

exo::ExodusDatabase

type::Type

name::String

Exodus.SideSet — Type

id::Any

elements::AbstractVector

sides::AbstractVector

Exodus.SideSet — Method

Exodus.SideSet — Method

Exodus.SideSetVariable — Type

Exodus.VariableIDException — Type

exo::ExodusDatabase

type::Type

id::Int64

Exodus.VariableNameException — Type

exo::ExodusDatabase

type::Type

name::String

Exodus.ex_entity_id — Type

Exodus.ex_entity_type — Type

ex_entity_type

Entity type enums (exentitytype in exodusII.h)

Exodus.ex_inquiry — Type

ex_inquiry

Inquiry enums (ex_inquiry in exodusII.h).

Exodus.void_int — Type

void_int = Cvoid

Base.close — Method

close(p::Exodus.ParallelExodusDatabase)

close(p)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ParallelExodus.jl:229.

Base.close — Method

close(exo::ExodusDatabase)

Used to close and ExodusDatabase.

close(exo)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExodusTypes.jl:560.

Base.copy — Method

copy(exo::ExodusDatabase, new_file_name::String)

Used to copy an ExodusDatabase. As of right now this is the best way to create a new ExodusDatabase for output. Not all of the put methods have been wrapped and properly tested. This one has though.

copy(exo, new_file_name)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExodusTypes.jl:569.

Base.length — Method

length(n::NodeSet) -> Any

length(n)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExodusTypes.jl:764.

Base.length — Method

length(s::SideSet) -> Any

length(s)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExodusTypes.jl:805.

Base.show — Method

show(io::IO, cmap_params::Exodus.CommunicationMapParameters)

show(io, cmap_params)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ParallelExodus.jl:98.

Base.show — Method

show(io::IO, e::Exodus.EPUException)

show(io, e)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Epu.jl:9.

Base.show — Method

show(io::IO, e::Exodus.ExodiffException)

show(io, e)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExoDiff.jl:9.

Base.show — Method

show(io::IO, e::Exodus.ExodusError)

show(io, e)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Errors.jl:10.

Base.show — Method

show(io::IO, init::Exodus.LoadBalanceParameters)

show(io, init)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ParallelExodus.jl:48.

Base.show — Method

show(io::IO, e::Exodus.NemSliceException)

show(io, e)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Decomp.jl:15.

Base.show — Method

show(io::IO, e::Exodus.NemSpreadException)

show(io, e)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Decomp.jl:21.

Base.show — Method

show(io::IO, e::Exodus.SetIDException)

show(io, e)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExodusTypes.jl:629.

Base.show — Method

show(io::IO, e::Exodus.SetNameException)

show(io, e)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExodusTypes.jl:647.

Base.show — Method

show(io::IO, e::Exodus.VariableIDException)

show(io, e)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExodusTypes.jl:665.

Base.show — Method

show(io::IO, e::Exodus.VariableNameException)

show(io, e)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExodusTypes.jl:684.

Base.show — Method

show(io::IO, init::Initialization)

show(io, init)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExodusTypes.jl:191.

Base.show — Method

show(io::IO, node_set::NodeSet)

show(io, node_set)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExodusTypes.jl:768.

Base.show — Method

show(io::IO, sset::SideSet)

show(io, sset)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExodusTypes.jl:809.

Base.show — Method

show(io::IO, block::Block)

show(io, block)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExodusTypes.jl:727.

Base.show — Method

show(
    io::IO,
    exo::Exodus.ParallelExodusDatabase{M, I, B, F, N}
)

show(io, exo)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ParallelExodus.jl:198.

Exodus.InitializationGlobal — Method

InitializationGlobal(
    exo::ExodusDatabase{M, I, B, F}
) -> Initialization

InitializationGlobal(exo)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ParallelExodus.jl:16.

Exodus.collect_element_connectivities — Method

collect_element_connectivities(
    exo::ExodusDatabase{M, I, B, F}
) -> Any

collects all blocks by default

collect_element_connectivities(exo)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Helpers.jl:14.

Exodus.collect_element_to_element_connectivities — Method

collect_element_to_element_connectivities(
    exo::ExodusDatabase{M, I, B, F}
) -> Any

collect all blocks by default

collect_element_to_element_connectivities(exo)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Helpers.jl:71.

Exodus.collect_node_to_element_connectivities — Method

collect_node_to_element_connectivities(
    exo::ExodusDatabase{M, I, B, F}
) -> Any

collect all blocks by default

collect_node_to_element_connectivities(exo)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Helpers.jl:40.

Exodus.copy_mesh — Method

copy_mesh(file_name::String, new_file_name::String)

Simpler copy method to only copy a mesh for output later on

copy_mesh(file_name, new_file_name)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExodusTypes.jl:588.

Exodus.decomp — Method

decomp(file_name::String, n_procs::Integer)

decomp(file_name, n_procs)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Decomp.jl:120.

Exodus.epu — Method

epu(file_name::String)

epu(file_name)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Epu.jl:28.

Exodus.epu — Method

epu() -> Any

Prints epu help message

epu()

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Epu.jl:20.

Exodus.epu_error — Method

epu_error(cmd::Cmd)

epu_error(cmd)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Epu.jl:14.

Exodus.exodiff — Method

exodiff(ex_1::String, ex_2::String; command_file) -> Bool

Return true if the two files pass the exodiff test. Otherwise it returns false

exodiff(ex_1, ex_2; command_file)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExoDiff.jl:28.

Exodus.exodiff — Method

exodiff() -> Any

Prints exodiff help message

exodiff()

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExoDiff.jl:19.

Exodus.exodiff_error — Method

exodiff_error(cmd::Cmd)

exodiff_error(cmd)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExoDiff.jl:14.

Exodus.exodus_error_check — Method

exodus_error_check(error_code::Integer, method_name::String)

Generic error handling method.

Arguments

error_code::T: error code, usually negative means something went bad

method_name::String: method name that called this

exodus_error_check(error_code, method_name)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Errors.jl:19.

Exodus.int_and_float_modes — Method

int_and_float_modes(exo::Int32) -> NTuple{4, DataType}

int_and_float_modes(exo)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExodusTypes.jl:106.

Exodus.nem_slice — Method

nem_slice() -> Any

nem_slice()

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Decomp.jl:35.

Exodus.nem_slice — Method

nem_slice(file_name::String, n_procs::Integer) -> Any

nem_slice(file_name, n_procs)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Decomp.jl:43.

Exodus.nem_slice_error — Method

nem_slice_error(cmd::Cmd)

nem_slice_error(cmd)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Decomp.jl:27.

Exodus.nem_spread — Method

nem_spread() -> Any

nem_spread()

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Decomp.jl:69.

Exodus.nem_spread — Method

nem_spread(file_name::String, n_procs::Integer) -> Any

nem_spread(file_name, n_procs)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Decomp.jl:77.

Exodus.nem_spread_error — Method

nem_spread_error(cmd::Cmd)

nem_spread_error(cmd)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Decomp.jl:31.

Exodus.open_exodus_file — Method

open_exodus_file(file_name::String, mode) -> Int32

Helper method for opening exodus database

open_exodus_file(file_name, mode)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExodusTypes.jl:402.

Exodus.read_block — Method

read_block(exo::ExodusDatabase, block_id::Integer) -> Block

read_block(exo, block_id)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Blocks.jl:96.

Exodus.read_block — Method

read_block(exo::ExodusDatabase, block_name::String) -> Block

read_block(exo, block_name)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Blocks.jl:100.

Exodus.read_block_connectivity — Method

read_block_connectivity(
    exo::ExodusDatabase{M, I, B, F},
    block_id::Integer,
    conn_length::Integer
) -> Vector

read_block_connectivity(exo, block_id, conn_length)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Blocks.jl:39.

Exodus.read_block_id_map — Method

read_block_id_map(
    exo::ExodusDatabase,
    block_id::Integer
) -> Any

TODO do this one later... it depends on a few things

TODO fix this to not use void_int... use a proper type

read_block_id_map(exo, block_id)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Blocks.jl:5.

Exodus.read_block_parameters — Method

read_block_parameters(
    exo::ExodusDatabase{M, I, B, F},
    block_id::Integer
) -> Tuple{String, Vararg{Any, 5}}

read_block_parameters(exo, block_id)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Blocks.jl:17.

Exodus.read_blocks! — Method

read_blocks!(
    blocks::Array{B<:Block, 1},
    exo::ExodusDatabase,
    block_ids::Array{I<:Integer, 1}
)

TODO: change name to readelementblocks!

read_blocks!(blocks, exo, block_ids)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Blocks.jl:105.

Exodus.read_blocks — Method

read_blocks(
    exo::ExodusDatabase,
    block_ids::Union{Vector{<:Integer}, var"#s127"} where var"#s127"<:Integer
) -> Any

Helper method for initializing blocks.

TODO: change name to readelementblocks

read_blocks(exo, block_ids)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Blocks.jl:118.

Exodus.read_coordinate_names — Method

read_coordinate_names(exo::ExodusDatabase) -> Any

Method to read coordinates names Returns a vector of strings

read_coordinate_names(exo)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Coordinates.jl:120.

Exodus.read_coordinates — Method

read_coordinates(exo::Exodus.ParallelExodusDatabase) -> Any

read_coordinates(exo)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ParallelExodus.jl:353.

Exodus.read_coordinates — Method

read_coordinates(exo::ExodusDatabase) -> Matrix

Method to read coordinates. Returns a matrix that is ndim x nnodes.

read_coordinates(exo)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Coordinates.jl:5.

Exodus.read_element_type — Method

read_element_type(
    exo::ExodusDatabase,
    block_id::Integer
) -> String

read_element_type(exo, block_id)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Blocks.jl:85.

Exodus.read_id_map — Method

read_id_map(
    exo::ExodusDatabase{M, I, B, F},
    type::Type{MAP<:Exodus.AbstractExodusMap}
) -> Any

read_id_map(exo, type)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Maps.jl:13.

Exodus.read_ids — Method

read_ids(
    exo::ExodusDatabase{M, I, B, F},
    _::Type{S<:Exodus.AbstractExodusSet}
) -> Any

read_ids(exo, _)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:3.

Exodus.read_ids — Method

read_ids(
    exo::Exodus.ParallelExodusDatabase,
    type::Type{T<:Exodus.AbstractExodusSet}
) -> Any

read_ids(exo, type)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ParallelExodus.jl:358.

Exodus.read_info — Method

read_info(exo::ExodusDatabase) -> Vector{String}

read_info(exo)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Info.jl:3.

Exodus.read_init_info — Method

read_init_info(
    exo::ExodusDatabase
) -> Tuple{Int32, Int32, String}

read_init_info(exo)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ParallelExodus.jl:3.

Exodus.read_map — Method

read_map(exo::ExodusDatabase{M, I, B, F}) -> Any

TODO change to not use void_int

read_map(exo)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Maps.jl:4.

Exodus.read_name — Method

read_name(
    exo::ExodusDatabase,
    _::Type{S<:Exodus.AbstractExodusSet},
    id::Integer
) -> String

read_name(exo, _, id)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:22.

Exodus.read_name — Method

read_name(
    exo::ExodusDatabase,
    _::Type{V<:Exodus.AbstractExodusVariable},
    var_index::Integer
) -> String

General method to read the name of a variable in index var_index for a given variable type V.

Examples: julia> readname(exo, ElementVariable, 1) "stressxx"

julia> readname(exo, GlobalVariable, 2) "reactionforce"

julia> readname(exo, NodalVariable, 1) "displx"

julia> readname(exo, NodeSetVariable, 1) "nsetdispl_x"

julia> read_name(exo, SideSetVariable, 1) "pressure"

read_name(exo, _, var_index)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:49.

Exodus.read_names — Method

read_names(
    exo::ExodusDatabase,
    _::Type{S<:Exodus.AbstractExodusSet}
) -> Any

read_names(exo, _)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:34.

Exodus.read_names — Method

read_names(
    exo::Exodus.ParallelExodusDatabase,
    type::Type{T<:Union{Exodus.AbstractExodusVariable, Exodus.AbstractExodusSet}}
) -> Any

read_names(exo, type)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ParallelExodus.jl:363.

Exodus.read_names — Method

read_names(
    exo::ExodusDatabase,
    _::Type{V<:Exodus.AbstractExodusVariable}
) -> Vector{String}

General method to read the names of variables for a given variable type V.

Examples: julia> readnames(exo, ElementVariable) "stressxx" "stressyy" "stresszz" "stressxy" "stressyz" "stress_zx"

julia> readnames(exo, GlobalVariable) "globaldispl" "reaction_force"

julia> readnames(exo, NodalVariable) "displx" "disply" "displz"

julia> readname(exo, NodeSetVariable) "nsetdisplx" "nsetdisply" "nsetdispl_z"

julia> read_name(exo, SideSetVariable) "pressure"

read_names(exo, _)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:90.

Exodus.read_node_set_nodes — Method

read_node_set_nodes(
    exo::ExodusDatabase{M, I, B, F},
    set_id::Integer
) -> Any

read_node_set_nodes(exo, set_id)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:63.

Exodus.read_number_of_time_steps — Method

read_number_of_time_steps(exo::ExodusDatabase) -> UInt32

read_number_of_time_steps(exo)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Times.jl:3.

Exodus.read_number_of_variables — Method

read_number_of_variables(
    exo::Exodus.ParallelExodusDatabase,
    type::Type{T<:Exodus.AbstractExodusVariable}
) -> Any

read_number_of_variables(exo, type)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ParallelExodus.jl:378.

Exodus.read_number_of_variables — Method

read_number_of_variables(
    exo::ExodusDatabase,
    _::Type{V<:Exodus.AbstractExodusVariable}
) -> Int32

General method to read the number of variables for a given variable type V.

Examples: julia> readnumberof_variables(exo, ElementVariable) 6

julia> readnumberof_variables(exo, GlobalVariable) 5

julia> readnumberof_variables(exo, NodalVariable) 3

julia> readnumberof_variables(exo, NodeSetVariable) 3

julia> readnumberof_variables(exo, SideSetVariable) 6

read_number_of_variables(exo, _)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:20.

Exodus.read_partial_block_connectivity — Method

read_partial_block_connectivity(
    exo::ExodusDatabase,
    block_id::Integer,
    start_num::Integer,
    num_ent::Integer
) -> Any

read_partial_block_connectivity(
    exo,
    block_id,
    start_num,
    num_ent
)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Blocks.jl:68.

Exodus.read_partial_coordinates — Method

read_partial_coordinates(
    exo::ExodusDatabase,
    start_node_num::Integer,
    num_nodes::Integer
) -> Matrix

Method to read a partial set of coordinates that are contiguous. Returns a matrix that is ndim x nnodes

read_partial_coordinates(exo, start_node_num, num_nodes)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Coordinates.jl:43.

Exodus.read_partial_coordinates_component — Method

read_partial_coordinates_component(
    exo::ExodusDatabase,
    start_node_num::Integer,
    num_nodes::Integer,
    component::Integer
) -> Vector

Method to read a specific component of a partial set of coordinates that are contiguous. Returns a vector of length n_nodes TODO change to not use Cvoid

read_partial_coordinates_component(
    exo,
    start_node_num,
    num_nodes,
    component
)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Coordinates.jl:91.

Exodus.read_qa — Method

read_qa(exo::ExodusDatabase) -> Matrix{String}

read_qa(exo)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/QA.jl:3.

Exodus.read_set — Method

read_set(
    exo::ExodusDatabase,
    type::Type{S<:Exodus.AbstractExodusSet},
    set_id
) -> Any

read_set(exo, type, set_id)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:113.

Exodus.read_set — Method

read_set(
    exo::Exodus.ParallelExodusDatabase,
    type::Type{T},
    id
) -> Any

read_set(exo, type, id)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ParallelExodus.jl:368.

Exodus.read_set_parameters — Method

read_set_parameters(
    exo::ExodusDatabase{M, I, B, F},
    set_id::Integer,
    _::Type{S<:Union{NodeSet, SideSet}}
) -> Tuple{Any, Any}

read_set_parameters(exo, set_id, _)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:46.

Exodus.read_sets! — Method

read_sets!(
    sets::Array{T<:Exodus.AbstractExodusSet, 1},
    exo::ExodusDatabase,
    ids::Array{I, 1}
)

read_sets!(sets, exo, ids)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:119.

Exodus.read_sets — Method

read_sets(
    exo::ExodusDatabase{M, I, B, F},
    type::Type{S<:Exodus.AbstractExodusSet}
) -> Any

read_sets(exo, type)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:135.

Exodus.read_sets — Method

read_sets(
    exo::Exodus.ParallelExodusDatabase,
    type::Type{T<:Exodus.AbstractExodusSet}
) -> Any

read_sets(exo, type)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ParallelExodus.jl:373.

Exodus.read_side_set_elements_and_sides — Method

read_side_set_elements_and_sides(
    exo::ExodusDatabase{M, I, B, F},
    set_id::Integer
) -> Tuple{Any, Any}

read_side_set_elements_and_sides(exo, set_id)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:77.

Exodus.read_side_set_node_list — Method

read_side_set_node_list(
    exo::ExodusDatabase{M, I, B, F},
    side_set_id::Integer
) -> Tuple{Any, Vector}

UNTESTED

read_side_set_node_list(exo, side_set_id)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:92.

Exodus.read_time — Method

read_time(exo::ExodusDatabase, time_step::Integer) -> Any

TODO figure out how to make this not use a vector of length 1 - either a ref or a ptr

read_time(exo, time_step)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Times.jl:12.

Exodus.read_times — Method

read_times(exo::ExodusDatabase) -> Vector

read_times(exo)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Times.jl:23.

Exodus.read_values — Method

read_values(
    exo::Exodus.ParallelExodusDatabase,
    t::Type{GlobalVariable},
    timestep::Integer
) -> Any

Wrapper method for global variables around the main readvalues method readvalues(exo::ParallelExodusDatabase, t::Type{GlobalVariable}, timestep::Integer) = read_values(exo, (t,), (timestep,), (1,), (1,))

Example: read_values(exo, GlobalVariable, 1)

read_values(exo, t, timestep)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ParallelExodus.jl:394.

Exodus.read_values — Method

read_values(
    exo::Exodus.ParallelExodusDatabase,
    t::Type{NodalVariable},
    timestep::Integer,
    index::Integer
) -> Any

Wrapper method for nodal variables

read_values(exo, t, timestep, index)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ParallelExodus.jl:400.

Exodus.read_values — Method

read_values(
    exo::Exodus.ParallelExodusDatabase,
    t::Type{NodalVariable},
    timestep::Integer,
    name::String
) -> Any

Wrapper method for nodal variables

read_values(exo, t, timestep, name)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ParallelExodus.jl:411.

Exodus.read_values — Method

read_values(
    exo::ExodusDatabase,
    t::Type{GlobalVariable},
    timestep::Integer
) -> Vector

Wrapper method for global variables around the main readvalues method readvalues(exo::ExodusDatabase, t::Type{GlobalVariable}, timestep::Integer) = read_values(exo, t, timestep, 1, 1)

Example: read_values(exo, GlobalVariable, 1)

read_values(exo, t, timestep)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:275.

Exodus.read_values — Method

read_values(
    exo::ExodusDatabase,
    t::Type{NodalVariable},
    timestep::Integer,
    index::Integer
) -> Any

Wrapper method for nodal variables

read_values(exo, t, timestep, index)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:280.

Exodus.read_values — Method

read_values(
    exo::ExodusDatabase,
    t::Type{NodalVariable},
    timestep::Integer,
    name::String
) -> Any

Wrapper method for nodal variables

read_values(exo, t, timestep, name)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:296.

Exodus.read_values — Method

read_values(
    exo::Exodus.ParallelExodusDatabase,
    type::Type{T<:Exodus.AbstractExodusSet},
    time_step::Int64,
    id::Int64,
    var_index::Int64
) -> Any

read_values(exo, type, time_step, id, var_index)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ParallelExodus.jl:383.

Exodus.read_values — Method

read_values(
    exo::Exodus.ParallelExodusDatabase,
    type::Type{V<:Exodus.AbstractExodusVariable},
    time_step::Integer,
    id::Integer,
    var_name::String
) -> Any

read_values(exo, type, time_step, id, var_name)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ParallelExodus.jl:405.

Exodus.read_values — Method

read_values(
    exo::Exodus.ParallelExodusDatabase,
    type::Type{V<:Exodus.AbstractExodusVariable},
    time_step::Integer,
    set_name::String,
    var_name::String
)

read_values(exo, type, time_step, set_name, var_name)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ParallelExodus.jl:419.

Exodus.read_values — Method

read_values(
    exo::ExodusDatabase,
    _::Type{V<:Union{ElementVariable, NodalVariable, NodeSetVariable, SideSetVariable}},
    time_step::Integer,
    id::Integer,
    var_name::String
) -> Any

read_values(exo, _, time_step, id, var_name)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:285.

Exodus.read_values — Method

read_values(
    exo::ExodusDatabase,
    _::Type{V<:Union{ElementVariable, NodeSetVariable, SideSetVariable}},
    time_step::Integer,
    set_name::String,
    var_name::String
) -> Any

read_values(exo, _, time_step, set_name, var_name)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:301.

Exodus.read_values — Method

read_values(
    exo::ExodusDatabase{M, I, B, F},
    _::Type{V<:ElementVariable},
    timestep::Integer,
    id::Integer,
    var_index::Integer
) -> Any

Method to read element variables

read_values(exo, _, timestep, id, var_index)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:159.

Exodus.read_values — Method

read_values(
    exo::ExodusDatabase{M, I, B, F},
    _::Type{V<:GlobalVariable},
    timestep::Integer,
    id::Integer,
    var_index::Integer
) -> Vector

Method to read global variables

read_values(exo, _, timestep, id, var_index)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:187.

Exodus.read_values — Method

read_values(
    exo::ExodusDatabase{M, I, B, F},
    _::Type{V<:NodalVariable},
    timestep::Integer,
    id::Integer,
    var_index::Integer
) -> Any

Method to read nodal variables

read_values(exo, _, timestep, id, var_index)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:215.

Exodus.read_values — Method

read_values(
    exo::ExodusDatabase{M, I, B, F},
    _::Type{V<:Union{NodeSetVariable, SideSetVariable}},
    timestep::Integer,
    id::Integer,
    var_index::Integer
) -> Any

Method to read nodeset/sideset variables

read_values(exo, _, timestep, id, var_index)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:243.

Exodus.set_exodus_max_name_length — Method

set_exodus_max_name_length(exoid::Int32, len::Int32)

set_exodus_max_name_length(exoid, len)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExodusTypes.jl:21.

Exodus.set_exodus_options — Method

set_exodus_options(options)

set_exodus_options(options)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExodusTypes.jl:14.

Exodus.write_block — Method

write_block(exo::ExodusDatabase, block::Block)

WARNING: currently does not support edges, faces and attributes

write_block(exo, block)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Blocks.jl:135.

Exodus.write_block_connectivity — Method

write_block_connectivity(
    exo::ExodusDatabase,
    block_id::Integer,
    conn::Array{I<:Integer, 2}
)

write_block_connectivity(exo, block_id, conn)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Blocks.jl:51.

Exodus.write_blocks — Method

write_blocks(exo::ExodusDatabase, blocks::Vector{<:Block})

write_blocks(exo, blocks)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Blocks.jl:161.

Exodus.write_coordinate_names — Method

write_coordinate_names(
    exo::ExodusDatabase,
    coord_names::Vector{String}
)

Method to write coordinate names, e.g. x, y, z

write_coordinate_names(exo, coord_names)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Coordinates.jl:179.

Exodus.write_coordinates — Method

write_coordinates(
    exo::ExodusDatabase,
    coords::Union{Array{F<:AbstractFloat, 1}, Array{F<:AbstractFloat, 2}}
)

Method to write coordinates

write_coordinates(exo, coords)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Coordinates.jl:138.

Exodus.write_id_map — Method

write_id_map(
    exo::ExodusDatabase{M, I, B, F},
    type::Type{MAP<:Exodus.AbstractExodusMap},
    map::Array{M, 1}
)

write_id_map(exo, type, map)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Maps.jl:35.

Exodus.write_info — Method

write_info(exo::ExodusDatabase, info::Vector{String})

write_info(exo, info)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Info.jl:24.

Exodus.write_initialization! — Method

write_initialization!(exoid::Int32, init::Initialization)

Used to set up a exodus database in write mode

The ccall signatures should reall be B (bulk int type of exo) instead of Clonglong

write_initialization!(exoid, init)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExodusTypes.jl:207.

Exodus.write_name — Method

write_name(
    exo::ExodusDatabase{M, I, B, F},
    set::Exodus.AbstractExodusSet,
    name::String
)

write_name(exo, set, name)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:204.

Exodus.write_name — Method

write_name(
    exo::ExodusDatabase{M, I, B, F},
    _::Type{S<:Exodus.AbstractExodusSet},
    set_id::Integer,
    name::String
)

write_name(exo, _, set_id, name)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:185.

Exodus.write_name — Method

write_name(
    exo::ExodusDatabase,
    _::Type{V<:Exodus.AbstractExodusVariable},
    var_index::Integer,
    var_name::String
)

write_name(exo, _, var_index, var_name)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:371.

Exodus.write_names — Method

write_names(
    exo::ExodusDatabase,
    _::Type{S<:Exodus.AbstractExodusSet},
    names::Vector{String}
)

WARNING: this methods likely does not have good safe guards

write_names(exo, _, names)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:223.

Exodus.write_names — Method

write_names(
    exo::ExodusDatabase,
    type::Type{V<:Exodus.AbstractExodusVariable},
    var_names::Vector{String}
)

write_names(exo, type, var_names)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:385.

Exodus.write_number_of_variables — Method

write_number_of_variables(
    exo::ExodusDatabase,
    _::Type{V<:Exodus.AbstractExodusVariable},
    num_vars::Integer
)

General method to write the number of variables for a given variable type V.

Examples: julia> writenumberof_variables(exo, ElementVariable, 6)

julia> writenumberof_variables(exo, GlobalVariable, 5)

julia> writenumberof_variables(exo, NodalVariable, 3)

julia> writenumberof_variables(exo, NodeSetVariable, 3)

julia> writenumberof_variables(exo, SideSetVariable, 6)

write_number_of_variables(exo, _, num_vars)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:362.

Exodus.write_partial_coordinates — Method

write_partial_coordinates(
    exo::ExodusDatabase,
    start_node_num::Integer,
    coords::Union{Array{F<:AbstractFloat, 1}, Array{F<:AbstractFloat, 2}}
)

write_partial_coordinates(exo, start_node_num, coords)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Coordinates.jl:190.

Exodus.write_partial_coordinates_component — Method

write_partial_coordinates_component(
    exo::ExodusDatabase,
    start_node_num::Integer,
    component::Integer,
    coords::Array{F<:AbstractFloat, 1}
)

write_partial_coordinates_component(
    exo,
    start_node_num,
    component,
    coords
)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Coordinates.jl:217.

Exodus.write_partial_coordinates_component — Method

write_partial_coordinates_component(
    exo::ExodusDatabase,
    start_node_num::Integer,
    component::String,
    coords::Array{F<:AbstractFloat, 1}
)

write_partial_coordinates_component(
    exo,
    start_node_num,
    component,
    coords
)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Coordinates.jl:228.

Exodus.write_qa — Method

write_qa(exo::ExodusDatabase, qa_record::Matrix{String})

write_qa(exo, qa_record)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/QA.jl:29.

Exodus.write_set — Method

write_set(
    exo::ExodusDatabase{M, I, B, F},
    set::Exodus.AbstractExodusSet
)

Typing ensures we don't write a set with non-matching types to the exodus file.

write_set(exo, set)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:166.

Exodus.write_set_parameters — Method

write_set_parameters(
    exo::ExodusDatabase{M, I, B, F},
    set::Exodus.AbstractExodusSet
)

WARNING: currently doesn't support distance factors

write_set_parameters(exo, set)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:153.

Exodus.write_sets — Method

write_sets(
    exo::ExodusDatabase,
    sets::Array{T<:Exodus.AbstractExodusSet, 1}
)

write_sets(exo, sets)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:177.

Exodus.write_time — Method

write_time(
    exo::ExodusDatabase,
    time_step::Integer,
    time_value::AbstractFloat
)

write_time(exo, time_step, time_value)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Times.jl:33.

Exodus.write_values — Method

write_values(
    exo::ExodusDatabase,
    t::Type{GlobalVariable},
    timestep::Integer,
    var_values::Vector{<:AbstractFloat}
)

Wrapper method for global variables around the main writevalues method writevalues( exo::ExodusDatabase, t::Type{GlobalVariable}, timestep::Integer, varvalues::Vector{<:AbstractFloat} ) = writevalues(exo, t, timestep, 1, 1, var_values)

Note: you need to first run writenumberof_variables(exo, GlobalVariable, n) where n is the number of variables.

Example: writenumberofvariables(exo, GlobalVariable, 5) writevalues(exo, GlobalVariable, 1, [10.0, 20.0, 30.0, 40.0, 50.0])

write_values(exo, t, timestep, var_values)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:441.

Exodus.write_values — Method

write_values(
    exo::ExodusDatabase,
    t::Type{NodalVariable},
    timestep::Integer,
    var_index::Integer,
    var_values::Vector{<:AbstractFloat}
)

Wrapper for writing nodal variables by index number

write_values(exo, t, timestep, var_index, var_values)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:450.

Exodus.write_values — Method

write_values(
    exo::ExodusDatabase,
    t::Type{NodalVariable},
    timestep::Integer,
    var_name::String,
    var_values::Vector{<:AbstractFloat}
)

Wrapper method for nodal variables

write_values(exo, t, timestep, var_name, var_values)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:471.

Exodus.write_values — Method

write_values(
    exo::ExodusDatabase,
    _::Type{V<:Exodus.AbstractExodusVariable},
    timestep::Integer,
    id::Integer,
    var_index::Integer,
    var_values::Vector{<:AbstractFloat}
)

write_values(exo, _, timestep, id, var_index, var_values)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:407.

Exodus.write_values — Method

write_values(
    exo::ExodusDatabase,
    _::Type{V<:Exodus.AbstractExodusVariable},
    timestep::Integer,
    id::Integer,
    var_name::String,
    var_value::Vector{<:AbstractFloat}
)

write_values(exo, _, timestep, id, var_name, var_value)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:454.

Exodus.write_values — Method

write_values(
    exo::ExodusDatabase,
    _::Type{V<:Exodus.AbstractExodusVariable},
    time_step::Integer,
    set_name::String,
    var_name::String,
    var_value::Vector{<:AbstractFloat}
)

write_values(
    exo,
    _,
    time_step,
    set_name,
    var_name,
    var_value
)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:475.