Exodus

Documentation for Exodus. <!–

Exodus.Exodus

Exodus.Block

Exodus.Block

Exodus.Block

Exodus.ElementVariable

Exodus.ExodusDatabase

Exodus.ExodusDatabase

Exodus.ExodusError

Exodus.GlobalVariable

Exodus.Initialization

Exodus.Initialization

Exodus.Initialization

Exodus.ModeException

Exodus.NodalVariable

Exodus.NodeSet

Exodus.NodeSet

Exodus.NodeSet

Exodus.NodeSetVariable

Exodus.SetIDException

Exodus.SetNameException

Exodus.SideSet

Exodus.SideSet

Exodus.SideSet

Exodus.SideSetVariable

Exodus.VariableIDException

Exodus.VariableNameException

Exodus.ex_entity_id

Exodus.ex_entity_type

Exodus.ex_inquiry

Exodus.void_int

Base.close

Base.copy

Base.length

Base.length

Base.show

Base.show

Base.show

Base.show

Base.show

Base.show

Base.show

Base.show

Base.show

Exodus.collect_block_connectivities

Exodus.decomp

Exodus.exodus_error_check

Exodus.int_and_float_modes

Exodus.read_block

Exodus.read_block

Exodus.read_block_connectivity

Exodus.read_block_id_map

Exodus.read_block_parameters

Exodus.read_blocks

Exodus.read_blocks!

Exodus.read_coordinate_names

Exodus.read_coordinates

Exodus.read_element_type

Exodus.read_info

Exodus.read_map

Exodus.read_name

Exodus.read_name

Exodus.read_names

Exodus.read_names

Exodus.read_node_set_nodes

Exodus.read_number_of_time_steps

Exodus.read_number_of_variables

Exodus.read_partial_block_connectivity

Exodus.read_partial_coordinates

Exodus.read_partial_coordinates_component

Exodus.read_qa

Exodus.read_set

Exodus.read_set_parameters

Exodus.read_sets

Exodus.read_sets!

Exodus.read_side_set_elements_and_sides

Exodus.read_side_set_node_list

Exodus.read_time

Exodus.read_times

Exodus.read_values

Exodus.read_values

Exodus.read_values

Exodus.read_values

Exodus.read_values

Exodus.read_values

Exodus.set_exodus_max_name_length

Exodus.set_exodus_options

Exodus.write_block

Exodus.write_block_connectivity

Exodus.write_blocks

Exodus.write_coordinate_names

Exodus.write_coordinates

Exodus.write_info

Exodus.write_initialization!

Exodus.write_name

Exodus.write_name

Exodus.write_name

Exodus.write_names

Exodus.write_names

Exodus.write_number_of_variables

Exodus.write_partial_coordinates

Exodus.write_partial_coordinates_component

Exodus.write_partial_coordinates_component

Exodus.write_qa

Exodus.write_set

Exodus.write_set_parameters

Exodus.write_sets

Exodus.write_time

Exodus.write_values

Exodus.write_values

Exodus.write_values

Exodus.write_values

Exodus.write_values

Exodus.write_values

Exodus.@decomp

Exodus.@epu

Exodus.@exodiff

–>

Exodus.Exodus — Module

Block

ElementVariable

ExodusDatabase

GlobalVariable

Initialization

NodalVariable

NodeSet

NodeSetVariable

SideSet

SideSetVariable

close

collect_block_connectivities

copy

@decomp

decomp

@epu

@exodiff

length

read_block

read_block_id_map

read_blocks

read_coordinates

read_element_type

read_ids

read_info

read_map

read_name

read_names

read_number_of_time_steps

read_number_of_variables

read_qa

read_set

read_sets

read_time

read_times

read_values

write_block

write_blocks

write_coordinates

write_info

write_name

write_names

write_number_of_variables

write_qa

write_set

write_sets

write_time

write_values

Exodus.Block — Type

id::Any

num_elem::Int64

num_nodes_per_elem::Int64

elem_type::String

conn::Matrix

Exodus.Block — Method

Init method for block container.

Exodus.Block — Method

Exodus.ElementVariable — Type

Exodus.ExodusDatabase — Method

Exodus.ExodusDatabase — Method

Exodus.ExodusError — Type

error_code::Int32

error_msg::String

Exodus.GlobalVariable — Type

Exodus.Initialization — Method

Workaround method

Exodus.Initialization — Method

Workaround method

Exodus.Initialization — Method

Exodus.ModeException — Type

mode::String

Exodus.NodalVariable — Type

Exodus.NodeSet — Type

id::Any

nodes::Vector

Exodus.NodeSet — Method

Exodus.NodeSet — Method

Exodus.NodeSetVariable — Type

Exodus.SetIDException — Type

exo::ExodusDatabase

type::Type

id::Integer

Exodus.SetNameException — Type

exo::ExodusDatabase

type::Type

name::String

Exodus.SideSet — Type

id::Any

elements::Vector

sides::Vector

Exodus.SideSet — Method

Exodus.SideSet — Method

Exodus.SideSetVariable — Type

Exodus.VariableIDException — Type

exo::ExodusDatabase

type::Type

id::Int64

Exodus.VariableNameException — Type

exo::ExodusDatabase

type::Type

name::String

Exodus.ex_entity_id — Type

Exodus.ex_entity_type — Type

ex_entity_type

Entity type enums (exentitytype in exodusII.h)

Exodus.ex_inquiry — Type

ex_inquiry

Inquiry enums (ex_inquiry in exodusII.h).

Exodus.void_int — Type

void_int = Cvoid

Base.close — Method

close(exo::ExodusDatabase)

Used to close and ExodusDatabase.

close(exo)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExodusTypes.jl:452.

Base.copy — Method

copy(exo::ExodusDatabase, new_file_name::String)

Used to copy an ExodusDatabase. As of right now this is the best way to create a new ExodusDatabase for output. Not all of the put methods have been wrapped and properly tested. This one has though.

copy(exo, new_file_name)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExodusTypes.jl:461.

Base.length — Method

length(n::NodeSet) -> Int64

length(n)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExodusTypes.jl:645.

Base.length — Method

length(s::SideSet) -> Int64

length(s)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExodusTypes.jl:685.

Base.show — Method

show(io::IO, e::Exodus.ExodusError)

show(io, e)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Errors.jl:10.

Base.show — Method

show(io::IO, e::Exodus.SetIDException)

show(io, e)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExodusTypes.jl:513.

Base.show — Method

show(io::IO, e::Exodus.SetNameException)

show(io, e)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExodusTypes.jl:531.

Base.show — Method

show(io::IO, e::Exodus.VariableIDException)

show(io, e)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExodusTypes.jl:549.

Base.show — Method

show(io::IO, e::Exodus.VariableNameException)

show(io, e)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExodusTypes.jl:568.

Base.show — Method

show(io::IO, init::Initialization)

show(io, init)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExodusTypes.jl:161.

Base.show — Method

show(io::IO, node_set::NodeSet)

show(io, node_set)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExodusTypes.jl:649.

Base.show — Method

show(io::IO, sset::SideSet)

show(io, sset)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExodusTypes.jl:689.

Base.show — Method

show(io::IO, block::Block)

show(io, block)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExodusTypes.jl:610.

Exodus.collect_block_connectivities — Method

collect_block_connectivities(exo::ExodusDatabase) -> Any

TODO reduce allocations and check for type instabilities

collect_block_connectivities(exo)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Blocks.jl:181.

Exodus.decomp — Method

decomp(file_name::String, n_procs::Integer) -> Any

decomp(file_name, n_procs)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Decomp.jl:85.

Exodus.exodus_error_check — Method

exodus_error_check(error_code::Integer, method_name::String)

Generic error handling method.

Arguments

error_code::T: error code, usually negative means something went bad

method_name::String: method name that called this

exodus_error_check(error_code, method_name)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Errors.jl:19.

Exodus.int_and_float_modes — Method

int_and_float_modes(exo::Int32) -> NTuple{4, Type}

int_and_float_modes(exo)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExodusTypes.jl:30.

Exodus.read_block — Method

read_block(exo::ExodusDatabase, block_id::Integer) -> Block

read_block(exo, block_id)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Blocks.jl:106.

Exodus.read_block — Method

read_block(exo::ExodusDatabase, block_name::String) -> Block

read_block(exo, block_name)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Blocks.jl:110.

Exodus.read_block_connectivity — Method

read_block_connectivity(
    exo::ExodusDatabase{M, I, B, F},
    block_id::Integer,
    conn_length::Integer
) -> Vector

read_block_connectivity(exo, block_id, conn_length)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Blocks.jl:47.

Exodus.read_block_id_map — Method

read_block_id_map(
    exo::ExodusDatabase,
    block_id::Integer
) -> Vector

read_block_id_map(exo, block_id)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Blocks.jl:3.

Exodus.read_block_parameters — Method

read_block_parameters(
    exo::ExodusDatabase{M, I, B, F},
    block_id::Integer
) -> Tuple{String, Vararg{Any, 5}}

read_block_parameters(exo, block_id)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Blocks.jl:24.

Exodus.read_blocks! — Method

read_blocks!(
    blocks::Array{B<:Block, 1},
    exo::ExodusDatabase,
    block_ids::Array{I<:Integer, 1}
)

TODO: change name to readelementblocks!

read_blocks!(blocks, exo, block_ids)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Blocks.jl:115.

Exodus.read_blocks — Method

read_blocks(
    exo::ExodusDatabase,
    block_ids::Union{Vector{<:Integer}, var"#s123"} where var"#s123"<:Integer
) -> Any

Helper method for initializing blocks.

TODO: change name to readelementblocks

read_blocks(exo, block_ids)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Blocks.jl:128.

Exodus.read_coordinate_names — Method

read_coordinate_names(exo::ExodusDatabase) -> Any

Method to read coordinates names Returns a vector of strings

read_coordinate_names(exo)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Coordinates.jl:142.

Exodus.read_coordinates — Method

read_coordinates(exo::ExodusDatabase) -> Matrix

Method to read coordinates. Returns a matrix that is ndim x nnodes.

read_coordinates(exo)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Coordinates.jl:5.

Exodus.read_element_type — Method

read_element_type(
    exo::ExodusDatabase,
    block_id::Integer
) -> String

read_element_type(exo, block_id)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Blocks.jl:95.

Exodus.read_info — Method

read_info(exo::ExodusDatabase) -> Vector{String}

read_info(exo)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Info.jl:3.

Exodus.read_map — Method

read_map(exo::ExodusDatabase) -> Vector

read_map(exo)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Maps.jl:3.

Exodus.read_name — Method

read_name(
    exo::ExodusDatabase,
    _::Type{S<:Exodus.AbstractSet},
    id::Integer
) -> String

read_name(exo, _, id)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:26.

Exodus.read_name — Method

read_name(
    exo::ExodusDatabase,
    _::Type{V<:Exodus.AbstractVariable},
    var_index::Integer
) -> String

General method to read the name of a variable in index var_index for a given variable type V.

Examples: julia> readname(exo, ElementVariable, 1) "stressxx"

julia> readname(exo, GlobalVariable, 2) "reactionforce"

julia> readname(exo, NodalVariable, 1) "displx"

julia> readname(exo, NodeSetVariable, 1) "nsetdispl_x"

julia> read_name(exo, SideSetVariable, 1) "pressure"

read_name(exo, _, var_index)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:49.

Exodus.read_names — Method

read_names(
    exo::ExodusDatabase,
    _::Type{S<:Exodus.AbstractSet}
) -> Vector{String}

read_names(exo, _)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:53.

Exodus.read_names — Method

read_names(
    exo::ExodusDatabase,
    _::Type{V<:Exodus.AbstractVariable}
) -> Vector{String}

General method to read the names of variables for a given variable type V.

Examples: julia> readnames(exo, ElementVariable) "stressxx" "stressyy" "stresszz" "stressxy" "stressyz" "stress_zx"

julia> readnames(exo, GlobalVariable) "globaldispl" "reaction_force"

julia> readnames(exo, NodalVariable) "displx" "disply" "displz"

julia> readname(exo, NodeSetVariable) "nsetdisplx" "nsetdisply" "nsetdispl_z"

julia> read_name(exo, SideSetVariable) "pressure"

read_names(exo, _)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:91.

Exodus.read_node_set_nodes — Method

read_node_set_nodes(
    exo::ExodusDatabase{M, I, B, F},
    set_id::Integer
) -> Vector

read_node_set_nodes(exo, set_id)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:83.

Exodus.read_number_of_time_steps — Method

read_number_of_time_steps(exo::ExodusDatabase) -> UInt32

read_number_of_time_steps(exo)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Times.jl:3.

Exodus.read_number_of_variables — Method

read_number_of_variables(
    exo::ExodusDatabase,
    _::Type{V<:Exodus.AbstractVariable}
) -> Int32

General method to read the number of variables for a given variable type V.

Examples: julia> readnumberof_variables(exo, ElementVariable) 6

julia> readnumberof_variables(exo, GlobalVariable) 5

julia> readnumberof_variables(exo, NodalVariable) 3

julia> readnumberof_variables(exo, NodeSetVariable) 3

julia> readnumberof_variables(exo, SideSetVariable) 6

read_number_of_variables(exo, _)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:20.

Exodus.read_partial_block_connectivity — Method

read_partial_block_connectivity(
    exo::ExodusDatabase,
    block_id::Integer,
    start_num::Integer,
    num_ent::Integer
) -> Any

read_partial_block_connectivity(
    exo,
    block_id,
    start_num,
    num_ent
)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Blocks.jl:78.

Exodus.read_partial_coordinates — Method

read_partial_coordinates(
    exo::ExodusDatabase,
    start_node_num::Integer,
    num_nodes::Integer
) -> Matrix

Method to read a partial set of coordinates that are contiguous. Returns a matrix that is ndim x nnodes

read_partial_coordinates(exo, start_node_num, num_nodes)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Coordinates.jl:49.

Exodus.read_partial_coordinates_component — Method

read_partial_coordinates_component(
    exo::ExodusDatabase,
    start_node_num::Integer,
    num_nodes::Integer,
    component::Integer
) -> Vector

Method to read a specific component of a partial set of coordinates that are contiguous. Returns a vector of length n_nodes

read_partial_coordinates_component(
    exo,
    start_node_num,
    num_nodes,
    component
)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Coordinates.jl:108.

Exodus.read_qa — Method

read_qa(exo::ExodusDatabase) -> Matrix{String}

read_qa(exo)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/QA.jl:3.

Exodus.read_set — Method

read_set(
    exo::ExodusDatabase,
    type::Type{S<:Exodus.AbstractSet},
    set_id
) -> Any

read_set(exo, type, set_id)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:158.

Exodus.read_set_parameters — Method

read_set_parameters(
    exo::ExodusDatabase{M, I, B, F},
    set_id::Integer,
    _::Type{S<:Exodus.AbstractSet}
) -> Tuple{Any, Any}

read_set_parameters(exo, set_id, _)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:70.

Exodus.read_sets! — Method

read_sets!(
    sets::Array{T<:Exodus.AbstractSet, 1},
    exo::ExodusDatabase,
    ids::Array{I, 1}
)

read_sets!(sets, exo, ids)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:164.

Exodus.read_sets — Method

read_sets(
    exo::ExodusDatabase{M, I, B, F},
    type::Type{S<:Exodus.AbstractSet}
) -> Vector

read_sets(exo, type)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:180.

Exodus.read_side_set_elements_and_sides — Method

read_side_set_elements_and_sides(
    exo::ExodusDatabase{M, I, B, F},
    set_id::Integer
) -> Tuple{Vector, Vector}

read_side_set_elements_and_sides(exo, set_id)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:103.

Exodus.read_side_set_node_list — Method

read_side_set_node_list(
    exo::ExodusDatabase{M, I, B, F},
    side_set_id::Integer
) -> Tuple{Vector, Vector}

UNTESTED

read_side_set_node_list(exo, side_set_id)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:128.

Exodus.read_time — Method

read_time(exo::ExodusDatabase, time_step::Integer) -> Any

TODO figure out how to make this not use a vector of length 1 - either a ref or a ptr

read_time(exo, time_step)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Times.jl:12.

Exodus.read_times — Method

read_times(exo::ExodusDatabase) -> Vector

read_times(exo)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Times.jl:23.

Exodus.read_values — Method

read_values(
    exo::ExodusDatabase,
    t::Type{GlobalVariable},
    timestep::Integer
) -> Vector

Wrapper method for global variables around the main readvalues method readvalues(exo::ExodusDatabase, t::Type{GlobalVariable}, timestep::Integer) = read_values(exo, t, timestep, 1, 1)

Example: read_values(exo, GlobalVariable, 1)

read_values(exo, t, timestep)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:192.

Exodus.read_values — Method

read_values(
    exo::ExodusDatabase,
    t::Type{NodalVariable},
    timestep::Integer,
    index::Integer
) -> Vector

Wrapper method for nodal variables

read_values(exo, t, timestep, index)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:197.

Exodus.read_values — Method

read_values(
    exo::ExodusDatabase,
    t::Type{NodalVariable},
    timestep::Integer,
    name::String
) -> Vector

Wrapper method for nodal variables

read_values(exo, t, timestep, name)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:213.

Exodus.read_values — Method

read_values(
    exo::ExodusDatabase,
    _::Type{V<:Union{ElementVariable, NodalVariable, NodeSetVariable, SideSetVariable}},
    time_step::Integer,
    id::Integer,
    var_name::String
) -> Vector

read_values(exo, _, time_step, id, var_name)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:202.

Exodus.read_values — Method

read_values(
    exo::ExodusDatabase,
    _::Type{V<:Union{ElementVariable, NodeSetVariable, SideSetVariable}},
    time_step::Integer,
    set_name::String,
    var_name::String
) -> Vector

read_values(exo, _, time_step, set_name, var_name)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:218.

Exodus.read_values — Method

read_values(
    exo::ExodusDatabase{M, I, B, F},
    _::Type{V<:Exodus.AbstractVariable},
    timestep::Integer,
    id::Integer,
    var_index::Integer
) -> Vector

General method to read variable values.

read_values(exo, _, timestep, id, var_index)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:128.

Exodus.set_exodus_max_name_length — Method

set_exodus_max_name_length(exoid::Int32, len::Int32)

set_exodus_max_name_length(exoid, len)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExodusTypes.jl:21.

Exodus.set_exodus_options — Method

set_exodus_options(options)

set_exodus_options(options)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExodusTypes.jl:14.

Exodus.write_block — Method

write_block(exo::ExodusDatabase, block::Block)

WARNING: currently does not support edges, faces and attributes

write_block(exo, block)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Blocks.jl:145.

Exodus.write_block_connectivity — Method

write_block_connectivity(
    exo::ExodusDatabase,
    block_id::Integer,
    conn::Array{I<:Integer, 2}
)

write_block_connectivity(exo, block_id, conn)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Blocks.jl:61.

Exodus.write_blocks — Method

write_blocks(exo::ExodusDatabase, blocks::Vector{<:Block})

write_blocks(exo, blocks)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Blocks.jl:171.

Exodus.write_coordinate_names — Method

write_coordinate_names(
    exo::ExodusDatabase,
    coord_names::Vector{String}
)

Method to write coordinate names, e.g. x, y, z

write_coordinate_names(exo, coord_names)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Coordinates.jl:201.

Exodus.write_coordinates — Method

write_coordinates(
    exo::ExodusDatabase,
    coords::Union{Array{F<:AbstractFloat, 1}, Array{F<:AbstractFloat, 2}}
)

Method to write coordinates

write_coordinates(exo, coords)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Coordinates.jl:160.

Exodus.write_info — Method

write_info(exo::ExodusDatabase, info::Vector{String})

write_info(exo, info)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Info.jl:24.

Exodus.write_initialization! — Method

write_initialization!(exoid::Int32, init::Initialization)

Used to set up a exodus database in write mode

The ccall signatures should reall be B (bulk int type of exo) instead of Clonglong

write_initialization!(exoid, init)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/ExodusTypes.jl:177.

Exodus.write_name — Method

write_name(
    exo::ExodusDatabase{M, I, B, F},
    set::Exodus.AbstractSet,
    name::String
)

write_name(exo, set, name)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:248.

Exodus.write_name — Method

write_name(
    exo::ExodusDatabase{M, I, B, F},
    _::Type{S<:Exodus.AbstractSet},
    set_id::Integer,
    name::String
)

write_name(exo, _, set_id, name)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:229.

Exodus.write_name — Method

write_name(
    exo::ExodusDatabase,
    _::Type{V<:Exodus.AbstractVariable},
    var_index::Integer,
    var_name::String
)

write_name(exo, _, var_index, var_name)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:288.

Exodus.write_names — Method

write_names(
    exo::ExodusDatabase,
    _::Type{S<:Exodus.AbstractSet},
    names::Vector{String}
)

WARNING: this methods likely does not have good safe guards

write_names(exo, _, names)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:267.

Exodus.write_names — Method

write_names(
    exo::ExodusDatabase,
    _::Type{V<:Exodus.AbstractVariable},
    var_names::Vector{String}
)

write_names(exo, _, var_names)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:303.

Exodus.write_number_of_variables — Method

write_number_of_variables(
    exo::ExodusDatabase,
    _::Type{V<:Exodus.AbstractVariable},
    num_vars::Integer
)

General method to write the number of variables for a given variable type V.

Examples: julia> writenumberof_variables(exo, ElementVariable, 6)

julia> writenumberof_variables(exo, GlobalVariable, 5)

julia> writenumberof_variables(exo, NodalVariable, 3)

julia> writenumberof_variables(exo, NodeSetVariable, 3)

julia> writenumberof_variables(exo, SideSetVariable, 6)

write_number_of_variables(exo, _, num_vars)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:279.

Exodus.write_partial_coordinates — Method

write_partial_coordinates(
    exo::ExodusDatabase,
    start_node_num::Integer,
    coords::Union{Array{F<:AbstractFloat, 1}, Array{F<:AbstractFloat, 2}}
)

write_partial_coordinates(exo, start_node_num, coords)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Coordinates.jl:212.

Exodus.write_partial_coordinates_component — Method

write_partial_coordinates_component(
    exo::ExodusDatabase,
    start_node_num::Integer,
    component::Integer,
    coords::Array{F<:AbstractFloat, 1}
)

write_partial_coordinates_component(
    exo,
    start_node_num,
    component,
    coords
)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Coordinates.jl:239.

Exodus.write_partial_coordinates_component — Method

write_partial_coordinates_component(
    exo::ExodusDatabase,
    start_node_num::Integer,
    component::String,
    coords::Array{F<:AbstractFloat, 1}
)

write_partial_coordinates_component(
    exo,
    start_node_num,
    component,
    coords
)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Coordinates.jl:250.

Exodus.write_qa — Method

write_qa(exo::ExodusDatabase, qa_record::Matrix{String})

write_qa(exo, qa_record)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/QA.jl:29.

Exodus.write_set — Method

write_set(
    exo::ExodusDatabase{M, I, B, F},
    set::Exodus.AbstractSet
)

Typing ensures we don't write a set with non-matching types to the exodus file.

write_set(exo, set)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:210.

Exodus.write_set_parameters — Method

write_set_parameters(
    exo::ExodusDatabase{M, I, B, F},
    set::Exodus.AbstractSet
)

WARNING: currently doesn't support distance factors

write_set_parameters(exo, set)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:197.

Exodus.write_sets — Method

write_sets(
    exo::ExodusDatabase,
    sets::Array{T<:Exodus.AbstractSet, 1}
)

write_sets(exo, sets)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Sets.jl:221.

Exodus.write_time — Method

write_time(
    exo::ExodusDatabase,
    time_step::Integer,
    time_value::AbstractFloat
)

write_time(exo, time_step, time_value)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Times.jl:33.

Exodus.write_values — Method

write_values(
    exo::ExodusDatabase,
    t::Type{GlobalVariable},
    timestep::Integer,
    var_values::Vector{<:AbstractFloat}
)

Wrapper method for global variables around the main writevalues method writevalues( exo::ExodusDatabase, t::Type{GlobalVariable}, timestep::Integer, varvalues::Vector{<:AbstractFloat} ) = writevalues(exo, t, timestep, 1, 1, var_values)

Note: you need to first run writenumberof_variables(exo, GlobalVariable, n) where n is the number of variables.

Example: writenumberofvariables(exo, GlobalVariable, 5) writevalues(exo, GlobalVariable, 1, [10.0, 20.0, 30.0, 40.0, 50.0])

write_values(exo, t, timestep, var_values)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:351.

Exodus.write_values — Method

write_values(
    exo::ExodusDatabase,
    t::Type{NodalVariable},
    timestep::Integer,
    var_index::Integer,
    var_values::Vector{<:AbstractFloat}
)

Wrapper for writing nodal variables by index number

write_values(exo, t, timestep, var_index, var_values)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:360.

Exodus.write_values — Method

write_values(
    exo::ExodusDatabase,
    t::Type{NodalVariable},
    timestep::Integer,
    var_name::String,
    var_values::Vector{<:AbstractFloat}
)

Wrapper method for nodal variables

write_values(exo, t, timestep, var_name, var_values)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:381.

Exodus.write_values — Method

write_values(
    exo::ExodusDatabase,
    _::Type{V<:Exodus.AbstractVariable},
    timestep::Integer,
    id::Integer,
    var_index::Integer,
    var_values::Vector{<:AbstractFloat}
)

write_values(exo, _, timestep, id, var_index, var_values)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:317.

Exodus.write_values — Method

write_values(
    exo::ExodusDatabase,
    _::Type{V<:Exodus.AbstractVariable},
    timestep::Integer,
    id::Integer,
    var_name::String,
    var_value::Vector{<:AbstractFloat}
)

write_values(exo, _, timestep, id, var_name, var_value)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:364.

Exodus.write_values — Method

write_values(
    exo::ExodusDatabase,
    _::Type{V<:Exodus.AbstractVariable},
    time_step::Integer,
    set_name::String,
    var_name::String,
    var_value::Vector{<:AbstractFloat}
)

write_values(
    exo,
    _,
    time_step,
    set_name,
    var_name,
    var_value
)

defined at /home/runner/work/Exodus.jl/Exodus.jl/src/Variables.jl:385.

Exodus.@decomp — Macro

Exodus.@epu — Macro

Exodus.@exodiff — Macro