Exodus.jl

Documentation for Exodus.jl

Documentation

Exodus.Exodus — Module

Exodus

Exodus.ExoFloat — Type

ExoFloat

Union of different Exodus float types

Exodus.ExoInt — Type

ExoInt

Union of different Exodus integer types.

Exodus.Block — Type

Block{I <: ExoInt, B <: ExoInt}

Container for reading in blocks

Exodus.Block — Method

Block(exo::ExodusDatabase{M, I, B, F}, block_id::I) where {M <: ExoInt, I <: ExoInt,
                                                           B <: ExoInt, F <: ExoFloat}

Init method for block container.

Wraps ex_get_block! and ex_get_conn!

Exodus.ExodusDatabase — Type

ExodusDatabase{M <: ExoInt, I <: ExoInt, B <: ExoInt, F <: ExoFloat}

Main entry point for the package whether it's in read or write mode.

Exodus.ExodusDatabase — Method

ExodusDatabase(file_name::String, mode::String; int_mode="32-bit", float_mode="64-bit")

Init method.

Arguments

file_name::String: absolute path to exodus file

mode::String: mode to read

int_mode: either 32-bit or 64-bit

float_mode: either 32-bit or 64-bit

Exodus.Initialization — Type

Initialization

Container that should be setup first thing after getting an exo ID

Exodus.Initialization — Method

Initialization(exo::ExodusDatabase{M, I, B, F}) where {M <: ExoInt, I <: ExoInt, B <: ExoInt, F <: ExoFloat}

Exodus.NodeSet — Type

NodeSet{I <: ExoInt, B <: ExoInt}

Container for node sets.

Exodus.NodeSet — Method

NodeSet(exo::ExodusDatabase{M, I, B, F}, node_set_id::I) where {M <: ExoInt, I <: ExoInt,
                                                                B <: ExoInt, F <: ExoFloat}

Init method for a NodeSet with ID nodesetid.

Exodus.ex_entity_id — Type

ex_entity_id = Clonglong

Exodus.ex_entity_type — Type

ex_entity_type

Entity type enums (exentitytype in exodusII.h)

Exodus.ex_inquiry — Type

ex_inquiry

Inquiry enums (ex_inquiry in exodusII.h).

Exodus.void_int — Type

void_int = Cvoid

Base.close — Method

Base.close(exo::ExodusDatabase{M, I, B, F}) where {M <: ExoInt, I <: ExoInt, B <: ExoInt, F <: ExoFloat}

Used to close and ExodusDatabase.

Base.copy — Method

Base.copy(exo::ExodusDatabase{M, I, B, F},
          new_file_name::String) where {M <: ExoInt, I <: ExoInt, B <: ExoInt, F <: ExoFloat}

Used to copy an ExodusDatabase. As of right now this is the best way to create a new ExodusDatabase for output. Not all of the put methods have been wrapped and properly tested. This one has though.

Base.length — Method

Base.length(nset::NodeSet)

Exodus.ex_close! — Method

ex_close!(exoid::Cint)

Exodus.ex_copy! — Method

ex_copy!(in_exoid::Cint, out_exoid::Cint)

Exodus.ex_get_coord! — Method

ex_get_coord!(exoid::Cint, x_coords::Vector{T}, y_coords::Ptr{Cvoid}, z_coords::Ptr{Cvoid}) where {T <: ExoFloat}

Exodus.ex_get_coord! — Method

ex_get_coord!(exoid::Cint, x_coords::Vector{T}, y_coords::Vector{T}, z_coords::Ptr{Cvoid}) where {T <: ExoFloat}

Exodus.ex_get_coord! — Method

ex_get_coord!(exoid::Cint, x_coords::Vector{T}, y_coords::Vector{T}, z_coords::Vector{T}) where {T <: ExoFloat}

Exodus.ex_get_coord_names! — Method

ex_get_coord_names!(exo_id::Cint, coord_names::Vector{Vector{UInt8}})

Exodus.ex_get_id_map! — Method

ex_get_id_map!(exoid::Cint, map_type::ex_entity_type, map::Vector{T}) where {T <: ExoInt}

Exodus.ex_get_ids! — Method

ex_get_ids!(exoid::Cint, exo_const::ex_entity_type, ids::Vector{T}) where {T <: ExoInt}

Exodus.ex_get_init! — Method

ex_get_init!(exoid::Cint, 
             title::Vector{UInt8},
             num_dim::Ref{Clonglong}, num_nodes::Ref{Clonglong}, num_elem::Ref{Clonglong}, 
             num_elem_blk::Ref{Clonglong}, num_node_sets::Ref{Clonglong}, num_side_sets::Ref{Clonglong})

Exodus.ex_get_map! — Method

ex_get_map!(exoid::Cint, elem_map::Vector{T}) where {T <: ExoInt}

Exodus.ex_get_set_param! — Method

ex_get_set_param!(exoid::Cint, set_type::ex_entity_type, set_id::Cint, 
                  num_entry_in_set::Ref{T}, num_dist_fact_in_set::Ref{T}) where {T <: ExoInt}

Exodus.ex_get_set_param! — Method

ex_get_set_param!(exoid::Cint, set_type::ex_entity_type, set_id::Clonglong,
                  num_entry_in_set::Ref{T}, num_dist_fact_in_set::Ref{T}) where {T <: ExoInt}

Exodus.ex_get_var! — Method

ex_get_var!(exoid::Cint, time_step, var_type::ex_entity_type, var_index,
            obj_id::ex_entity_id, num_entry_this_obj, var_vals)

Exodus.ex_get_variable_name! — Method

ex_get_variable_name!(exoid::Cint, obj_type::ex_entity_type, var_num, var_name)

Exodus.ex_get_variable_param! — Method

ex_get_variable_param!(exoid::Cint, obj_type::ex_entity_type, num_vars)

Exodus.ex_inquire_int — Method

ex_inquire_int(exoid::Cint, req_info::ex_inquiry)

Exodus.ex_int64_status — Method

ex_int64_status(exoid::Cint)

Exodus.ex_open — Method

ex_open(path, mode, comp_ws, io_ws)::Cint

NOT USED

Exodus.ex_open_int — Method

ex_open_int(path, mode, comp_ws, io_ws, version, run_version)::Cint

FIX TYPES

Exodus.ex_opts — Method

ex_opts(options)

Exodus.ex_put_coord! — Method

ex_put_coord!(exoid::Cint, x_coords, y_coords, z_coords)

NOT THAT WELL TESTED

Exodus.ex_put_coord_names! — Method

ex_put_coord_names!(exoid::Cint, coord_names::Vector{Vector{UInt8}})

Exodus.exodus_error_check — Method

exodus_error_check(error_code::T)

Generic error handling method.

Arguments

error_code::T: error code, usually negative means something went bad

method_name::String: method name that called this

Exodus.put_coordinate_names — Method

put_coordinate_names(exo::ExodusDatabase{M, I, B, F}, 
                     coord_names::Vector{String}) where {M <: ExoInt, I <: ExoInt,
                                                         B <: ExoInt, F <: ExoFloat}

Work in progress...

Exodus.put_coordinates — Method

put_coordinates(exo::ExodusDatabase{M, I, B, F}, 
                     coords::Matrix{F}) where {M <: ExoInt, I <: ExoInt,
                                               B <: ExoInt, F <: ExoFloat}

Work in progress... not that well tested

Exodus.put_initialization — Method

put_initialization(exo::ExodusDatabase{M, I, B, F}, 
                   init::Initialization) where {M <: ExoInt, I <: ExoInt,
                                                B <: ExoInt, F <: ExoFloat}

Exodus.read_block_ids — Method

read_block_ids(exo::ExodusDatabase{M, I, B, F}, 
               init::Initialization) where {M <: ExoInt, I <: ExoInt, 
                                            B <: ExoInt, F <: ExoFloat}

Retrieves numerical block ids.

Wraps exgetids!

Exodus.read_blocks — Method

read_blocks(exo::ExodusDatabase{M, I, B, F}, 
            block_ids::Vector{I}) where {M <: ExoInt, I <: ExoInt,
                                         B <: ExoInt, F <: ExoFloat}

Helper method for initializing blocks.

Exodus.read_coordinate_names — Method

read_coordinate_names(exo::ExodusDatabase{M, I, B, F}, 
                      init::Initialization) where {M <: ExoInt, I <: ExoInt,
                                                   B <: ExoInt, F <: ExoFloat}

Exodus.read_coordinates — Method

read_coordinates(exo::ExodusDatabase{M, I, B, F}, 
                 init::Initialization) where {M <: ExoInt, I <: ExoInt,
                                              B <: ExoInt, F <: ExoFloat}

Method to read coordinates. Returns a matrix that is nnodes x ndim.

TODO... This method should really return ndim x nnodes but there's TODO... issues encountered downstream with some views found in Tardigrade

Exodus.read_element_map — Method

read_element_map(exo::ExodusDatabase{M, I, B, F}, init::Initialization) where {M <: ExoInt, I <: ExoInt,
                                                                               B <: ExoInt, F <: ExoFloat}

Exodus.read_node_set_ids — Method

read_node_set_ids(exo::ExodusDatabase{M, I, B, F},
                  init::Initialization) where {M <: ExoInt, I <: ExoInt,
                                               B <: ExoInt, F <: ExoFloat}

Exodus.read_node_sets — Method

read_node_sets(exo::ExodusDatabase{M, I, B, F}, 
               node_set_ids::Array{I}) where {M <: ExoInt, I <: ExoInt,
                                              B <: ExoInt, F <: ExoFloat}

Exodus.read_number_of_nodal_variables — Method

read_number_of_nodal_variables(exo::ExodusDatabase{M, I, B, F}) where {M <: ExoInt, I <: ExoInt,
                                                                       B <: ExoInt, F <: ExoFloat}

Exodus.read_number_of_time_steps — Method

read_number_of_time_steps(exo::ExodusDatabase{M, I, B, F}) where {M <: ExoInt, I <: ExoInt,
                                                                  B <: ExoInt, F <: ExoFloat}

Exodus.read_times — Method

read_times(exo::ExodusDatabase{M, I, B, F}) where {M <: ExoInt, I <: ExoInt,
                                                   B <: ExoInt, F <: ExoFloat}

Exodus.write_time — Method

write_time(exo::ExodusDatabase{M, I, B, F}, 
           time_step, time_value::F) where {M <: ExoInt, I <: ExoInt,
                                            B <: ExoInt, F <: ExoFloat}