Exodus.jl

Documentation for Exodus.jl

Documentation

Exodus.Exodus — Module

Block

ExodusDatabase

ExodusDatabase!

Initialization

NodeSet

SideSet

close

copy

@decomp

@epu

exo_float_type

exo_int_types

@exodiff

read_block_connectivity

read_block_ids

read_block_names

read_blocks

read_coordinate_names

read_coordinates

read_element_block_parameters

read_element_map

read_element_variable_name

read_element_variable_names

read_element_variable_values

read_global_variable_name

read_global_variable_names

read_global_variable_values

read_nodal_variable_name

read_nodal_variable_names

read_nodal_variable_values

read_node_set_ids

read_node_set_names

read_node_set_parameters

read_node_sets

read_number_of_element_variables

read_number_of_global_variables

read_number_of_nodal_variables

read_number_of_time_steps

read_side_set_elements_and_sides

read_side_set_ids

read_side_set_names

read_side_set_node_list

read_side_set_parameters

read_times

write_coordinate_names

write_coordinates

write_element_variable_name

write_element_variable_names

write_element_variable_values

write_global_variable_name

write_global_variable_names

write_global_variable_values

write_initialization!

write_nodal_variable_name

write_nodal_variable_names

write_nodal_variable_values

write_node_set_names

write_number_of_element_variables

write_number_of_global_variables

write_number_of_nodal_variables

write_time

Exodus.Block — Type

block_id::Integer: ID of block

num_elem::Int64: Number of elements in block

num_nodes_per_elem::Int64: Number of nodes per element in block

elem_type::String: Type of finite element, e.g. QUAD4

conn::Matrix{B} where B<:Integer: Connectivty of integer type consistent with exodus database

Exodus.Block — Method

Init method for block container.

Wraps ex_get_block! and ex_get_conn!

Exodus.Block — Method

Exodus.ExodusDatabase — Method

ExodusDatabase(
    file_name::String,
    mode::String
) -> ExodusDatabase

Init method for read/read-write

ExodusDatabase(file_name, mode)

Exodus.Initialization — Type

num_dim::Int64: Dimension of exodus database

num_nodes::Int64: Number of nodes in exodus database

num_elems::Int64: Number of elements in exodus database

num_elem_blks::Int64: Number of element blocks in exodus database

num_node_sets::Int64: Number of node sets in exodus database

num_side_sets::Int64: Number of side sets in exodus database

Exodus.Initialization — Method

Initialization(exo_id::Integer) -> Initialization

Initialization(exo_id)

Exodus.NodeSet — Type

node_set_id::Integer: ID of node set

num_nodes::Int64: Number of nodes in the node set

nodes::Vector{B} where B<:Integer: Node IDs for each node in the node set

Exodus.NodeSet — Method

Init method for a NodeSet with ID nodesetid.

Exodus.NodeSet — Method

Exodus.SideSet — Type

side_set_id::Integer: ID of side set

num_elements::Int64: Number of elements in the side set

elements::Vector{B} where B<:Integer: Element IDs

sides::Vector{B} where B<:Integer: Sides

Exodus.SideSet — Method

Exodus.SideSet — Method

Exodus.ex_entity_id — Type

exentityid = Clonglong

Exodus.ex_entity_type — Type

ex_entity_type

Entity type enums (exentitytype in exodusII.h)

Exodus.ex_inquiry — Type

ex_inquiry

Inquiry enums (ex_inquiry in exodusII.h).

Exodus.void_int — Type

void_int = Cvoid

Base.close — Method

close(exo::ExodusDatabase)

Used to close and ExodusDatabase.

close(exo)

Base.copy — Method

copy(exo::ExodusDatabase, new_file_name::String)

Used to copy an ExodusDatabase. As of right now this is the best way to create a new ExodusDatabase for output. Not all of the put methods have been wrapped and properly tested. This one has though.

copy(exo, new_file_name)

Base.length — Method

length(nset::NodeSet) -> Int64

length(nset)

Base.length — Method

length(sset::SideSet) -> Int64

length(sset)

Base.show — Method

show(io::IO, init::Initialization)

show(io, init)

Base.show — Method

show(io::IO, sset::SideSet)

show(io, sset)

Base.show — Method

show(io::IO, node_set::NodeSet)

show(io, node_set)

Exodus.ex_put_coord! — Method

ex_put_coord!(
    exoid::Int32,
    x_coords::Union{Ptr, Vector{<:Real}},
    y_coords::Union{Ptr, Vector{<:Real}},
    z_coords::Union{Ptr, Vector{<:Real}}
)

NOT THAT WELL TESTED

ex_put_coord!(exoid, x_coords, y_coords, z_coords)

Exodus.exodus_error_check — Method

exodus_error_check(error_code::Integer, method_name::String)

Generic error handling method.

Arguments

error_code::T: error code, usually negative means something went bad

method_name::String: method name that called this

exodus_error_check(error_code, method_name)

Exodus.read_block_connectivity — Method

read_block_connectivity(
    exo::ExodusDatabase,
    block_id::Integer
) -> Any

read_block_connectivity(exo, block_id)

Exodus.read_block_ids — Method

read_block_ids(exo::ExodusDatabase) -> Vector

Retrieves numerical block ids.

Wraps exgetids!

read_block_ids(exo)

Exodus.read_block_names — Method

read_block_names(exo::ExodusDatabase) -> Vector{String}

read_block_names(exo)

Exodus.read_blocks! — Method

read_blocks!(
    blocks::Array{B<:Block, 1},
    exo::ExodusDatabase,
    block_ids::Array{I<:Integer, 1}
)

read_blocks!(blocks, exo, block_ids)

Exodus.read_blocks — Method

read_blocks(
    exo::ExodusDatabase,
    block_ids::Union{Vector{<:Integer}, var"#s34"} where var"#s34"<:Integer
) -> Any

Helper method for initializing blocks.

read_blocks(exo, block_ids)

Exodus.read_coordinate_names — Method

read_coordinate_names(exo::ExodusDatabase) -> Vector{String}

read_coordinate_names(exo)

Exodus.read_coordinates — Method

read_coordinates(exo::ExodusDatabase) -> Any

Method to read coordinates. Returns a matrix that is ndim x nnodes.

read_coordinates(exo)

Exodus.read_element_block_parameters — Method

read_element_block_parameters(
    exo::ExodusDatabase,
    block_id::Integer
) -> Tuple{String, Any, Any, Any, Any, Any}

read_element_block_parameters(exo, block_id)

Exodus.read_element_map — Method

read_element_map(exo::ExodusDatabase) -> Vector

read_element_map(exo)

Exodus.read_element_variable_name — Method

read_element_variable_name(
    exo::ExodusDatabase,
    var_index::Integer
) -> String

read_element_variable_name(exo, var_index)

Exodus.read_element_variable_names — Method

read_element_variable_names(
    exo::ExodusDatabase
) -> Vector{String}

read_element_variable_names(exo)

Exodus.read_element_variable_values — Method

read_element_variable_values(
    exo::ExodusDatabase,
    time_step,
    block_id::Integer,
    variable_index::Integer
) -> Any

read_element_variable_values(
    exo,
    time_step,
    block_id,
    variable_index
)

Exodus.read_element_variable_values — Method

read_element_variable_values(
    exo::ExodusDatabase,
    time_step,
    block_id::Integer,
    var_name::String
) -> Any

read_element_variable_values(
    exo,
    time_step,
    block_id,
    var_name
)

Exodus.read_global_variable_name — Method

read_global_variable_name(
    exo::ExodusDatabase,
    var_index::Integer
) -> String

read_global_variable_name(exo, var_index)

Exodus.read_global_variable_names! — Method

read_global_variable_names!(
    exo::ExodusDatabase,
    num_vars::Int32,
    var_name::Vector{UInt8},
    var_names::Vector{String}
)

read_global_variable_names!(
    exo,
    num_vars,
    var_name,
    var_names
)

Exodus.read_global_variable_names — Method

read_global_variable_names(
    exo::ExodusDatabase
) -> Vector{String}

read_global_variable_names(exo)

Exodus.read_global_variable_values — Method

read_global_variable_values(
    exo::ExodusDatabase,
    timestep::Integer,
    num_vars::Integer
) -> Vector

read_global_variable_values(exo, timestep, num_vars)

Exodus.read_nodal_variable_name — Method

read_nodal_variable_name(
    exo::ExodusDatabase,
    var_index::Integer
) -> String

read_nodal_variable_name(exo, var_index)

Exodus.read_nodal_variable_names — Method

read_nodal_variable_names(
    exo::ExodusDatabase
) -> Vector{String}

read_nodal_variable_names(exo)

Exodus.read_nodal_variable_values — Method

read_nodal_variable_values(
    exo::ExodusDatabase,
    time_step,
    var_name::String
) -> Vector

read_nodal_variable_values(exo, time_step, var_name)

Exodus.read_nodal_variable_values — Method

read_nodal_variable_values(
    exo::ExodusDatabase,
    time_step,
    variable_index::Integer
) -> Vector

read_nodal_variable_values(exo, time_step, variable_index)

Exodus.read_node_set_ids — Method

read_node_set_ids(exo::ExodusDatabase) -> Vector

read_node_set_ids(exo)

Exodus.read_node_set_names — Method

read_node_set_names(exo::ExodusDatabase) -> Vector{String}

read_node_set_names(exo)

Exodus.read_node_set_nodes — Method

read_node_set_nodes(
    exo::ExodusDatabase,
    node_set_id::Integer
) -> Any

read_node_set_nodes(exo, node_set_id)

Exodus.read_node_set_parameters — Method

read_node_set_parameters(
    exo::ExodusDatabase,
    node_set_id::Integer
) -> Tuple{Any, Any}

read_node_set_parameters(exo, node_set_id)

Exodus.read_node_sets! — Method

read_node_sets!(
    node_sets::Vector{NodeSet},
    exo::ExodusDatabase,
    node_set_ids::Vector{<:Integer}
)

read_node_sets!(node_sets, exo, node_set_ids)

Exodus.read_node_sets — Method

read_node_sets(
    exo::ExodusDatabase,
    node_set_ids::Array{<:Integer}
) -> Any

read_node_sets(exo, node_set_ids)

Exodus.read_number_of_element_variables — Method

read_number_of_element_variables(
    exo::ExodusDatabase
) -> Int32

read_number_of_element_variables(exo)

Exodus.read_number_of_global_variables — Method

read_number_of_global_variables(
    exo::ExodusDatabase
) -> Int32

read_number_of_global_variables(exo)

Exodus.read_number_of_nodal_variables — Method

read_number_of_nodal_variables(exo::ExodusDatabase) -> Int32

read_number_of_nodal_variables(exo)

Exodus.read_number_of_time_steps — Method

read_number_of_time_steps(exo::ExodusDatabase) -> Int32

read_number_of_time_steps(exo)

Exodus.read_side_set_elements_and_sides — Method

read_side_set_elements_and_sides(
    exo::ExodusDatabase,
    side_set_id::Integer
) -> Tuple{Any, Any}

read_side_set_elements_and_sides(exo, side_set_id)

Exodus.read_side_set_ids — Method

read_side_set_ids(exo::ExodusDatabase) -> Any

read_side_set_ids(exo)

Exodus.read_side_set_names — Method

read_side_set_names(exo::ExodusDatabase) -> Any

read_side_set_names(exo)

Exodus.read_side_set_parameters — Method

read_side_set_parameters(
    exo::ExodusDatabase,
    side_set_id::Integer
) -> Tuple{Any, Any}

read_side_set_parameters(exo, side_set_id)

Exodus.read_times — Method

read_times(exo::ExodusDatabase) -> Vector

read_times(exo)

Exodus.write_coordinate_names — Method

write_coordinate_names(
    exo::ExodusDatabase,
    coord_names::Vector{String}
)

Work in progress...

write_coordinate_names(exo, coord_names)

Exodus.write_coordinates — Method

write_coordinates(
    exo::ExodusDatabase,
    coords::Array{F<:Real, 2}
)

Work in progress... not that well tested

write_coordinates(exo, coords)

Exodus.write_element_variable_name — Method

write_element_variable_name(
    exo::ExodusDatabase,
    var_index::Integer,
    var_name::String
)

write_element_variable_name(exo, var_index, var_name)

Exodus.write_element_variable_names — Method

write_element_variable_names(
    exo::ExodusDatabase,
    var_indices::Vector{<:Integer},
    var_names::Vector{String}
)

write_element_variable_names(exo, var_indices, var_names)

Exodus.write_element_variable_values — Method

write_element_variable_values(
    exo::ExodusDatabase,
    time_step::Integer,
    block_id::Integer,
    var_index::Integer,
    var_values::Vector{<:Real}
)

write_element_variable_values(
    exo,
    time_step,
    block_id,
    var_index,
    var_values
)

Exodus.write_element_variable_values — Method

write_element_variable_values(
    exo::ExodusDatabase,
    time_step::Integer,
    block_id::Integer,
    var_name::String,
    var_values::Vector{<:Real}
)

write_element_variable_values(
    exo,
    time_step,
    block_id,
    var_name,
    var_values
)

Exodus.write_global_variable_name — Method

write_global_variable_name(
    exo::ExodusDatabase,
    var_index::Integer,
    var_name::String
)

write_global_variable_name(exo, var_index, var_name)

Exodus.write_global_variable_names — Method

write_global_variable_names(
    exo::ExodusDatabase,
    var_indices::Vector{<:Integer},
    var_names::Vector{String}
)

write_global_variable_names(exo, var_indices, var_names)

Exodus.write_global_variable_values — Method

write_global_variable_values(
    exo::ExodusDatabase,
    timestep::Integer,
    num_vars::Integer,
    var_values::Vector{<:Real}
)

write_global_variable_values(
    exo,
    timestep,
    num_vars,
    var_values
)

Exodus.write_initialization! — Method

write_initialization!(exoid::Int32, init::Initialization)

Used to set up a exodus database in write mode

write_initialization!(exoid, init)

Exodus.write_nodal_variable_name — Method

write_nodal_variable_name(
    exo::ExodusDatabase,
    var_index::Integer,
    var_name::String
)

write_nodal_variable_name(exo, var_index, var_name)

Exodus.write_nodal_variable_names — Method

write_nodal_variable_names(
    exo::ExodusDatabase,
    var_indices::Vector{<:Integer},
    var_names::Vector{String}
)

write_nodal_variable_names(exo, var_indices, var_names)

Exodus.write_nodal_variable_values — Method

write_nodal_variable_values(
    exo::ExodusDatabase,
    time_step,
    var_index::Integer,
    var_values::Vector{<:Real}
)

write_nodal_variable_values(
    exo,
    time_step,
    var_index,
    var_values
)

Exodus.write_nodal_variable_values — Method

write_nodal_variable_values(
    exo::ExodusDatabase,
    time_step,
    var_name::String,
    var_value::Vector{<:Real}
)

write_nodal_variable_values(
    exo,
    time_step,
    var_name,
    var_value
)

Exodus.write_number_of_element_variables — Method

write_number_of_element_variables(
    exo::ExodusDatabase,
    num_vars::Integer
)

write_number_of_element_variables(exo, num_vars)

Exodus.write_number_of_global_variables — Method

write_number_of_global_variables(
    exo::ExodusDatabase,
    num_vars
)

write_number_of_global_variables(exo, num_vars)

Exodus.write_number_of_nodal_variables — Method

write_number_of_nodal_variables(
    exo::ExodusDatabase,
    num_vars
)

write_number_of_nodal_variables(exo, num_vars)

Exodus.write_time — Method

write_time(
    exo::ExodusDatabase,
    time_step::Integer,
    time_value::Real
)

write_time(exo, time_step, time_value)

Exodus.@decomp — Macro

Exodus.@exodiff — Macro