Exodus.jl

Documentation for Exodus.jl

Documentation

Exodus.Exodus — Module

Block

ExodusDatabase

Initialization

NodeSet

SideSet

close

copy

@decomp

@epu

exo_float_type

exo_int_types

@exodiff

length

read_coordinate_names

read_coordinates

read_element_block_connectivity

read_element_block_id_map

read_element_block_ids

read_element_block_names

read_element_block_parameters

read_element_blocks

read_element_map

read_element_type

read_element_variable_name

read_element_variable_names

read_element_variable_values

read_global_variable_name

read_global_variable_names

read_global_variable_values

read_info

read_nodal_variable_name

read_nodal_variable_names

read_nodal_variable_values

read_node_set_ids

read_node_set_names

read_node_set_parameters

read_node_set_variable_name

read_node_set_variable_names

read_node_set_variable_values

read_node_sets

read_number_of_element_variables

read_number_of_global_variables

read_number_of_nodal_variables

read_number_of_node_set_variables

read_number_of_side_set_variables

read_number_of_time_steps

read_partial_coordinates

read_partial_coordinates_component

read_partial_element_block_connectivity

read_partial_nodal_variable_values

read_qa

read_side_set_elements_and_sides

read_side_set_ids

read_side_set_names

read_side_set_node_list

read_side_set_parameters

read_side_set_variable_name

read_side_set_variable_names

read_side_set_variable_values

read_side_sets

read_time

read_times

set_exodus_max_name_length

set_exodus_options

write_coordinate_names

write_coordinates

write_element_block

write_element_block_connectivity

write_element_block_name

write_element_block_names

write_element_blocks

write_element_variable_name

write_element_variable_names

write_element_variable_values

write_global_variable_name

write_global_variable_names

write_global_variable_values

write_info

write_initialization!

write_nodal_variable_name

write_nodal_variable_names

write_nodal_variable_values

write_node_set

write_node_set_name

write_node_set_names

write_node_set_variable_name

write_node_set_variable_names

write_node_set_variable_values

write_node_sets

write_number_of_element_variables

write_number_of_global_variables

write_number_of_nodal_variables

write_number_of_node_set_variables

write_number_of_side_set_variables

write_partial_coordinates

write_partial_coordinates_component

write_qa

write_side_set

write_side_set_name

write_side_set_names

write_side_set_variable_name

write_side_set_variable_names

write_side_set_variable_values

write_side_sets

write_time

Exodus.AbstractSet — Type

Exodus.Block — Type

block_id::Integer: ID of block

num_elem::Int64: Number of elements in block

num_nodes_per_elem::Int64: Number of nodes per element in block

elem_type::String: Type of finite element, e.g. QUAD4

conn::Matrix{B} where B<:Integer: Connectivty of integer type consistent with exodus database

Exodus.Block — Method

Init method for block container.

Exodus.Block — Method

Exodus.ExodusDatabase — Method

Exodus.ExodusDatabase — Method

Init method for read/read-write

Exodus.ExodusDatabase — Method

Exodus.Initialization — Method

Workaround method

Exodus.Initialization — Method

Workaround method

Exodus.Initialization — Method

Exodus.NodeSet — Type

id::Any

nodes::Vector

Exodus.NodeSet — Method

Exodus.NodeSet — Method

Exodus.SideSet — Type

id::Any

elements::Vector

sides::Vector

Exodus.SideSet — Method

Exodus.SideSet — Method

Exodus.ex_entity_id — Type

Exodus.ex_entity_type — Type

ex_entity_type

Entity type enums (exentitytype in exodusII.h)

Exodus.ex_inquiry — Type

ex_inquiry

Inquiry enums (ex_inquiry in exodusII.h).

Exodus.void_int — Type

void_int = Cvoid

Base.close — Method

close(exo::ExodusDatabase)

Used to close and ExodusDatabase.

close(exo)

Base.copy — Method

copy(exo::ExodusDatabase, new_file_name::String)

Used to copy an ExodusDatabase. As of right now this is the best way to create a new ExodusDatabase for output. Not all of the put methods have been wrapped and properly tested. This one has though.

copy(exo, new_file_name)

Base.length — Method

length(n::NodeSet) -> Int64

length(n)

Base.length — Method

length(s::SideSet) -> Int64

length(s)

Base.show — Method

show(io::IO, init::Initialization)

show(io, init)

Base.show — Method

show(io::IO, node_set::NodeSet)

show(io, node_set)

Base.show — Method

show(io::IO, sset::SideSet)

show(io, sset)

Base.show — Method

show(io::IO, block::Block)

show(io, block)

Exodus.entries — Method

entries(n::NodeSet) -> Vector

entries(n)

Exodus.entries — Method

entries(s::SideSet) -> Vector

entries(s)

Exodus.exo_float_type — Method

exo_float_type(
    exoid::Int32
) -> Union{Type{Float32}, Type{Float64}}

exo_float_type(exoid)

Exodus.exo_int_types — Method

exo_int_types(
    exoid::Int32
) -> Tuple{DataType, DataType, DataType}

exo_int_types(exoid)

Exodus.exodus_error_check — Method

exodus_error_check(error_code::Integer, method_name::String)

Generic error handling method.

Arguments

error_code::T: error code, usually negative means something went bad

method_name::String: method name that called this

exodus_error_check(error_code, method_name)

Exodus.extras — Method

extras(n::NodeSet) -> Ptr{Nothing}

extras(n)

Exodus.extras — Method

extras(s::SideSet) -> Vector

extras(s)

Exodus.read_coordinate_names — Method

read_coordinate_names(exo::ExodusDatabase) -> Any

Method to read coordinates names Returns a vector of strings

read_coordinate_names(exo)

Exodus.read_coordinates — Method

read_coordinates(exo::ExodusDatabase) -> Matrix

Method to read coordinates. Returns a matrix that is ndim x nnodes.

read_coordinates(exo)

Exodus.read_element_block_connectivity — Method

read_element_block_connectivity(
    exo::ExodusDatabase{M, I, B, F},
    block_id::Integer
) -> Any

read_element_block_connectivity(exo, block_id)

Exodus.read_element_block_id_map — Method

read_element_block_id_map(
    exo::ExodusDatabase,
    block_id::Integer
) -> Vector

read_element_block_id_map(exo, block_id)

Exodus.read_element_block_ids — Method

read_element_block_ids(
    exo::ExodusDatabase{M, I, B, F}
) -> Any

Retrieves numerical block ids.

TODO: this should probably be wrapped into sets

or put the method re-definition here. It's the same method for betting nset ids

read_element_block_ids(exo)

Exodus.read_element_block_names — Method

read_element_block_names(exo::ExodusDatabase) -> Any

read_element_block_names(exo)

Exodus.read_element_block_parameters — Method

read_element_block_parameters(
    exo::ExodusDatabase{M, I, B, F},
    block_id::Integer
) -> Tuple{String, Any, Any, Any, Any, Any}

read_element_block_parameters(exo, block_id)

Exodus.read_element_blocks! — Method

read_element_blocks!(
    blocks::Array{B<:Block, 1},
    exo::ExodusDatabase,
    block_ids::Array{I<:Integer, 1}
)

TODO: change name to readelementblocks!

read_element_blocks!(blocks, exo, block_ids)

Exodus.read_element_blocks — Method

read_element_blocks(
    exo::ExodusDatabase,
    block_ids::Union{Vector{<:Integer}, var"#s78"} where var"#s78"<:Integer
) -> Any

Helper method for initializing blocks.

TODO: change name to readelementblocks

read_element_blocks(exo, block_ids)

Exodus.read_element_map — Method

read_element_map(exo::ExodusDatabase) -> Any

read_element_map(exo)

Exodus.read_element_type — Method

read_element_type(
    exo::ExodusDatabase,
    block_id::Integer
) -> String

read_element_type(exo, block_id)

Exodus.read_element_variable_name — Method

read_element_variable_name(
    exo::ExodusDatabase,
    var_index::Integer
) -> String

read_element_variable_name(exo, var_index)

Exodus.read_element_variable_names — Method

read_element_variable_names(
    exo::ExodusDatabase
) -> Vector{String}

read_element_variable_names(exo)

Exodus.read_element_variable_values — Method

read_element_variable_values(
    exo::ExodusDatabase,
    timestep::Integer,
    id::Integer,
    var_index::Integer
) -> Any

read_element_variable_values(exo, timestep, id, var_index)

Exodus.read_element_variable_values — Method

read_element_variable_values(
    exo::ExodusDatabase,
    timestep::Integer,
    id::Integer,
    var_name::String
) -> Any

read_element_variable_values(exo, timestep, id, var_name)

Exodus.read_global_variable_name — Method

read_global_variable_name(
    exo::ExodusDatabase,
    var_index::Integer
) -> String

read_global_variable_name(exo, var_index)

Exodus.read_global_variable_names — Method

read_global_variable_names(
    exo::ExodusDatabase
) -> Vector{String}

read_global_variable_names(exo)

Exodus.read_global_variable_values — Method

read_global_variable_values(
    exo::ExodusDatabase,
    timestep::Integer
) -> Any

read_global_variable_values(exo, timestep)

Exodus.read_nodal_variable_name — Method

read_nodal_variable_name(
    exo::ExodusDatabase,
    var_index::Integer
) -> String

read_nodal_variable_name(exo, var_index)

Exodus.read_nodal_variable_names — Method

read_nodal_variable_names(
    exo::ExodusDatabase
) -> Vector{String}

read_nodal_variable_names(exo)

Exodus.read_nodal_variable_values — Method

read_nodal_variable_values(
    exo::ExodusDatabase,
    timestep::Integer,
    var_index::Integer
) -> Any

read_nodal_variable_values(exo, timestep, var_index)

Exodus.read_nodal_variable_values — Method

read_nodal_variable_values(
    exo::ExodusDatabase,
    timestep::Integer,
    var_name::String
) -> Any

read_nodal_variable_values(exo, timestep, var_name)

Exodus.read_node_set_ids — Method

read_node_set_ids(exo::ExodusDatabase) -> Any

read_node_set_ids(exo)

Exodus.read_node_set_names — Method

read_node_set_names(exo::ExodusDatabase) -> Any

read_node_set_names(exo)

Exodus.read_node_set_nodes — Method

read_node_set_nodes(
    exo::ExodusDatabase{M, I, B, F},
    set_id::Integer
) -> Any

read_node_set_nodes(exo, set_id)

Exodus.read_node_set_parameters — Method

read_node_set_parameters(
    exo::ExodusDatabase,
    set_id::Integer
) -> Tuple{Any, Any}

read_node_set_parameters(exo, set_id)

Exodus.read_node_set_variable_name — Method

read_node_set_variable_name(
    exo::ExodusDatabase,
    var_index::Integer
) -> String

read_node_set_variable_name(exo, var_index)

Exodus.read_node_set_variable_names — Method

read_node_set_variable_names(
    exo::ExodusDatabase
) -> Vector{String}

read_node_set_variable_names(exo)

Exodus.read_node_set_variable_values — Method

read_node_set_variable_values(
    exo::ExodusDatabase,
    timestep::Integer,
    id::Integer,
    var_index::Integer
) -> Any

read_node_set_variable_values(exo, timestep, id, var_index)

Exodus.read_node_set_variable_values — Method

read_node_set_variable_values(
    exo::ExodusDatabase,
    timestep::Integer,
    id::Integer,
    var_name::String
) -> Any

read_node_set_variable_values(exo, timestep, id, var_name)

Exodus.read_node_set_variable_values — Method

read_node_set_variable_values(
    exo::ExodusDatabase,
    timestep::Integer,
    set_name::String,
    var_name::String
) -> Any

read_node_set_variable_values(
    exo,
    timestep,
    set_name,
    var_name
)

Exodus.read_node_sets — Method

read_node_sets(
    exo::ExodusDatabase,
    set_ids::Vector{<:Integer}
) -> Vector{NodeSet}

read_node_sets(exo, set_ids)

Exodus.read_number_of_element_variables — Method

read_number_of_element_variables(
    exo::ExodusDatabase
) -> Int32

read_number_of_element_variables(exo)

Exodus.read_number_of_global_variables — Method

read_number_of_global_variables(
    exo::ExodusDatabase
) -> Int32

read_number_of_global_variables(exo)

Exodus.read_number_of_nodal_variables — Method

read_number_of_nodal_variables(exo::ExodusDatabase) -> Int32

read_number_of_nodal_variables(exo)

Exodus.read_number_of_node_set_variables — Method

read_number_of_node_set_variables(
    exo::ExodusDatabase
) -> Int32

read_number_of_node_set_variables(exo)

Exodus.read_number_of_side_set_variables — Method

read_number_of_side_set_variables(
    exo::ExodusDatabase
) -> Int32

read_number_of_side_set_variables(exo)

Exodus.read_number_of_time_steps — Method

read_number_of_time_steps(exo::ExodusDatabase) -> UInt32

read_number_of_time_steps(exo)

Exodus.read_number_of_variables — Method

read_number_of_variables(
    exo::ExodusDatabase,
    type::Exodus.ex_entity_type
) -> Int32

read_number_of_variables(exo, type)

Exodus.read_partial_coordinates — Method

read_partial_coordinates(
    exo::ExodusDatabase,
    start_node_num::Integer,
    num_nodes::Integer
) -> Matrix

Method to read a partial set of coordinates that are contiguous. Returns a matrix that is ndim x nnodes

read_partial_coordinates(exo, start_node_num, num_nodes)

Exodus.read_partial_coordinates_component — Method

read_partial_coordinates_component(
    exo::ExodusDatabase,
    start_node_num::Integer,
    num_nodes::Integer,
    component::Integer
) -> Vector

Method to read a specific component of a partial set of coordinates that are contiguous. Returns a vector of length n_nodes

read_partial_coordinates_component(
    exo,
    start_node_num,
    num_nodes,
    component
)

Exodus.read_partial_element_block_connectivity — Method

read_partial_element_block_connectivity(
    exo::ExodusDatabase,
    block_id::Integer,
    start_num::Integer,
    num_ent::Integer
) -> Any

read_partial_element_block_connectivity(
    exo,
    block_id,
    start_num,
    num_ent
)

Exodus.read_partial_nodal_variable_values — Method

read_partial_nodal_variable_values(
    exo::ExodusDatabase,
    time_step::Integer,
    var_index::Integer,
    start_node::Integer,
    num_nodes::Integer
) -> Vector

read_partial_nodal_variable_values(
    exo,
    time_step,
    var_index,
    start_node,
    num_nodes
)

Exodus.read_partial_nodal_variable_values — Method

read_partial_nodal_variable_values(
    exo::ExodusDatabase,
    time_step::Integer,
    var_name::String,
    start_node::Integer,
    num_nodes::Integer
) -> Vector

read_partial_nodal_variable_values(
    exo,
    time_step,
    var_name,
    start_node,
    num_nodes
)

Exodus.read_partial_variable_values — Method

read_partial_variable_values(
    exo::ExodusDatabase,
    time_step::Integer,
    id::Integer,
    var_name::String,
    start_node::Integer,
    num_nodes::Integer,
    type::Exodus.ex_entity_type
) -> Vector

read_partial_variable_values(
    exo,
    time_step,
    id,
    var_name,
    start_node,
    num_nodes,
    type
)

Exodus.read_partial_variable_values — Method

read_partial_variable_values(
    exo::ExodusDatabase{M, I, B, F},
    time_step::Integer,
    id::Integer,
    var_index::Integer,
    start_node::Integer,
    num_nodes::Integer,
    type::Exodus.ex_entity_type
) -> Vector

read_partial_variable_values(
    exo,
    time_step,
    id,
    var_index,
    start_node,
    num_nodes,
    type
)

Exodus.read_set_ids — Method

read_set_ids(
    exo::ExodusDatabase{M, I, B, F},
    type::Exodus.ex_entity_type
) -> Any

read_set_ids(exo, type)

Exodus.read_set_names — Method

read_set_names(
    exo::ExodusDatabase,
    type::Exodus.ex_entity_type
) -> Any

read_set_names(exo, type)

Exodus.read_set_parameters — Method

read_set_parameters(
    exo::ExodusDatabase{M, I, B, F},
    set_id::Integer,
    type::Exodus.ex_entity_type
) -> Tuple{Any, Any}

read_set_parameters(exo, set_id, type)

Exodus.read_sets! — Method

read_sets!(
    sets::Array{T<:Exodus.AbstractSet, 1},
    exo::ExodusDatabase,
    set_ids::Array{I<:Integer, 1}
)

read_sets!(sets, exo, set_ids)

Exodus.read_sets — Method

read_sets(
    exo::ExodusDatabase,
    set_ids::Array{I<:Integer, 1},
    type::Type{T<:Exodus.AbstractSet}
) -> Vector

read_sets(exo, set_ids, type)

Exodus.read_side_set_elements_and_sides — Method

read_side_set_elements_and_sides(
    exo::ExodusDatabase{M, I, B, F},
    set_id::Integer
) -> Tuple{Any, Any}

read_side_set_elements_and_sides(exo, set_id)

Exodus.read_side_set_ids — Method

read_side_set_ids(exo::ExodusDatabase) -> Any

read_side_set_ids(exo)

Exodus.read_side_set_names — Method

read_side_set_names(exo::ExodusDatabase) -> Any

read_side_set_names(exo)

Exodus.read_side_set_node_list — Method

read_side_set_node_list(
    exo::ExodusDatabase{M, I, B, F},
    side_set_id::Integer
) -> Tuple{Any, Vector}

read_side_set_node_list(exo, side_set_id)

Exodus.read_side_set_parameters — Method

read_side_set_parameters(
    exo::ExodusDatabase,
    set_id::Integer
) -> Tuple{Any, Any}

read_side_set_parameters(exo, set_id)

Exodus.read_side_set_variable_name — Method

read_side_set_variable_name(
    exo::ExodusDatabase,
    var_index::Integer
) -> String

read_side_set_variable_name(exo, var_index)

Exodus.read_side_set_variable_names — Method

read_side_set_variable_names(
    exo::ExodusDatabase
) -> Vector{String}

read_side_set_variable_names(exo)

Exodus.read_side_set_variable_values — Method

read_side_set_variable_values(
    exo::ExodusDatabase,
    timestep::Integer,
    id::Integer,
    var_index::Integer
) -> Any

read_side_set_variable_values(exo, timestep, id, var_index)

Exodus.read_side_set_variable_values — Method

read_side_set_variable_values(
    exo::ExodusDatabase,
    timestep::Integer,
    id::Integer,
    var_name::String
) -> Any

read_side_set_variable_values(exo, timestep, id, var_name)

Exodus.read_side_set_variable_values — Method

read_side_set_variable_values(
    exo::ExodusDatabase,
    timestep::Integer,
    set_name::String,
    var_name::String
) -> Any

read_side_set_variable_values(
    exo,
    timestep,
    set_name,
    var_name
)

Exodus.read_side_sets — Method

read_side_sets(
    exo::ExodusDatabase,
    set_ids::Vector{<:Integer}
) -> Vector{SideSet}

read_side_sets(exo, set_ids)

Exodus.read_time — Method

read_time(exo::ExodusDatabase, time_step::Integer) -> Any

TODO figure out how to make this not use a vector of length 1 - either a ref or a ptr

read_time(exo, time_step)

Exodus.read_times — Method

read_times(exo::ExodusDatabase) -> Vector

read_times(exo)

Exodus.read_variable_name — Method

read_variable_name(
    exo::ExodusDatabase,
    var_index::Integer,
    type::Exodus.ex_entity_type
) -> String

read_variable_name(exo, var_index, type)

Exodus.read_variable_names — Method

read_variable_names(
    exo::ExodusDatabase,
    type::Exodus.ex_entity_type
) -> Vector{String}

read_variable_names(exo, type)

Exodus.read_variable_values — Method

read_variable_values(
    exo::ExodusDatabase,
    time_step::Integer,
    id::Integer,
    var_name::String,
    type::Exodus.ex_entity_type
) -> Any

read_variable_values(exo, time_step, id, var_name, type)

Exodus.read_variable_values — Method

read_variable_values(
    exo::ExodusDatabase,
    time_step::Integer,
    var_name::String,
    nset_name::String,
    type::Exodus.ex_entity_type
) -> Any

read_variable_values(
    exo,
    time_step,
    var_name,
    nset_name,
    type
)

Exodus.read_variable_values — Method

read_variable_values(
    exo::ExodusDatabase{M, I, B, F},
    timestep::Integer,
    id::Integer,
    var_index::Integer,
    type::Exodus.ex_entity_type
) -> Any

read_variable_values(exo, timestep, id, var_index, type)

Exodus.set_exodus_max_name_length — Method

set_exodus_max_name_length(exoid::Int32, len::Int32)

set_exodus_max_name_length(exoid, len)

Exodus.set_exodus_options — Method

set_exodus_options(options)

set_exodus_options(options)

Exodus.write_coordinate_names — Method

write_coordinate_names(
    exo::ExodusDatabase,
    coord_names::Vector{String}
)

Method to write coordinate names, e.g. x, y, z

write_coordinate_names(exo, coord_names)

Exodus.write_coordinates — Method

write_coordinates(
    exo::ExodusDatabase,
    coords::Array{F<:Real, 2}
)

Method to write coordinates

write_coordinates(exo, coords)

Exodus.write_element_block — Method

write_element_block(exo::ExodusDatabase, block::Block)

WARNING: currently does not support edges, faces and attributes

write_element_block(exo, block)

Exodus.write_element_block_connectivity — Method

write_element_block_connectivity(
    exo::ExodusDatabase,
    block_id::Integer,
    conn::Array{I<:Integer, 2}
)

write_element_block_connectivity(exo, block_id, conn)

Exodus.write_element_block_name — Method

write_element_block_name(
    exo::ExodusDatabase,
    block::Block,
    name::String
)

write_element_block_name(exo, block, name)

Exodus.write_element_block_names — Method

write_element_block_names(
    exo::ExodusDatabase,
    names::Vector{String}
)

write_element_block_names(exo, names)

Exodus.write_element_blocks — Method

write_element_blocks(
    exo::ExodusDatabase,
    blocks::Vector{<:Block}
)

write_element_blocks(exo, blocks)

Exodus.write_element_variable_name — Method

write_element_variable_name(
    exo::ExodusDatabase,
    var_index::Integer,
    var_name::String
)

write_element_variable_name(exo, var_index, var_name)

Exodus.write_element_variable_names — Method

write_element_variable_names(
    exo::ExodusDatabase,
    var_names::Vector{String}
)

write_element_variable_names(exo, var_names)

Exodus.write_element_variable_values — Method

write_element_variable_values(
    exo::ExodusDatabase,
    timestep::Integer,
    id::Integer,
    var_index::Integer,
    var_values::Vector{<:Real}
)

write_element_variable_values(
    exo,
    timestep,
    id,
    var_index,
    var_values
)

Exodus.write_element_variable_values — Method

write_element_variable_values(
    exo::ExodusDatabase,
    timestep::Integer,
    id::Integer,
    var_name::String,
    var_values::Vector{<:Real}
)

write_element_variable_values(
    exo,
    timestep,
    id,
    var_name,
    var_values
)

Exodus.write_global_variable_name — Method

write_global_variable_name(
    exo::ExodusDatabase,
    var_index::Integer,
    var_name::String
)

write_global_variable_name(exo, var_index, var_name)

Exodus.write_global_variable_names — Method

write_global_variable_names(
    exo::ExodusDatabase,
    var_names::Vector{String}
)

write_global_variable_names(exo, var_names)

Exodus.write_global_variable_values — Method

write_global_variable_values(
    exo::ExodusDatabase,
    timestep::Integer,
    var_values::Vector{<:Real}
)

write_global_variable_values(exo, timestep, var_values)

Exodus.write_initialization! — Method

write_initialization!(exoid::Int32, init::Initialization)

Used to set up a exodus database in write mode

The ccall signatures should reall be B (bulk int type of exo) instead of Clonglong

write_initialization!(exoid, init)

Exodus.write_nodal_variable_name — Method

write_nodal_variable_name(
    exo::ExodusDatabase,
    var_index::Integer,
    var_name::String
)

write_nodal_variable_name(exo, var_index, var_name)

Exodus.write_nodal_variable_names — Method

write_nodal_variable_names(
    exo::ExodusDatabase,
    var_names::Vector{String}
)

write_nodal_variable_names(exo, var_names)

Exodus.write_nodal_variable_values — Method

write_nodal_variable_values(
    exo::ExodusDatabase,
    timestep::Integer,
    var_index::Integer,
    var_values::Vector{<:Real}
)

write_nodal_variable_values(
    exo,
    timestep,
    var_index,
    var_values
)

Exodus.write_nodal_variable_values — Method

write_nodal_variable_values(
    exo::ExodusDatabase,
    timestep::Integer,
    var_name::String,
    var_values::Vector{<:Real}
)

write_nodal_variable_values(
    exo,
    timestep,
    var_name,
    var_values
)

Exodus.write_node_set — Method

write_node_set(exo::ExodusDatabase, set::NodeSet)

write_node_set(exo, set)

Exodus.write_node_set_name — Method

write_node_set_name(
    exo::ExodusDatabase,
    set::NodeSet,
    name::String
)

write_node_set_name(exo, set, name)

Exodus.write_node_set_names — Method

write_node_set_names(
    exo::ExodusDatabase,
    names::Vector{String}
)

write_node_set_names(exo, names)

Exodus.write_node_set_variable_name — Method

write_node_set_variable_name(
    exo::ExodusDatabase,
    var_index::Integer,
    var_name::String
)

write_node_set_variable_name(exo, var_index, var_name)

Exodus.write_node_set_variable_names — Method

write_node_set_variable_names(
    exo::ExodusDatabase,
    var_names::Vector{String}
)

write_node_set_variable_names(exo, var_names)

Exodus.write_node_set_variable_values — Method

write_node_set_variable_values(
    exo::ExodusDatabase,
    timestep::Integer,
    id::Integer,
    var_index::Integer,
    var_values::Vector{<:Real}
)

write_node_set_variable_values(
    exo,
    timestep,
    id,
    var_index,
    var_values
)

Exodus.write_node_set_variable_values — Method

write_node_set_variable_values(
    exo::ExodusDatabase,
    timestep::Integer,
    id::Integer,
    var_name::String,
    var_values::Vector{<:Real}
)

write_node_set_variable_values(
    exo,
    timestep,
    id,
    var_name,
    var_values
)

Exodus.write_node_set_variable_values — Method

write_node_set_variable_values(
    exo::ExodusDatabase,
    timestep::Integer,
    set_name::String,
    var_name::String,
    var_values::Vector{<:Real}
)

write_node_set_variable_values(
    exo,
    timestep,
    set_name,
    var_name,
    var_values
)

Exodus.write_node_sets — Method

write_node_sets(exo::ExodusDatabase, sets::Vector{NodeSet})

write_node_sets(exo, sets)

Exodus.write_number_of_element_variables — Method

write_number_of_element_variables(
    exo::ExodusDatabase,
    num_vars::Integer
)

write_number_of_element_variables(exo, num_vars)

Exodus.write_number_of_global_variables — Method

write_number_of_global_variables(
    exo::ExodusDatabase,
    num_vars::Integer
)

write_number_of_global_variables(exo, num_vars)

Exodus.write_number_of_nodal_variables — Method

write_number_of_nodal_variables(
    exo::ExodusDatabase,
    num_vars::Integer
)

write_number_of_nodal_variables(exo, num_vars)

Exodus.write_number_of_node_set_variables — Method

write_number_of_node_set_variables(
    exo::ExodusDatabase,
    num_vars::Integer
)

write_number_of_node_set_variables(exo, num_vars)

Exodus.write_number_of_side_set_variables — Method

write_number_of_side_set_variables(
    exo::ExodusDatabase,
    num_vars::Integer
)

write_number_of_side_set_variables(exo, num_vars)

Exodus.write_number_of_variables — Method

write_number_of_variables(
    exo::ExodusDatabase,
    num_vars::Integer,
    type::Exodus.ex_entity_type
)

write_number_of_variables(exo, num_vars, type)

Exodus.write_partial_coordinates — Method

write_partial_coordinates(
    exo::ExodusDatabase,
    start_node_num::Integer,
    coords::Array{F<:Real, 2}
)

write_partial_coordinates(exo, start_node_num, coords)

Exodus.write_partial_coordinates_component — Method

write_partial_coordinates_component(
    exo::ExodusDatabase,
    start_node_num::Integer,
    component::Integer,
    coords::Array{F<:Real, 1}
)

write_partial_coordinates_component(
    exo,
    start_node_num,
    component,
    coords
)

Exodus.write_partial_coordinates_component — Method

write_partial_coordinates_component(
    exo::ExodusDatabase,
    start_node_num::Integer,
    component::String,
    coords::Array{F<:Real, 1}
)

write_partial_coordinates_component(
    exo,
    start_node_num,
    component,
    coords
)

Exodus.write_set — Method

write_set(
    exo::ExodusDatabase{M, I, B, F},
    set::Exodus.AbstractSet
)

Typing ensures we don't write a set with non-matching types to the exodus file.

write_set(exo, set)

Exodus.write_set_name — Method

write_set_name(
    exo::ExodusDatabase{M, I, B, F},
    set::Exodus.AbstractSet,
    name::String
)

write_set_name(exo, set, name)

Exodus.write_set_names — Method

write_set_names(
    exo::ExodusDatabase,
    names::Vector{String},
    type::Exodus.ex_entity_type
)

write_set_names(exo, names, type)

Exodus.write_set_parameters — Method

write_set_parameters(
    exo::ExodusDatabase{M, I, B, F},
    set::Exodus.AbstractSet
)

WARNING: currently doesn't support distance factors

write_set_parameters(exo, set)

Exodus.write_sets — Method

write_sets(
    exo::ExodusDatabase,
    sets::Array{T<:Exodus.AbstractSet, 1}
)

write_sets(exo, sets)

Exodus.write_side_set — Method

write_side_set(exo::ExodusDatabase, set::SideSet)

write_side_set(exo, set)

Exodus.write_side_set_name — Method

write_side_set_name(
    exo::ExodusDatabase,
    set::SideSet,
    name::String
)

write_side_set_name(exo, set, name)

Exodus.write_side_set_names — Method

write_side_set_names(
    exo::ExodusDatabase,
    names::Vector{String}
)

write_side_set_names(exo, names)

Exodus.write_side_set_variable_name — Method

write_side_set_variable_name(
    exo::ExodusDatabase,
    var_index::Integer,
    var_name::String
)

write_side_set_variable_name(exo, var_index, var_name)

Exodus.write_side_set_variable_names — Method

write_side_set_variable_names(
    exo::ExodusDatabase,
    var_names::Vector{String}
)

write_side_set_variable_names(exo, var_names)

Exodus.write_side_set_variable_values — Method

write_side_set_variable_values(
    exo::ExodusDatabase,
    timestep::Integer,
    id::Integer,
    var_index::Integer,
    var_values::Vector{<:Real}
)

write_side_set_variable_values(
    exo,
    timestep,
    id,
    var_index,
    var_values
)

Exodus.write_side_set_variable_values — Method

write_side_set_variable_values(
    exo::ExodusDatabase,
    timestep::Integer,
    id::Integer,
    var_name::String,
    var_values::Vector{<:Real}
)

write_side_set_variable_values(
    exo,
    timestep,
    id,
    var_name,
    var_values
)

Exodus.write_side_set_variable_values — Method

write_side_set_variable_values(
    exo::ExodusDatabase,
    timestep::Integer,
    set_name::String,
    var_name::String,
    var_values::Vector{<:Real}
)

write_side_set_variable_values(
    exo,
    timestep,
    set_name,
    var_name,
    var_values
)

Exodus.write_side_sets — Method

write_side_sets(exo::ExodusDatabase, sets::Vector{SideSet})

write_side_sets(exo, sets)

Exodus.write_time — Method

write_time(
    exo::ExodusDatabase,
    time_step::Integer,
    time_value::Real
)

write_time(exo, time_step, time_value)

Exodus.write_variable_name — Method

write_variable_name(
    exo::ExodusDatabase,
    var_index::Integer,
    var_name::String,
    type::Exodus.ex_entity_type
)

write_variable_name(exo, var_index, var_name, type)

Exodus.write_variable_names — Method

write_variable_names(
    exo::ExodusDatabase,
    var_names::Vector{String},
    type::Exodus.ex_entity_type
)

write_variable_names(exo, var_names, type)

Exodus.write_variable_values — Method

write_variable_values(
    exo::ExodusDatabase,
    timestep::Integer,
    id::Integer,
    var_index::Integer,
    var_values::Vector{<:Real},
    type::Exodus.ex_entity_type
)

write_variable_values(
    exo,
    timestep,
    id,
    var_index,
    var_values,
    type
)

Exodus.write_variable_values — Method

write_variable_values(
    exo::ExodusDatabase,
    time_step::Integer,
    id::Integer,
    var_name::String,
    var_value::Vector{<:Real},
    type::Exodus.ex_entity_type
)

write_variable_values(
    exo,
    time_step,
    id,
    var_name,
    var_value,
    type
)

Exodus.write_variable_values — Method

write_variable_values(
    exo::ExodusDatabase,
    time_step::Integer,
    var_name::String,
    set_name::String,
    var_value::Vector{<:Real},
    type::Exodus.ex_entity_type
)

write_variable_values(
    exo,
    time_step,
    var_name,
    set_name,
    var_value,
    type
)

Exodus.@decomp — Macro

Exodus.@epu — Macro

Exodus.@exodiff — Macro