Exodus.jl Methods

Public methods

Base.close — Method

Used to close and ExodusDatabase.

Base.copy — Method

Used to copy an ExodusDatabase. As of right now this is the best way to create a new ExodusDatabase for output. Not all of the put methods have been wrapped and properly tested. This one has though.

Base.length — Method

Base.length — Method

Exodus.collect_element_connectivities — Method

collects all blocks by default

Exodus.collect_element_to_element_connectivities — Method

collect all blocks by default

Exodus.collect_node_to_element_connectivities — Method

collect all blocks by default

Exodus.copy_mesh — Method

Simpler copy method to only copy a mesh for output later on

Exodus.decomp — Method

Exodus.epu — Method

Exodus.epu — Method

Prints epu help message

Exodus.exodiff — Method

Return true if the two files pass the exodiff test. Otherwise it returns false

Exodus.exodiff — Method

Prints exodiff help message

Exodus.read_block — Method

Exodus.read_block — Method

Exodus.read_block_id_map — Method

TODO do this one later... it depends on a few things

TODO fix this to not use void_int... use a proper type

Exodus.read_blocks — Method

Helper method for initializing blocks.

TODO: change name to readelementblocks

Exodus.read_coordinates — Method

Method to read coordinates. Returns a matrix that is ndim x nnodes.

Exodus.read_element_type — Method

Exodus.read_id_map — Method

Exodus.read_ids — Method

Exodus.read_info — Method

Exodus.read_map — Method

TODO change to not use void_int

Exodus.read_name — Method

Exodus.read_name — Method

General method to read the name of a variable in index var_index for a given variable type V.

Examples: julia> readname(exo, ElementVariable, 1) "stressxx"

julia> readname(exo, GlobalVariable, 2) "reactionforce"

julia> readname(exo, NodalVariable, 1) "displx"

julia> readname(exo, NodeSetVariable, 1) "nsetdispl_x"

julia> read_name(exo, SideSetVariable, 1) "pressure"

Exodus.read_names — Method

Exodus.read_names — Method

General method to read the names of variables for a given variable type V.

Examples: julia> readnames(exo, ElementVariable) "stressxx" "stressyy" "stresszz" "stressxy" "stressyz" "stress_zx"

julia> readnames(exo, GlobalVariable) "globaldispl" "reaction_force"

julia> readnames(exo, NodalVariable) "displx" "disply" "displz"

julia> readname(exo, NodeSetVariable) "nsetdisplx" "nsetdisply" "nsetdispl_z"

julia> read_name(exo, SideSetVariable) "pressure"

Exodus.read_number_of_time_steps — Method

Exodus.read_number_of_variables — Method

General method to read the number of variables for a given variable type V.

Examples: julia> readnumberof_variables(exo, ElementVariable) 6

julia> readnumberof_variables(exo, GlobalVariable) 5

julia> readnumberof_variables(exo, NodalVariable) 3

julia> readnumberof_variables(exo, NodeSetVariable) 3

julia> readnumberof_variables(exo, SideSetVariable) 6

Exodus.read_qa — Method

Exodus.read_set — Method

Exodus.read_sets — Method

Exodus.read_side_set_node_list — Method

UNTESTED

Exodus.read_time — Method

TODO figure out how to make this not use a vector of length 1 - either a ref or a ptr

Exodus.read_times — Method

Exodus.read_values — Method

Wrapper method for global variables around the main readvalues method readvalues(exo::ExodusDatabase, t::Type{GlobalVariable}, timestep::Integer) = read_values(exo, t, timestep, 1, 1)

Example: read_values(exo, GlobalVariable, 1)

Exodus.read_values — Method

Wrapper method for nodal variables

Exodus.read_values — Method

Wrapper method for nodal variables

Exodus.read_values — Method

Exodus.read_values — Method

Exodus.read_values — Method

Method to read element variables

Exodus.read_values — Method

Method to read global variables

Exodus.read_values — Method

Method to read nodal variables

Exodus.read_values — Method

Method to read nodeset/sideset variables

Exodus.write_block — Method

WARNING: currently does not support edges, faces and attributes

Exodus.write_blocks — Method

Exodus.write_coordinates — Method

Method to write coordinates

Exodus.write_id_map — Method

Exodus.write_info — Method

Exodus.write_name — Method

Exodus.write_name — Method

Exodus.write_name — Method

Exodus.write_names — Method

WARNING: this methods likely does not have good safe guards

Exodus.write_names — Method

Exodus.write_number_of_variables — Method

General method to write the number of variables for a given variable type V.

Examples: julia> writenumberof_variables(exo, ElementVariable, 6)

julia> writenumberof_variables(exo, GlobalVariable, 5)

julia> writenumberof_variables(exo, NodalVariable, 3)

julia> writenumberof_variables(exo, NodeSetVariable, 3)

julia> writenumberof_variables(exo, SideSetVariable, 6)

Exodus.write_qa — Method

Exodus.write_set — Method

Typing ensures we don't write a set with non-matching types to the exodus file.

Exodus.write_sets — Method

Exodus.write_time — Method

Exodus.write_values — Method

Wrapper method for global variables around the main writevalues method writevalues( exo::ExodusDatabase, t::Type{GlobalVariable}, timestep::Integer, varvalues::Vector{<:AbstractFloat} ) = writevalues(exo, t, timestep, 1, 1, var_values)

Note: you need to first run writenumberof_variables(exo, GlobalVariable, n) where n is the number of variables.

Example: writenumberofvariables(exo, GlobalVariable, 5) writevalues(exo, GlobalVariable, 1, [10.0, 20.0, 30.0, 40.0, 50.0])

Exodus.write_values — Method

Wrapper for writing nodal variables by index number

Exodus.write_values — Method

Wrapper method for nodal variables

Exodus.write_values — Method

Exodus.write_values — Method

Exodus.write_values — Method

Private methods

Base.show — Method

Base.show — Method

Base.show — Method

Base.show — Method

Base.show — Method

Base.show — Method

Base.show — Method

Base.show — Method

Base.show — Method

Base.show — Method

Base.show — Method

Base.show — Method

Base.show — Method

Base.show — Method

Base.show — Method

Exodus.InitializationGlobal — Method

Exodus.collect_global_to_color — Function

For collecting globaltocolor

Exodus.epu_error — Method

Exodus.exodiff_error — Method

Exodus.exodus_error_check — Method

Generic error handling method.

Arguments

error_code::T: error code, usually negative means something went bad

method_name::String: method name that called this

Exodus.nem_slice — Method

Exodus.nem_slice — Method

Exodus.nem_slice_error — Method

Exodus.nem_spread — Method

Exodus.nem_spread — Method

Exodus.nem_spread_error — Method

Exodus.open_exodus_file — Method

Helper method for opening exodus database

Exodus.read_block_connectivity — Method

Exodus.read_block_parameters — Method

Exodus.read_blocks! — Method

TODO: change name to readelementblocks!

Exodus.read_coordinate_names — Method

Method to read coordinates names Returns a vector of strings

Exodus.read_init_info — Method

Exodus.read_node_set_nodes — Method

Exodus.read_partial_block_connectivity — Method

Exodus.read_partial_coordinates — Method

Method to read a partial set of coordinates that are contiguous. Returns a matrix that is ndim x nnodes

Exodus.read_partial_coordinates_component — Method

Method to read a specific component of a partial set of coordinates that are contiguous. Returns a vector of length n_nodes TODO change to not use Cvoid

Exodus.read_set_parameters — Method

Exodus.read_sets! — Method

Exodus.read_side_set_elements_and_sides — Method

Exodus.set_exodus_max_name_length — Method

Exodus.set_exodus_options — Method

Exodus.write_block_connectivity — Method

Exodus.write_coordinate_names — Method

Method to write coordinate names, e.g. x, y, z

Exodus.write_initialization! — Method

Used to set up a exodus database in write mode

The ccall signatures should reall be B (bulk int type of exo) instead of Clonglong

Exodus.write_partial_coordinates — Method

Exodus.write_partial_coordinates_component — Method

Exodus.write_partial_coordinates_component — Method

Exodus.write_set_parameters — Method

WARNING: currently doesn't support distance factors