Exodus.jl Methods

Public methods

Base.close — Method

close(exo::ExodusDatabase)

Used to close and ExodusDatabase.

Base.copy — Method

copy(
    exo::ExodusDatabase,
    new_file_name::String;
    mesh_only_flag
)

Used to copy an ExodusDatabase. As of right now this is the best way to create a new ExodusDatabase for output. Not all of the put methods have been wrapped and properly tested. This one has though.

Base.length — Method

length(n::NodeSet) -> Any

Base.length — Method

length(s::SideSet) -> Any

Exodus.collect_element_connectivities — Method

collect_element_connectivities(
    exo::ExodusDatabase{M, I, B, F}
) -> Any

collects all blocks by default

Exodus.collect_element_to_element_connectivities — Method

collect_element_to_element_connectivities(
    exo::ExodusDatabase{M, I, B, F}
) -> Any

collect all blocks by default

Exodus.collect_node_to_element_connectivities — Method

collect_node_to_element_connectivities(
    exo::ExodusDatabase{M, I, B, F}
) -> Any

collect all blocks by default

Exodus.copy_mesh — Method

copy_mesh(file_name::String, new_file_name::String)

Simpler copy method to only copy a mesh for output later on

Exodus.decomp — Method

decomp(file_name::String, n_procs::Integer)

Exodus.epu — Method

epu(file_name::String) -> Base.Process

Exodus.epu — Method

epu() -> Base.Process

Prints epu help message

Exodus.exodiff — Method

exodiff(
    ex_1::String,
    ex_2::String,
    cli_args::Vector{String}
) -> Bool

Exodus.exodiff — Method

exodiff(ex_1::String, ex_2::String; command_file) -> Bool

Return true if the two files pass the exodiff test. Otherwise it returns false

Exodus.exodiff — Method

exodiff() -> Base.Process

Prints exodiff help message

Exodus.read_block — Method

read_block(exo::ExodusDatabase, block_id::Integer) -> Block

Exodus.read_block — Method

read_block(exo::ExodusDatabase, block_name::String) -> Block

Exodus.read_block_id_map — Method

read_block_id_map(
    exo::ExodusDatabase{M, I, B, F},
    block_id
) -> Any

TODO do this one later... it depends on a few things

TODO fix this to not use void_int... use a proper type

Exodus.read_blocks — Method

read_blocks(
    exo::ExodusDatabase,
    block_ids::Union{Vector{<:Integer}, var"#s122"} where var"#s122"<:Integer
) -> Any

Helper method for initializing blocks.

TODO: change name to readelementblocks

Exodus.read_coordinates — Method

read_coordinates(exo::ExodusDatabase) -> Matrix

Method to read coordinates. Returns a matrix that is ndim x nnodes.

Exodus.read_element_type — Method

read_element_type(
    exo::ExodusDatabase,
    block_id::Integer
) -> String

Exodus.read_id_map — Method

read_id_map(
    exo::ExodusDatabase{M, I, B, F},
    type::Type{MAP<:Exodus.AbstractExodusMap}
) -> Any

Exodus.read_ids — Method

read_ids(
    exo::ExodusDatabase{M, I, B, F},
    _::Type{S<:Exodus.AbstractExodusSet}
) -> Any

Exodus.read_info — Method

read_info(exo::ExodusDatabase) -> Vector{String}

Exodus.read_map — Method

read_map(exo::ExodusDatabase{M, I, B, F}) -> Any

TODO change to not use void_int

Exodus.read_name — Method

read_name(
    exo::ExodusDatabase,
    _::Type{S<:Exodus.AbstractExodusSet},
    id::Integer
) -> String

Exodus.read_name — Method

read_name(
    exo::ExodusDatabase,
    _::Type{V<:Exodus.AbstractExodusVariable},
    var_index::Integer
) -> String

General method to read the name of a variable in index var_index for a given variable type V.

Examples: julia> readname(exo, ElementVariable, 1) "stressxx"

julia> readname(exo, GlobalVariable, 2) "reactionforce"

julia> readname(exo, NodalVariable, 1) "displx"

julia> readname(exo, NodeSetVariable, 1) "nsetdispl_x"

julia> read_name(exo, SideSetVariable, 1) "pressure"

Exodus.read_names — Method

read_names(
    exo::ExodusDatabase,
    _::Type{S<:Exodus.AbstractExodusSet}
) -> Any

Exodus.read_names — Method

read_names(
    exo::ExodusDatabase,
    _::Type{V<:Exodus.AbstractExodusVariable}
) -> Vector{String}

General method to read the names of variables for a given variable type V.

Examples: julia> readnames(exo, ElementVariable) "stressxx" "stressyy" "stresszz" "stressxy" "stressyz" "stress_zx"

julia> readnames(exo, GlobalVariable) "globaldispl" "reaction_force"

julia> readnames(exo, NodalVariable) "displx" "disply" "displz"

julia> readname(exo, NodeSetVariable) "nsetdisplx" "nsetdisply" "nsetdispl_z"

julia> read_name(exo, SideSetVariable) "pressure"

Exodus.read_number_of_time_steps — Method

read_number_of_time_steps(exo::ExodusDatabase) -> UInt32

Exodus.read_number_of_variables — Method

read_number_of_variables(
    exo::ExodusDatabase,
    _::Type{V<:Exodus.AbstractExodusVariable}
) -> Int32

General method to read the number of variables for a given variable type V.

Examples: julia> readnumberof_variables(exo, ElementVariable) 6

julia> readnumberof_variables(exo, GlobalVariable) 5

julia> readnumberof_variables(exo, NodalVariable) 3

julia> readnumberof_variables(exo, NodeSetVariable) 3

julia> readnumberof_variables(exo, SideSetVariable) 6

Exodus.read_qa — Method

read_qa(exo::ExodusDatabase) -> Matrix{String}

Exodus.read_set — Method

read_set(
    exo::ExodusDatabase,
    type::Type{S<:Exodus.AbstractExodusSet},
    set_id
) -> Any

Exodus.read_sets — Method

read_sets(
    exo::ExodusDatabase{M, I, B, F},
    type::Type{S<:Exodus.AbstractExodusSet}
) -> Any

Exodus.read_side_set_node_list — Method

read_side_set_node_list(
    exo::ExodusDatabase{M, I, B, F},
    side_set_id::Integer
) -> Tuple{Any, Vector}

UNTESTED

Exodus.read_time — Method

read_time(exo::ExodusDatabase, time_step::Integer) -> Any

TODO figure out how to make this not use a vector of length 1 - either a ref or a ptr

Exodus.read_times — Method

read_times(exo::ExodusDatabase) -> Vector

Exodus.read_values — Method

read_values(
    exo::ExodusDatabase,
    t::Type{GlobalVariable},
    timestep::Integer
) -> Vector

Wrapper method for global variables around the main readvalues method readvalues(exo::ExodusDatabase, t::Type{GlobalVariable}, timestep::Integer) = read_values(exo, t, timestep, 1, 1)

Example: read_values(exo, GlobalVariable, 1)

Exodus.read_values — Method

read_values(
    exo::ExodusDatabase,
    t::Type{NodalVariable},
    timestep::Integer,
    index::Integer
) -> Any

Wrapper method for nodal variables

Exodus.read_values — Method

read_values(
    exo::ExodusDatabase,
    t::Type{NodalVariable},
    timestep::Integer,
    name::String
) -> Any

Wrapper method for nodal variables

Exodus.read_values — Method

read_values(
    exo::ExodusDatabase,
    _::Type{NodalVectorVariable},
    timestep::Integer,
    base_name::String
) -> Any

Wrapper method for nodal vector variables

Exodus.read_values — Method

read_values(
    exo::ExodusDatabase,
    _::Type{V<:Union{ElementVariable, NodalVariable, NodeSetVariable, SideSetVariable}},
    time_step::Integer,
    id::Integer,
    var_name::String
) -> Any

Exodus.read_values — Method

read_values(
    exo::ExodusDatabase,
    _::Type{V<:Union{ElementVariable, NodeSetVariable, SideSetVariable}},
    time_step::Integer,
    set_name::String,
    var_name::String
) -> Any

Exodus.read_values — Method

read_values(
    exo::ExodusDatabase{M, I, B, F},
    type::Type{V<:ElementVariable},
    timestep::Integer,
    id::Integer,
    var_index::Integer
) -> Any

Method to read element variables

Exodus.read_values — Method

read_values(
    exo::ExodusDatabase{M, I, B, F},
    type::Type{V<:GlobalVariable},
    timestep::Integer,
    id::Integer,
    var_index::Integer
) -> Vector

Method to read global variables

Exodus.read_values — Method

read_values(
    exo::ExodusDatabase{M, I, B, F},
    type::Type{V<:NodalVariable},
    timestep::Integer,
    id::Integer,
    var_index::Integer
) -> Any

Method to read nodal variables

Exodus.read_values — Method

read_values(
    exo::ExodusDatabase{M, I, B, F},
    type::Type{V<:Union{NodeSetVariable, SideSetVariable}},
    timestep::Integer,
    id::Integer,
    var_index::Integer
) -> Any

Method to read nodeset/sideset variables

Exodus.write_block — Method

write_block(exo::ExodusDatabase, block::Block)

WARNING: currently does not support edges, faces and attributes

Exodus.write_block — Method

write_block(
    exo::ExodusDatabase,
    block_id::Integer,
    elem_type::String,
    conn::Array{I<:Integer, 2}
)

Exodus.write_blocks — Method

write_blocks(exo::ExodusDatabase, blocks::Vector{<:Block})

Exodus.write_coordinates — Method

write_coordinates(
    exo::ExodusDatabase,
    coords::Union{Array{F<:AbstractFloat, 1}, Array{F<:AbstractFloat, 2}}
)

Method to write coordinates

Exodus.write_id_map — Method

write_id_map(
    exo::ExodusDatabase{M, I, B, F},
    type::Type{MAP<:Exodus.AbstractExodusMap},
    map::Array{M, 1}
)

Exodus.write_info — Method

write_info(exo::ExodusDatabase, info::Vector{String})

Exodus.write_name — Method

write_name(
    exo::ExodusDatabase{M, I, B, F},
    set::Exodus.AbstractExodusSet,
    name::String
)

Exodus.write_name — Method

write_name(
    exo::ExodusDatabase{M, I, B, F},
    _::Type{S<:Exodus.AbstractExodusSet},
    set_id::Integer,
    name::String
)

Exodus.write_name — Method

write_name(
    exo::ExodusDatabase,
    _::Type{V<:Exodus.AbstractExodusVariable},
    var_index::Integer,
    var_name::String
)

Exodus.write_names — Method

write_names(
    exo::ExodusDatabase,
    _::Type{S<:Exodus.AbstractExodusSet},
    names::Vector{String}
)

WARNING: this methods likely does not have good safe guards

Exodus.write_names — Method

write_names(
    exo::ExodusDatabase,
    type::Type{V<:Exodus.AbstractExodusVariable},
    var_names::Vector{String}
)

Exodus.write_number_of_variables — Method

write_number_of_variables(
    exo::ExodusDatabase,
    _::Type{V<:Exodus.AbstractExodusVariable},
    num_vars::Integer
)

General method to write the number of variables for a given variable type V.

Examples: julia> writenumberof_variables(exo, ElementVariable, 6)

julia> writenumberof_variables(exo, GlobalVariable, 5)

julia> writenumberof_variables(exo, NodalVariable, 3)

julia> writenumberof_variables(exo, NodeSetVariable, 3)

julia> writenumberof_variables(exo, SideSetVariable, 6)

Exodus.write_qa — Method

write_qa(exo::ExodusDatabase, qa_record::Matrix{String})

Exodus.write_set — Method

write_set(
    exo::ExodusDatabase{M, I, B, F},
    set::Exodus.AbstractExodusSet
)

Typing ensures we don't write a set with non-matching types to the exodus file.

Exodus.write_sets — Method

write_sets(
    exo::ExodusDatabase,
    sets::Array{T<:Exodus.AbstractExodusSet, 1}
)

Exodus.write_time — Method

write_time(
    exo::ExodusDatabase,
    time_step::Integer,
    time_value::AbstractFloat
)

Exodus.write_values — Method

write_values(
    exo::ExodusDatabase{M, I, B, F},
    t::Type{GlobalVariable},
    timestep::Integer,
    var_values::Array{F, 1}
)

Wrapper method for global variables around the main writevalues method writevalues( exo::ExodusDatabase, t::Type{GlobalVariable}, timestep::Integer, varvalues::Vector{F} ) = writevalues(exo, t, timestep, 1, 1, var_values)

Note: you need to first run writenumberof_variables(exo, GlobalVariable, n) where n is the number of variables.

Example: writenumberofvariables(exo, GlobalVariable, 5) writevalues(exo, GlobalVariable, 1, [10.0, 20.0, 30.0, 40.0, 50.0])

Exodus.write_values — Method

write_values(
    exo::ExodusDatabase{M, I, B, F},
    t::Type{NodalVariable},
    timestep::Integer,
    var_index::Integer,
    var_values::Array{F, 1}
)

Wrapper for writing nodal variables by index number

Exodus.write_values — Method

write_values(
    exo::ExodusDatabase{M, I, B, F},
    t::Type{NodalVariable},
    timestep::Integer,
    var_name::String,
    var_values::Array{F, 1}
)

Wrapper method for nodal variables

Exodus.write_values — Method

write_values(
    exo::ExodusDatabase{M, I, B, F},
    _::Type{V<:Exodus.AbstractExodusVariable},
    timestep::Integer,
    id::Integer,
    var_index::Integer,
    var_values::Array{F, 1}
)

Exodus.write_values — Method

write_values(
    exo::ExodusDatabase{M, I, B, F},
    _::Type{V<:Exodus.AbstractExodusVariable},
    timestep::Integer,
    id::Integer,
    var_name::String,
    var_value::Array{F, 1}
)

Exodus.write_values — Method

write_values(
    exo::ExodusDatabase{M, I, B, F},
    _::Type{V<:Exodus.AbstractExodusVariable},
    time_step::Integer,
    set_name::String,
    var_name::String,
    var_value::Array{F, 1}
)

Private methods

Base.show — Method

Base.show — Method

Base.show — Method

Base.show — Method

Base.show — Method

Base.show — Method

Base.show — Method

Base.show — Method

Base.show — Method

Base.show — Method

Base.show — Method

Base.show — Method

Base.show — Method

Base.show — Method

Base.show — Method

Exodus.InitializationGlobal — Method

InitializationGlobal(
    exo::ExodusDatabase{M, I, B, F}
) -> Initialization

Exodus.bulk_int_mode — Method

bulk_int_mode(exo::Int32) -> Union{Type{Int32}, Type{Int64}}

Exodus.collect_element_connectivities! — Method

collect_element_connectivities!(
    conns::Array{Array{B, 1}, 1},
    blocks::Array{Block{I, Array{B, 2}}, 1}
)

Exodus.collect_element_to_element_connectivities! — Method

collect_element_to_element_connectivities!(
    elem_to_elem::Array{Array{B, 1}, 1},
    node_to_elem::Array{Array{B, 1}, 1},
    conns::Array{Array{B, 1}, 1}
)

Exodus.collect_global_to_color — Function

For collecting globaltocolor

collect_global_to_color(
    file_name::String,
    n_procs::Int64
) -> Vector{Int64}
collect_global_to_color(
    file_name::String,
    n_procs::Int64,
    n_dofs::Int64
) -> Vector{Int64}

Exodus.collect_node_to_element_connectivities! — Method

collect_node_to_element_connectivities!(
    node_to_elem::Array{Array{B, 1}, 1},
    conns::Array{Array{B, 1}, 1}
)

Exodus.copy_transient — Method

copy_transient(file_name::String, new_file_name::String)

Exodus.entries — Method

entries(s::SideSet) -> AbstractVector

Exodus.epu_error — Method

epu_error(cmd::Cmd)

Exodus.exodiff_error — Method

exodiff_error(cmd::Cmd)

Exodus.exodus_error_check — Method

exodus_error_check(
    exo::ExodusDatabase,
    error_code::Integer,
    method_name::String
)

Exodus.exodus_error_check — Method

exodus_error_check(
    exo::Int32,
    error_code::Integer,
    method_name::String
)

Generic error handling method.

Arguments

error_code::T: error code, usually negative means something went bad

method_name::String: method name that called this

Exodus.exodus_type_check — Method

exodus_type_check(sym, context, type1, type2)

Exodus.extras — Method

extras(s::SideSet) -> AbstractVector

Exodus.float_mode — Method

float_mode(
    exo::Int32
) -> Union{Type{Float32}, Type{Float64}}

Exodus.get_bulk_int_type — Method

get_bulk_int_type(_::ExodusDatabase{M, I, B, F}) -> Any

Exodus.get_file_id — Method

get_file_id(exo::ExodusDatabase) -> Int32

Exodus.get_float_type — Method

get_float_type(_::ExodusDatabase{M, I, B, F}) -> Any

Exodus.get_id_int_type — Method

get_id_int_type(_::ExodusDatabase{M, I, B, F}) -> Any

Exodus.get_init — Method

get_init(exo::ExodusDatabase) -> Initialization

Exodus.get_map_int_type — Method

get_map_int_type(_::ExodusDatabase{M, I, B, F}) -> Any

Exodus.get_mode — Method

get_mode(exo::ExodusDatabase) -> String

Exodus.id_error — Method

id_error(exo, _::Type{t<:Exodus.AbstractExodusSet}, id)

Exodus.id_error — Method

id_error(exo, _::Type{t<:Exodus.AbstractExodusVariable}, id)

Exodus.id_int_mode — Method

id_int_mode(exo::Int32) -> Union{Type{Int32}, Type{Int64}}

Exodus.initialization — Method

initialization(exo::ExodusDatabase) -> Initialization

Exodus.map_int_mode — Method

map_int_mode(exo::Int32) -> Union{Type{Int32}, Type{Int64}}

Exodus.mode_error — Method

mode_error(mode::String)

Exodus.name_error — Method

name_error(exo, _::Type{t<:Exodus.AbstractExodusSet}, name)

Exodus.name_error — Method

name_error(
    exo,
    _::Type{t<:Exodus.AbstractExodusVariable},
    name
)

Exodus.nem_slice — Method

nem_slice(
    file_name::String,
    n_procs::Integer
) -> Base.Process

Exodus.nem_slice_error — Method

nem_slice_error(cmd::Cmd)

Exodus.nem_spread — Method

nem_spread(
    file_name::String,
    n_procs::Integer
) -> Base.Process

Exodus.nem_spread_error — Method

nem_spread_error(cmd::Cmd)

Exodus.num_dimensions — Method

num_dimensions(init::Initialization) -> Any

Exodus.num_element_blocks — Method

num_element_blocks(init::Initialization) -> Any

Exodus.num_elements — Method

num_elements(init::Initialization) -> Any

Exodus.num_node_sets — Method

num_node_sets(init::Initialization) -> Any

Exodus.num_nodes — Method

num_nodes(init::Initialization) -> Any

Exodus.num_side_sets — Method

num_side_sets(init::Initialization) -> Any

Exodus.open_exodus_file — Method

open_exodus_file(file_name::String, mode) -> Int32

Helper method for opening exodus database

Exodus.read_block_connectivity — Method

read_block_connectivity(
    exo::ExodusDatabase{M, I, B, F},
    block_id::Integer,
    conn_length::Integer
) -> Vector

Exodus.read_block_parameters — Method

read_block_parameters(
    exo::ExodusDatabase{M, I, B, F},
    block_id::Integer
) -> Tuple{String, Vararg{Any, 5}}

Exodus.read_blocks! — Method

read_blocks!(
    blocks::Array{B<:Block, 1},
    exo::ExodusDatabase,
    block_ids::Array{I<:Integer, 1}
)

TODO: change name to readelementblocks!

Exodus.read_coordinate_names — Method

read_coordinate_names(exo::ExodusDatabase) -> Any

Method to read coordinates names Returns a vector of strings

Exodus.read_ghost_nodes_and_procs — Method

read_ghost_nodes_and_procs(rank, exo) -> Tuple{Any, Any}

Exodus.read_init_info — Method

read_init_info(
    exo::ExodusDatabase
) -> Tuple{Int32, Int32, String}

Exodus.read_internal_nodes_and_procs — Method

read_internal_nodes_and_procs(rank, exo) -> Tuple{Any, Any}

Exodus.read_node_cmaps — Method

read_node_cmaps(rank, exo) -> Any

Exodus.read_node_set_nodes — Method

read_node_set_nodes(
    exo::ExodusDatabase{M, I, B, F},
    set_id::Integer
) -> Any

Exodus.read_partial_block_connectivity — Method

read_partial_block_connectivity(
    exo::ExodusDatabase,
    block_id::Integer,
    start_num::Integer,
    num_ent::Integer
) -> Any

Exodus.read_partial_coordinates — Method

read_partial_coordinates(
    exo::ExodusDatabase,
    start_node_num::Integer,
    n_nodes::Integer
) -> Matrix

Method to read a partial set of coordinates that are contiguous. Returns a matrix that is ndim x nnodes

Exodus.read_partial_coordinates_component — Method

read_partial_coordinates_component(
    exo::ExodusDatabase,
    start_node_num::Integer,
    n_nodes::Integer,
    component::Integer
) -> Vector

Method to read a specific component of a partial set of coordinates that are contiguous. Returns a vector of length n_nodes TODO change to not use Cvoid

Exodus.read_partial_coordinates_component — Method

read_partial_coordinates_component(
    exo::ExodusDatabase,
    start_node_num::Integer,
    n_nodes::Integer,
    component::String
) -> Vector

Exodus.read_set_parameters — Method

read_set_parameters(
    exo::ExodusDatabase{M, I, B, F},
    set_id::Integer,
    _::Type{S<:Union{NodeSet, SideSet}}
) -> Tuple{Any, Any}

Exodus.read_sets! — Method

read_sets!(
    sets::Array{T<:Exodus.AbstractExodusSet, 1},
    exo::ExodusDatabase,
    ids::Array{I, 1}
)

Exodus.read_side_set_elements_and_sides — Method

read_side_set_elements_and_sides(
    exo::ExodusDatabase{M, I, B, F},
    set_id::Integer
) -> Tuple{Any, Any}

Exodus.read_values! — Method

read_values!(
    values::Array{F, 1},
    exo::ExodusDatabase{M, I, B, F},
    _::Type{V<:Exodus.AbstractExodusVariable},
    timestep::Integer,
    id::Integer,
    var_index::Integer
) -> Vector

Method to read element variables with pre-allocated array

Exodus.set_exodus_max_name_length — Method

set_exodus_max_name_length(exoid::Int32, len::Int32)

Exodus.set_exodus_options — Method

Exodus.set_name_dict — Method

set_name_dict(
    exo::ExodusDatabase,
    _::Type{Block}
) -> Dict{String}

Exodus.set_name_dict — Method

set_name_dict(
    exo::ExodusDatabase,
    _::Type{NodeSet}
) -> Dict{String}

Exodus.set_name_dict — Method

set_name_dict(
    exo::ExodusDatabase,
    _::Type{SideSet}
) -> Dict{String}

Exodus.var_name_dict — Method

var_name_dict(
    exo::ExodusDatabase,
    _::Type{ElementVariable}
) -> Dict{String}

Exodus.var_name_dict — Method

var_name_dict(
    exo::ExodusDatabase,
    _::Type{GlobalVariable}
) -> Dict{String}

Exodus.var_name_dict — Method

var_name_dict(
    exo::ExodusDatabase,
    _::Type{NodalVariable}
) -> Dict{String}

Exodus.var_name_dict — Method

var_name_dict(
    exo::ExodusDatabase,
    _::Type{NodeSetVariable}
) -> Dict{String}

Exodus.var_name_dict — Method

var_name_dict(
    exo::ExodusDatabase,
    _::Type{SideSetVariable}
) -> Dict{String}

Exodus.write_block_connectivity — Method

write_block_connectivity(
    exo::ExodusDatabase,
    block_id::Integer,
    conn::Array{I<:Integer, 2}
)

Exodus.write_coordinate_names — Method

write_coordinate_names(
    exo::ExodusDatabase,
    coord_names::Vector{String}
)

Method to write coordinate names, e.g. x, y, z

Exodus.write_initialization! — Method

write_initialization!(exoid::Int32, init::Initialization)

Used to set up a exodus database in write mode

The ccall signatures should reall be B (bulk int type of exo) instead of Clonglong

Exodus.write_partial_coordinates — Method

write_partial_coordinates(
    exo::ExodusDatabase,
    start_node_num::Integer,
    coords::Union{Array{F<:AbstractFloat, 1}, Array{F<:AbstractFloat, 2}}
)

Exodus.write_partial_coordinates_component — Method

write_partial_coordinates_component(
    exo::ExodusDatabase,
    start_node_num::Integer,
    component::Integer,
    coords::Array{F<:AbstractFloat, 1}
)

Exodus.write_partial_coordinates_component — Method

write_partial_coordinates_component(
    exo::ExodusDatabase,
    start_node_num::Integer,
    component::String,
    coords::Array{F<:AbstractFloat, 1}
)

Exodus.write_set_parameters — Method

write_set_parameters(
    exo::ExodusDatabase{M, I, B, F},
    set::Exodus.AbstractExodusSet
)

WARNING: currently doesn't support distance factors