Exodus.jl Types

Public methods

Exodus.Block — Type

Exodus.Block — Method

Init method for block container.

Exodus.Block — Method

Exodus.EdgeMap — Type

Exodus.ElementMap — Type

Exodus.ElementVariable — Type

Exodus.ExodusDatabase — Method

Type unstable helper to eliminate annoying lines of code to get type stability.

If you're looking for a type stable way to to open an exodus file, Simple copy past some of this into a barrier function

Exodus.ExodusDatabase — Method

Exodus.ExodusDatabase — Method

Exodus.ExodusDatabase — Method

Exodus.FaceMap — Type

Exodus.GlobalVariable — Type

Exodus.Initialization — Type

struct Initialization{ND, NN, NE, NEB, NNS, NSS}

Type that holds initialization information.

Exodus.Initialization — Method

Exodus.Initialization — Method

Initialization(_::Type{B<:Integer}) -> Initialization

Initialization(_)

Exodus.NodalVariable — Type

Exodus.NodeMap — Type

Exodus.NodeSet — Type

Exodus.NodeSet — Method

Exodus.NodeSet — Method

Exodus.NodeSetVariable — Type

Exodus.SideSet — Type

Exodus.SideSet — Method

Exodus.SideSet — Method

Exodus.SideSetVariable — Type

Private methods

Exodus.CommunicationMapParameters — Type

Exodus.CommunicationMapParameters — Method

Exodus.EPUException — Type

Exodus.ElementCommunicationMap — Type

Exodus.ElementCommunicationMap — Method

Exodus.ExodiffException — Type

Exodus.ExodusError — Type

Exodus.ExodusWindowsError — Type

Exodus.LoadBalanceParameters — Type

Exodus.LoadBalanceParameters — Method

Exodus.ModeException — Type

Exodus.NemSliceException — Type

Exodus.NemSpreadException — Type

Exodus.NodeCommunicationMap — Type

Exodus.ProcessorElementMaps — Type

Exodus.ProcessorElementMaps — Method

Exodus.ProcessorNodeMaps — Type

Exodus.ProcessorNodeMaps — Method

Exodus.SetIDException — Type

Exodus.SetNameException — Type

Exodus.VariableIDException — Type

Exodus.VariableNameException — Type

Exodus.ex_entity_id — Type

Exodus.ex_entity_type — Type

ex_entity_type

Entity type enums (exentitytype in exodusII.h)

Exodus.ex_inquiry — Type

ex_inquiry

Inquiry enums (ex_inquiry in exodusII.h).

Exodus.void_int — Type

void_int = Cvoid